HCI: Fix improper rfkill handling
Current behavior with kernel rfkill module loaded produces a cascade of
crash events starting from the following logcat error:

    unable to bind bluetooth user channel: Operation not possible due to RF-kill

However, the same device configuration worked well with btlinux HAL,
which I assume was used as a base for this HAL.

After some investigation, I found that the rfkill(1) function call in the
btlinux call changed the state to 1, which is RFKILL_STATE_UNBLOCKED.
The current rfkill(1) sets the soft block value to 1, which corresponds
to the state value 0 (RFKILL_STATE_SOFT_BLOCKED).

Fix it by providing the correct values to the rfkill() call.

Change-Id: I7a62744f9c21d80b9f8fa37f6c172afa2a5b6c8a
Signed-off-by: Roman Stratiienko <r.stratiienko@gmail.com>