Add new listener to perform screen recording.
The ScreenRecorderManager uses ADB to start an independent thread to
perform screen recording during the test execution. The recording is
automatically stopped at test teardown and the video file is saved to
the logging folder. The recording has a max duration of 180 secs.

Add @ScreenRecord annotation to be used to attach the recording listener
to a given method.

BUG: b/377927896
Change-Id: Ibb4a35bdbc30c373e70f97367092468db873f791