ui: don't move focus to notes input when created
This is a tricky UX decision because you could argue you want this.
However, IMO most of the time people create a note to mark that time
point, not because they actually want to annotate it. By moving focus,
we also make it much harder to actually delete the note as you need to
click the note *twice* to be able to delete it (once to select, once to
remove focus from text box).

Making the focus explicit makes this much more obvious to the user that
the focus is on the box so other hotkeys will not work: tbh I did not
even notice that the focus was being stolen for several minutes while
debugging this issue.

Change-Id: I01ae9318b59b166ddd7d7c244d0c287d25f6857f
Bug: https://github.com/google/perfetto/issues/965