serialize syncThread creation
try, to avoid driver lock up in make current
call on nvidia gpu.

this does not consistently happen, but when
it happens, there are a few sync threads calling
into SyncThread::initSyncEGLContext, and one
is always sleep on make current into the driver;
others blocked by the segl lock half way through
their completion of initSyncEGLContext task.

The hypothesis is that the initSyncEGLContext code
is not thread safe, hence the mutex here to
force them to complete initSyncEGLContext one by
one.

This could slow down the starting up a bit, but
a good enough trade off for stability.


Bug: 383543476

Change-Id: If0f02f448dff51cafb3ae861732ff421c1a04c0d