Add fetch caching toggle to local instance case
Allow control of this behavior via `acloud` to allow test configs to be
updated for the desired behavior.

Defaulting to `false` because of the issue in the parent bug of the one
linked below.  When an effective cache management strategy has been put
in place the default can be flipped to `true`.

Bug: 383775908
Test: m acloud
Test: acloud-dev create --branch=aosp-main
--build_target=aosp_cf_x86_64_phone-trunk_staging-userdebug
--local-instance -vv

Scanning through the logs should show the `cvd fetch` argument
list with the default value of `--enable_caching=false`.

Test: ls -l /tmp/cvd/<user_id>/cache

Should not show any directories with a timestamp matching the `acloud`
fetch.

Test: rm -r /tmp/acloud_image_artifacts
acloud-dev create --branch=aosp-main
--build_target=aosp_cf_x86_64_phone-trunk_staging-userdebug
--local-instance -vv --enable_fetch_local_caching

Logs show the flag as true, and the cache should show new files.

Change-Id: I2d4c50a77943d887c61796d4ea3679137c349d1f