Add PatternNode.scan(...) to scan for files that match patterns
Building on the previous work to add `PatternNode` this change adds
the `scan(...)` method to support using the `PatternNode` hierarchy
to efficiently scan for files that match the patterns encapsulated in
the `PatternNode` hierarchy.

Basically, it passes `PatternFileState` objects that contain the `File`
(a directory) to scan to a `PatternNode`. It checks that directory for
files that match the part of the pattern that it encapsulates. If it
finds any that it will create new `PatternFileState` objects for those
`File`s and pass it on to its children. The `ApiVersionPatternNode`
will also extract the `ApiVersion` from the file name and store that
in the `PatternFileState`.

Leaf `PatternNode`s, i.e. those who do not have any children, will
convert the `PatternFileState` into `MatchedPatternFile` which will be
returned to the caller. That encapsulates the `File` and `ApiVersion`.
It will always contain an `ApiVersion` as every pattern added must
contain `%` and so can only match if `ApiVersionPatternNode` matches.

A follow-up change will use this in `findAndroidJars(...)`.

Bug: 383288863
Test: ./gradlew
Change-Id: I09c80dc9b33265286890f6eff22899e76bd09b79