Add support for sequential video recording to the screen record tool.
The screen record listener records a single video with a max duration of
180 secs from adb screenrecord. This update modified the behavior to
record sequential videos of 180 secs each until the test completes. Each
video is saved with an incrementing filename suffic and stored in the
log folder.

Rename ScreenRecorderManager to ScreenRecorder.i

Add annotation marker to enableTwentyFourHourFormat test in API 32.

BUG: b/377927896
Change-Id: I02bd050879a6f3498b02583270fdf2de5aad70ea