Introduce a Copybara consistency file
This file was generated by running a modified version of Copybara with
a special patch to add custom baseline support for folder origins in
squash mode. Copybara was pointed at a clean checkout of Chromium
126.0.6423.0 for both the origin and the baseline; this results in the
production of an initial copybara consistency file with no other
changes.

Copybara consistency files are a way to keep track of any changes that
have only been made to the destination (here, AOSP) and have not been
upstreamed to the origin (here, Chromium). Consistency files contain,
for each file in the destination:

 - A checksum of the file in the destination - this is just a sanity
   checking mechanism for subsequent copybara runs to reduce the risk of
   human error;

 - The difference (if any) between the file in the destination and the
   (transformed) origin, at the time the consistency file was generated.
   This makes it possible to reconstruct the original baseline by
   reversing the diff.

Consistency files are meant to be used in subsequent Copybara runs. In
future runs, Copybara will check out the destination repo *at the time
the consistency file was last modified*, and use it to reconstruct the
merge import baseline, i.e. what the (transformed) origin looked like at
the time. This makes it possible to reconstruct the diff between the old
and new origins, something that wouldn't be normally be possible due
to our use of folder origin which doesn't support history (and also
other caveats related to intervening copybara configuration changes).
From there, copybara can automatically execute a 3-way merge between
the origin diff and the destination-only changes. This is an alternative
to the manual use of git branches and brings us closer to being able
to automate these merges.

Test: N/A
Bug: 384688443
Bug: 381421702
Bug: 382268057
Change-Id: I0565381f84d58eb249d92e064693e56268809210