transport_socket_server: Check return value before errno
socket_spec_listen() is used to create a listen socket to accept
incoming adb connection in adbd.

Currently there is an issue that adbd cannot listen on a localhost
socket even if `service.adb.listen_addrs` is set to
`tcp:localhost:5555`. The reason is that socket_spec_listen() will call
base::ParseInt() in its implementation,  which will set errno in this
case (since it cannot parse `localhost:5555` as an int), and errno will
not be reset after that. As a result, although the socket is created
successfully, the errno check will be matched at first and it will be
considered as an "unrecoverable error".

Since errno can be set by any function calls in the middle, while the
return value actually indicates the execution result of this function,
we should check the return value at first.

Bug: 385868238
Test: Verified that localhost addr works on cf.
Change-Id: I1af39fd9e3e3a1268fefabdd10584d996fd86a46