tp: stop overriding args for tracks which are not owned by track event
This CL makes a change to the decision about whether to update the args
for track_event tracks. Specifically, we no longer handle any updates
of args for the default thread tracks which are shared with all the
other data sources in trace processor. It's unfair to do so as track
event does not own them.

This breaks a bunch of features but we have a readymade workaround with
disallow_merging_with_system_tracks which we can recommend in a pinch

Change-Id: If99e03b493fe56824fe04289a99d65b42bbd9e01