alloctop
alloctop is a tool that simplifies the handling of /proc/allocinfo data.
It takes the slabtop as an interface reference, and uses similar command
line parameters.

The current version of the tool implements:
- Limitation of the output lines;
- Threshold for the minimum allocation size that a row needs to have to
  be printed;
- Sorting by size (numeric, descending), calls (numeric, descending) or
  tag (alphabetical, ascending);
- Tree view: the data is aggregated by the components of the tag field.

The only output mode implemented right now is --once, meaning that
currently there is no interface that periodically refreshes showing the
latest data.

Bug: 384777823
Test: adb shell dumpstate
Change-Id: Ib83ecb62e25d086551973ddab3f6ad73bb97cc0c
Signed-off-by: Alessio Balsini <balsini@google.com>