ui: rewrite how trace processor tracks are added to the UI
This CL completely overhauls the way in which trace processor tracks are
added to the UI. The new system is *opt-in* not *opt-out* as the
previous system was. This allows enforcing that people do not pollute
the top-level workspace with random tracks, instead grouping things
properly into top level groups.

Track event is also separated into its own plugin due to its complexity
and are now handled properly.

Change-Id: Ie68965df927ac682d57da6306bcdc05442a6560c