ui: Update Wattson idle cost name
Coincidentally, there are two identical table names in Wattson UI
aggregation and Wattson metrics, and normally these are two independent
callflows. However, in the UI, there is a possibility of running both
the metric as well as doing various UI actions, which causes SQL errors
when referencing the ostensibly different SQL table/view. Update the UI
table to name, and so that this naming collision doesn't occur.

Bug: 387329860
Change-Id: If1c40efde0fd774927f904e3354c75f13a08c4c4
Signed-off-by: Samuel Wu <wusamuel@google.com>