ui: Fix misaligned WASD zoom center after window resize
If the browser window width is changed, the new timeline width is
usually  not taken into account which throws the zoom center out of
whack when zooming with W & S keys. This results in the zoom center
being no longer aligned to the mouse location.

This CL fixes the issue by updating the timeline rect on the every
render cycle rather than just when the canvas changes size, which is
purposely done lazily to avoid expensive resize operations.

Updating the timeline rect is a very cheap operation compared to
rendering so the impact on performance is immeasurable.

VirtualCanvasOverlay::onCanvasResized() has also been removed as it is
no longer used, and most likely not useful/will cause confusion going
forward.

Change-Id: Ibe448fffdbf5ade577b039b9a63ec6cb73f17cc0
Bug: 387445739