Improve testing of VisibleForTesting
Adds a separate test class and splits the existing test into separate
Java and Kotlin tests and added new tests that use the numeric values
of the `VisibleForTesting` constants.

The new number tests fail on all providers as the code does not handle
them. The new Java only test fails on Turbine. The previous combined
Java and Kotlin test did not fail on Turbine because it was only run
for providers that support Kotlin, which Turbine does not.

Follow-up changes will fix the tests.

Bug: 383521289
Test: ./gradlew
Change-Id: I6dd61a76632932f2e1bbd11179c1b59e8f864edd