Migrate BaseDeviceMetricCollector to use SearchArtifactUtil
BaseDeviceMetricCollector has it's own logic for finding files in the
test artifacts, which was originally copied from PushFilePreparer but
has now diverged. Migrate it to use the new common implementation.

One problem this should fix is that the current implementation doesn't
handle cases where the module directory is different to the module name,
e.g. when using extra_test_configs.

Bug: 382547516
Bug: 368261467
Test: atest power-idle-display-on

Change-Id: I7b8625b98a7900fb377ee1eb5579e8c317ee44b4