Remove support for using % as a placeholder for an API level
Historically, `%` was the only way to specify a placeholder for an API
level in an `--android-jar-pattern` option. However, as part of the
work to allow other placeholders a more explicit format for the
placeholder was added, i.e. `{version:level}`. This change removes
support for `%` and replaces all uses of that with `{version:level}`.

Bug: 383288863
Test: ./gradlew
Change-Id: I90eaf1d9f6013a3970c2dc9e62f95246c7e22951