Add placeholder {version:major.minor?}
Adds new `Placeholder` `{version:major.minor?}`.

It was necessary as part of this to improve the check for detecting
an invalid number of instances of `{version:level}` within the same
pattern. Previously, the check just counted the number of
`PlaceholderPatternNode`s and made sure that there was exactly 1.
If there was 0 it assumed that `{version:level}` was missing and if
there were >1 then it assumed that there were multiple instances of
`{version:level}`. However, with this change it could also be due to
issues with `{version:major.minor?}`.

The check is there to ensure that the `PatternFileState`'s `version`
property is set exactly once. So, it was changed to track how many
`Placeholder`s were used for each `Property`.

Bug: 378479241
Test: ./gradlew
Change-Id: Idf01f724da2ba8eb74475e8d421c02f4879581a3