verifier: Reduce duplication for `kVerifierDebug`.
Most functions do not use this template argument and the
linker is able to deduplicate identical instantiations but
compiling these instantiations takes a significant time.

Timing with
    touch art/runtime/verifier/method_verifier.cc && \
    time art/tools/buildbot-build.sh --target
after this target was already built, shows
 - before:
    real 2m9.716s
    user 35m58.720s
    sys  3m11.564s
 - after:
    real 1m39.871s
    user 33m0.007s
    sys  3m40.512s
While the multi-threaded linking stage does not change much,
the single-threaded compilation of `method_verifier.cc` is
greatly improved, especially for the included debug builds.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Change-Id: I58ab16937fa7ca2d28c5cc0e979c7f8e6042926b