Cleanup TemporaryFolderOwner methods for new folder/file
Previously, the `newFolder(...)` and `newFile(...)` methods did not
make it clear in their name that they created new folders/files
respectively only if they did not already exist. This change renames
and documents them to make that clearer. It also makes sure that
when creating a new file that the parent directory exists first.

Bug: 388441272
Test: ./gradlew
Change-Id: I426e4b3701273a86eee3ff911fd1f267686deaab