verifier: Do not mark registers as conflict for return.
The conflict marking originates in the July 2013 commit
    b8c7859f21f5ae4c9b90f2ef2effc51967299737
    Revert "Revert "Remove non-live vregs from GC map on return.""
but the intent is no longer valid. We do not use verifier
information for GC maps in the compiler.

Remove monitor stack verification from `UpdateRegisters()`,
it is done when visiting the return instruction anyway.

Similarly remove the `CheckConstructorReturn()` check that
used to be in the now removed `AdjustReturnLine()` as we do
that when visiting the `return-void` anyway.

After this, `UpdateRegisters()` cannot report hard failure
(note that `RegTypeCache::MergeRegisters()` never reports
any hard failure), so remove the unnecessary error handling.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Change-Id: I8fca564a98b676a671e41275102bb397e4c04712