Fix for default margins of PWG-raster PDL
This update extends the recently merged fix for content trimming.
In the existing code, the default margins for PWG-Raster PDL are
incorrectly set to 0, which can result in content clipping at the
edges. To address this issue, we are aligning the margins for
PWG-Raster with those of PCLm to prevent clipping.

Test: Print a letter sized PDF that has content at the edges to a
PWG-raster printer loaded with Letter paper. Confirm printout
does not have clipping.
Signed-off-by: Saurabh Chauhan <mopriadevteam@gmail.com>

Change-Id: Idb8190ba9d5f70b011c064c5d82fb0d1ae473f0e