BUILD: translate gn source_sets to Bazel/Blaze as cc_library
This CL changes the logic for how we translate C++ code from gn ->
Bazel. Instead of translating targets into filegroups, bubbling
up dependencies manually in Python, make source_sets just be cc_library
targets and rely on GN to do all the dependency tracking and bubbling
for us.

This solves a major issue with ODR violations if multiple Perfetto
libraries are included in the same project.

Change-Id: If37acaa222596cac7d9b805229278eb6572acd3a