Autoupdate of 7 submodules
*external/mobile-data-download updated to 2ce58fde3ee5a2df7e1da989937d7c5ce37e8703
Removing upstream flag as it is now the default.
Also moving to internal robolectric projects to no longer use the  _upstream suffix.

Test: CI
Bug: 386652866
Change-Id: Ic707ef14704b219065cc178ad7e1abb92e327027

*frameworks/opt/localepicker updated to aa0c17a33ea7e377ed7dbfcce2168cae6e198d3a
Removing upstream flag as it is now the default.
Also moving to internal robolectric projects to no longer use the  _upstream suffix.

Test: CI
Bug: 386652866
Change-Id: If210e213cf2e34ff94c4e0f34cc21cd1d00a2337

*frameworks/opt/net/wifi updated to 99c1ff3222affd859a0aa19ffbec86f275ee80c2
Removing upstream flag as it is now the default.
Also moving to internal robolectric projects to no longer use the  _upstream suffix.

Test: CI
Bug: 386652866
Change-Id: Ib5d0c15de98d7c1dc393d83c88c1806de50610e1

*packages/apps/Settings updated to 7aa0ef40498933460bc38041f98c0fca74b22c39
Merge "Removing upstream flag as it is now the default. Also moving to internal robolectric projects to no longer use the  _upstream suffix." into main
*packages/apps/TV updated to 06a88c4e331579c5084112648ceacf90d44cb4e8
Removing upstream flag as it is now the default.
Also moving to internal robolectric projects to no longer use the  _upstream suffix.

Test: CI
Bug: 386652866
Change-Id: I605ae9e5833d5539aff64a672b24b836f4b27ff0

*packages/apps/WallpaperPicker2 updated to cb7195ae5199f68bf5dd2b62e37f95adc03b39f7
Removing upstream flag as it is now the default.
Also moving to internal robolectric projects to no longer use the  _upstream suffix.

Test: CI
Bug: 386652866
Change-Id: I4a3ce1cdc32a917600d2da89cf78015a0ee116fd

*prebuilts/sdk updated to c694be60070bf857e6c723e78bade17cab3f9120
Merge "Removing upstream flag as it is now the default. Also moving to internal robolectric projects to no longer use the  _upstream suffix." into main