[Chrome Scroll Jank plugin] Colorize scroll update slices
This CL updates the colors and names of "Scroll Update" slices on the
Chrome Scroll Timeline track based on the following classification:

  JANKY (red)
  FIRST_SCROLL_UPDATE_IN_FRAME (dark green)
  COALESCED (gray)
  INERTIAL (teal)
  DEFAULT (indigo)

For name, we combine all applicable scroll update classifications into
the name. For example, if a scroll update is both janky and inertial,
its name will be 'Janky Inertial Scroll Update'.

For color, we pick the highest-priority applicable scroll update
classification. For example, if a scroll update is both janky and
inertial, we classify it as janky (red).

The above logic is implemented in SQL for better performance and
scalability (e.g. it avoids duplicating the naming logic across the
track and the details panel).

Bug: 383995805
Change-Id: Idbcec33c1db6f9cf4793ccf65d111bd8da67c66f