Extract appendDefinitionListHelp(...) from nonInlineEnumOption(...)
Allows what is effectively an HTML definition list in the help text,
i.e. a term paired with a definition, that was previously added for
options that took an enum to explain each allowable value, to be reused
elsewhere in the help text.

This will be used for generating help for `FileFormat` related enums.

Bug: 388809101
Test: ./gradlew
Change-Id: I6366297aa5121d84dc65480aa1ca0dd7c7e13e9e