ANDROID: Refactor symlink generation in blueprints
There are many places in linux-firmware where we need to create a
symlink to another file. This could be to handle multiple kernel
versions that look for the file with a different name or in a
different location or it could be because multiple drivers can
currently use the same file and so we just use symlinks to point to
one central file.

When we have multiple drivers using symlinks to point ot the same
file, we'll do this:
- Factor the common file out into its own "workflow".
- Have all drivers depend on the factored out workflow and each driver
  will install a symlink to that common file.

Unfortunately, with the way the current code handled symlinks we
tripped an error path in Soong. It wouldn't allow us to install a
pre-created symlink to a file that was part of another module.

Fix this problem by using an "install_symlink" module, which appears
to be fine with this case. We'll use the "required" feature of Soong
to depend on any symlinks created so that things work automatically.

Bug: 361834699
Test: Install a module with a symlink and see symlink get created
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:92ab110fa711ce6387b0457a4098ae531e0c0781)
Merged-In: Ic6437e84ab7391d6973c255d6d3f318dceb33a5a
Change-Id: Ic6437e84ab7391d6973c255d6d3f318dceb33a5a