Refactor replaceChoiceOption(...)
Cleans up the method to make it easier to support adding prefixes in
a follow-up change. Previously, the part of the match that needed
replacing was computed from the whole match and relied on the match
having a single hard newline character before the part that needed
replacing. Adding a prefix after the hard newline character before
the text that needs replacing would break that.

This change adds a group to identify the part of the match that needs
replacing, removing the limitation of only allowing a single character
before the replacement.

Bug: 388809101
Test: ./gradlew
Change-Id: Id0971656876490722de2a008fbdc66747a70a872