Avoid using trace_data_lock_ in TraceStartCheckpoint
We used to record methods currently on thread stack into a string in
trace_data_. This required using a lock to synchronize across different
threads. This CL changes this by recording this information into the
per-thread buffer avoiding the need of locks. Since we record them in
the per-thread buffer we would have fewer free spaces in the buffer.
We don't expect this to be a big problem.

Bug: 352518093
Test: art/test.py
Change-Id: Ia9960456a17cfabf5ff6680925f4243b8bb892fb