ui: display CPU and machine for multi-machine traces
This revision implements display of CPU numbers in multiple-machine
traces (e.g. ChromeOS host + ARCVM guest). The UI currently shows the
ucpu number for remote machine CPU tracks. For example, CPU 0 of remote
machine 1 is labeled as CPU 4096. This revision fixes the UI so that
the CPU is labeled as CPU 0 (machine 1).

Bug: 284258446
Test: screenshot test arcvm_trace.test.ts
Change-Id: Ia30ac5bc547033e3ea03a9c45880a90c4af94cc4