verifier: Minor clean up of hard error reporting.
Move last failure check from `CodeFlowVerifyInstruction()`
to the caller. Move the "Rejecting opcode" message as well
and guard it with `if (kVerifierDebug)` as most other info
messages - it's not very important as we get the dex pc in
the main failure message anyway.

Adjust failure messages that relate to the `kContinue` flag
to make it clear that the failing instruction itself was OK.
Note that these shall now go through the last failure check
and info message in the caller.

Simplify `check-cast`/`instance-of` failure printing.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Change-Id: If2ec8dee50d6cf6d5ac7969367a838b7fe9702ce