verifier: Clean up `HandleMoveException`.
Stop processing the instruction on hard failure. Previously,
we could have hit a soft failure after a hard failure. (And
we would also unnecessarily update the register type.)

Pull the `move-exception` handling code out of the
`PotentiallyMarkRuntimeThrow()`. The calls between this
function and `Fail()` were confusing and seemingly recursive
even though there was no recursion thanks to the parameters
we were passing. The `PotentiallyMarkRuntimeThrow()` was
also keeping `flags_.have_pending_runtime_throw_failure_` as
false for `move-exception`, further adding to the confusion.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Change-Id: I991ff9b93c4d50a206cb0964de875e9687c07aba