test: Mark tests with emulator exceptions as skipped
Adds a pytest hook to modify test reports for EmulatorExceptions.
If a test fails due to an EmulatorException, the hook modifies
the report to mark the test as skipped and adds information about
the infrastructure error.

This change also includes adding a "infrastructure-error" property
to the JUnit XML report for tests that failed due to an emulator exception,
making it easier to identify and track these failures.

Bug: 390037698
Change-Id: I0738c1075882e25a61210a14746ac2d061466f6a