Split findAndroidJars(...) into two
Previously, `findAndroidJars(...)` would scan for files and then create
`VersionedJarApi` instances from those files. This change splits it
into two methods, one that scans for the files and the other which is
given a list of files and creates `VersionedJarApi`s. This is in
preparation for combining the scanning for extension jars and android
jars.

Bug: 259115852
Test: ./gradlew
Change-Id: I75eb5fe1515aa020b03f5b2dc8e49e4e0865801d