Add {version:extension} placeholder
All other `{version:*}` placeholders are ignored if an `ApiVersion`
range was provided and they were outside of it. However, the range only
applies to SDK versions and not extension versions. e.g. if generating
API for versions 10 to 35 then it should still include all extension
versions.

The goal of this bug is to support providing the pattern for extension
jars on the command line using the existing `--android-jar--pattern`.
That will require combining the separate `PatternNode.scan(...)` calls
for Android SDK and extension jars into one call.

While `{version:level}` currently works for scanning for extension
versions as no version range is provided on that call it will not work
when the scanning is combined and a version range is provided.

This change adds a new `{version:extension}` placeholder which behaves
liks `{version:level}` but ignores the range. It replaces the use of
`{version:level}` in `ExtensionSdkJarReader`.

Bug: 259115852
Test: ./gradlew
Change-Id: I32645f195607fc619f291c1ff68067a0abbae17d