Combine scanning for Android and extension jars
Previously, the scanning for Android jars and extension jars were done
separately. This change combines the two by adding the hard coded
pattern for the extension jars to the list of patterns provided to the
`--android-jar-pattern` command line option. The resulting files are
split apart based on whether they have a `module` name or not.

This will allow patterns for extension jars to be provided on the
command line using `--android-jar-pattern` and remove the need for the
`--sdk-extensions-root`. That will be cleaned up in following changes.

It was necessary to tweak the creation of `MatchedPatternFile` to
handle patterns that are not descendants of the current directory to
avoid this breaking the `ApiLevelsGenerationOptionTest` test method
`Test extension jar files in forAndroidConfig`.

Bug: 259115852
Test: ./gradlew
Change-Id: I93ef908c70dde12ea8f2cae27079c4c219ebbc91