Add PatternNode test for empty list of patterns
Previously, passing an empty list to `PatternNode.parsePatterns(...)`
would always result in a failure as it would match the directory that
was being scanned but fail as no version was found.

Currently, the list passed to `PatternNode.parsePatterns(...)` is never
empty because of the fallback that is added so this is not a problem.
However, a follow-up change will remove that fallback and those tests
that do not provide an explicit `--android-jar-pattern` will start to
fail without this change.

So, this change makes sure that an empty list of patterns is handled
correctly and just returns an empty sequence. That was chosen instead
of failing on an empty list of patterns as that was effectively the
previous behavior.

Bug: 390113278
Test: ./gradlew
Change-Id: I9cb50893a6481156e6b8fc63faa20d014feb6345