kernel_test: update expectations for min LTS for 25Q2
25Q2 is expected to release in the final month of the second quarter of
2025, ie. June 2025

Looking at:
https://android.googlesource.com/kernel/configs/+/refs/heads/main/kernel-lifetimes.xml

and fast forwarding the date to 2025-06-02 (the first is a Sunday, and
kernel expiration dates are usually on the first of the month), we can
notice that the minimum supported kernels are:
  android12-5.10: 5.10.218
  android13-5.10: 5.10.210
  android13-5.15: 5.15.149
  android14-5.15: 5.15.149
  android14-6.1:  6.1.78
  android15-6.6:  6.6.30
  android16-6.12: isn't yet listed, so we can use the latest LTS from kernel.org 6.12.9 (it's bound to be even higher yet)

Moving the cutoff one month forward to 2025-07-02 would only change:
  android13-5.10: 5.10.214
  android14-5.15: 5.15.153
  android14-6.1:  6.1.84
which only bumps 5.10 from 5.10.210 to 5.10.214 and 6.1 from 6.1.78 to
6.1.84 and is probably thus not worth it...

5.4.277 is chosen as the minimum for non-GKI2.0 5.4 kernel, as that was
released 2024-05-25, and will thus be > 1 year old on 2025-06-02 (unlike
5.4.278 from 2024-06-16, and the current 5.4.289 from 2025-01-09).

I'm not aware of us actually having any 5.4.x testing though, so 5.4 is
on shaky ground. With the Pixel 6+ kernel uprev to 6.1, even 5.10 and
5.15 are looking a bit iffy in terms of our ability to keep things
working reliably.

Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: Ia1302152937852c985b5f0fbb043f722d6aa5f0b