Skip 6E180015072901 at HAL.
This is required to FIX a CTS Test failure.
When CTS Test is run, SR2XX FW seems to treat this
as an invalid/unsupported packet and it returns a
failure NTF to upper layer.

This NTF has to be masked by Hal and must not be
sent to upper layers.

Bug: 381330041
Test: atest android.uwb.cts.UwbManagerTest#testSendVendorUciMessageWithMessageType android.uwb.cts.UwbManagerTest#testSendVendorUciMessageVendorGid
Change-Id: I5d27789b7c9cf4c0c46706844ff99be6d1737a74