Deprecate --sdk-extensions-root command line option
Previously, that was provided to specify the location of the SDK
extensions root directory that was then scanned for extension jars.
However, the extension jar location should now be specified using a
pattern in the `--android-jar-pattern` just as is done for
`android.jar` files.

This deprecates the option and stops it being required when the
`--sdk-extensions-info` option is specified which will allow the
Android build to stop using it and using `--android-jar-pattern`
instead.

Bug: 259115852
Test: ./gradlew
Change-Id: I1eb49737ca81c254ec31b4c2bad15c33208f802f