tp: add structured query proto and conversion from proto -> SQL
This CL adds the new concept of "structured queries" to trace processor.
These allow for representing relatively simple SQL queries as a
tree/graph of proto objects, making it possible to build visualizations
of Perfetto queries.

As well as this, this CL adds code to convert these protos to
PerfettoSQL queries, a necessary step to making the protos actually
useful

Change-Id: I0df20a11e53a554fe4237036e5970df502b70d77