Include minor version in @apiSince/@deprecatedSince tags
Previously, even if the `api-versions.xml` file included minor versions
they were not included in the `@apiSince` or `@deprecatedSince` tags.
This change fixes that by retrieving the `minor` value from `ApiLookup`
alongside the `major` value and creating a combined `major.minor`
`ApiVersion` from them.

The `ApiLookup` will treat version `X` and `X.0` as the same returning
a minor value of `0`. If that was simply passed to `ApiVersion` then it
would result in `@apiSince` tags for existing versions changing from
`@apiSince X` to `@apiSince X.0`. That would potentially be confusing
to developers when reading the resulting documentation in addition to
breaking lots of existing tests. So, this replaces a `minor` value of
`0` with `null` to maintain the existing behavior.

Bug: 302189716
Test: ./gradlew
Change-Id: I35af293bcf29f53807bd27096a8b394086e182ee