Add test for extracting class annotations
The test shows that class annotations are not currently extracted. That
will be fixed in a follow-up.

The test purposely includes an annotation on a method as otherwise the
would be no generated `annotation.xml` file for `test.pkg` and there is
currently no way to express and test for that in the `check(...)`
method. At best the test could catch the exception that is thrown in
that case but it complicates the follow-up change. Having an
`annotation.xml` file that just has the wrong content makes the delta
in the follow-up change simpler.

The test fails on Turbine because `ExtractAnnotations` does not yet
support it.

Bug: 391413938
Test: ./gradlew
Change-Id: I0f8230ab1467e3028068184d1c00c204ba34ab4c