Update signature verification files
Updates the signature verification files prior to adding a new
dependency to make it easier to see the effects of adding the
dependency. This adds more `<component>` elements and reorders an out
of sequence `<trusted-key>` element in the metadata file and removes
an unnecessary version header for BouncyCastle in the keyring file
courtesy of the fix for https://github.com/gradle/gradle/issues/31035.

Generated with:
    ./gradlew --write-verification-metadata sha256 --export-keys

Bug: 391554590
Test: ./gradlew
Change-Id: I503ab34dd888497fc33422e840577611e859f042