video: Enable deviceless unit tests in cros-codecs
This CL adds rust_test_host to enable deviceless unit tests for
cros-codecs. Once enabled, the unit tests are exercised in presubmit for
changes under cros-codecs only.

This CL also defines the rust_defaults module for other modules to
share.

Note that the changes to the Android.bp file are manual, instead of
using the tool cargo_embargo, which does not generate the expected entry
for the existing and new targets.

Bug: 385215152
Test: atest libcros_codecs_test
Change-Id: Ic1ea1d50876633c174a959ba75f448c4ab470a40