Move stats reporting from PreRebootDriver to PreRebootDexoptJob.
In the async flow for OTA, we need to call an update_engine API
in PreRebootDexoptJob, so we have to report stats there.

Also, this CL changes where a fatal error is handled, so that a fatal
error can be propagated to the shell command handler and shown to the
user.

Bug: 377557749
Test: atest ArtServiceTests
Change-Id: I2e939b147706d4d35c4b813c37db3225871870cd