Initialize the lowoverhead trace entry points to nop entry points
We update them when we register the thread, so it doesn't impact what we
initialize these values to. They will be updated based on the
lowoverhead trace type when we register the thread. However, since the
expectation is to have nop entry points when there is no trace running
it is good to initialize them with nop entry points to start with.

Bug: 352518093
Test: art/test.py
Change-Id: I931646a7203ccab8cae6e037a763660b61921e3d