Ensure consistency between command line and config surfaces
Previously, the configured `<api-surface>`s were not used by Metalava.
This change ensures that if the `--api-surface` option is specified
that it references a configured `<api-surface>` and there is
consistency between the command line options and the configuration.

At the moment that consistency just verifies that the `showUnannotated`
property is `true` if and only if the corresponding `<api-surface>`
does not extend another surface.

Follow-up changes will make more use of the configured
`<api-surface>`s.

Bug: 391554590
Test: ./gradlew
Change-Id: I34925983c0f56efd70b10cfbf8163e8c58a09d6f