feat: Enable local runs of build scripts and update documentation
We now allow for the configuration of the build target. This allows
local developers to specify 'rcache' vs. 'ci'. This will allow
developers to locally experiment with the build script without needing to
rely on the build infrastructure.

The 'ci' configuration requires a series of credentials that are not
available to local developers, resulting in build failures.

The documentation has also been updated to reflect these changes and
provide more detailed instructions on how to run the build script
locally.  This includes information on the new `--config` flag and how
to configure credentials for `rcache`.

Change-Id: Iac1021c506befd6d8e42416c00c9f81ef5441a94