Allow runTest(...) to take empty arguments
Previously, passing in an empty array of arguments to `runTest(...)`
would cause it to print the help message. That meant that it was not
possible to test the default behavior of an option without providing
at least one other option which may not always be possible.

This change adds an additional `--help` option to use instead. That
does get included in the generated help so it was necessary to remove
that before checking the help.

Bug: 391554590
Test: ./gradlew
Change-Id: I9f3c0883f96ccb4db1bf727569c7d6bf2161c76b