Add enum value for when an InvalidKeyException is thrown during log list verification
Previously, we were just logging the same FAILURE_SIGNATURE_VERIFICATION, but it would be good to have more granularity to know when this is because the signature might be altogether invalid (e.g. wrong format or algorithm) as opposed to not being able to verify the log list.

Change-Id: Id05fe93fbd84d06b7df848b2651bc9ed73e190a8
Flag: com.android.net.ct.flags.certificate_transparency_service
Bug: 378626065
Bug: 391327942
Test: builds