Ignore --show*-annotation when generating from signature files
Previously, `ApiSelectionOptions.createApiSurfaces(...)` would always
check consistency of the `--show*-annotation` options against the
configuration when an `--api-surface` option was specified. However,
show annotations are only relevant when generating APIs from source
files. When generating stubs from signature files the annotations are
irrelevant and so are not provided, meaning `showUnannotated` defaults
to `true` even when generating stubs for a surface like `system` that
extends another. That leads to consistency check failure.

One way of addressing that would be to not supply the `--api-surface`
option when generating stubs but while the annotations are not
important in that case the relationship between surfaces is important
as signature files must be processed in order such that the file for a
surface comes before the file of the surface that extends it. e.g.
`public` must be processed before `system` which must be processed
before `module-lib`.

This change fixes that by just ignoring the show annotations and so not
performing a consistency check with them when generating from signature
files

Bug: 391554590
Test: ./gradlew
Change-Id: Ieeebe5c003e9783000f1a157055f7438cdf38235