Extend suspend timeout for debug activities
Extend the debugging suspend timeout substantially if it appears we
are running concurrently with an ANR dump. There is some concern
that we may be regularly getting suspend timeouts because too many
cycles are devoted to an ANR dump at the time.

Extend it almost indefinitely if the cuplrit thread is in a tracing
stop 't' state. This makes debugging slightly easier, and should reduce
the number of spurious bug reports we have to process.

Add GetOsThreadStatQuick test and include minimal test for
GetStateFromStatString.

Bug: 330444460
Bug: 392053307
Test: Treehugger
Change-Id: I820e7283fda84fa9fd9820ff9948b992e10547ec