lib: rust_support: add spinlock bindings
This adds `spin_lock_t` bindings and a Rust `SpinLock<T>` wrapper.

The main API of `.(try_)lock_save` disables interrupts
and saves the state before locking,
as this has minimal overhead and is the safest way.
`.(try_)lock_unsaved` (`spin_(try_)lock`) are still public,
but with warnings that interrupts must first be disabled
to avoid deadlocks, and suggesting `.(try_)lock_save` instead.

The interrupt flags for a `SpinLock` are set in `SpinLock::new`
so that all locks of the same `SpinLock` use the same flags,
and the flags are checked vs. what's currently disabled
to help avoid deadlocks.

An `IRQSpinLock` alias is also defined as a default
corresponding to `spin_lock_irq{save,restore}` in C.

`spin_lock_held` is also not used, as our `Mutex` wrapper
and `std::sync::Mutex` also do not have a `.is_locked` method,
as there is little use for one vs. the atomic `.try_lock`.

Bug: None
Test: build.py generic-x86_64
Change-Id: I07c9bf38a4d18d9fe3928d45372902296f2e9e45