Add test to show incorrect handling of extension when finalizing
When finalizing an API the `--current-codename` is either not set or is
set to `REL`. In either case the `--current-version` should be the new
finalized version, and extension APIs that were added since the last
finalized version must be included as being added in the new finalized
version. The latter is achieved by passing the new finalized version to
`ExtensionUpdater.nextSdkVersion`.

However, that is not the current behavior as shown by the test added in
this change. That shows that the `ExtensionUpdater.nextSdkVersion` is
actually set to new finalized version + 1. That causes exceptions to be
thrown by the `ApiXmlPrinter` when it looks for the next version after
the version in which an item was last present to compute the `removed`
version.

A follow-up change will fix this.

Bug: 392565932
Test: ./gradlew
Change-Id: I7b948515aa7a29439676f48f3bc4e6b35f2994af