Partially order ApiSurfaceConfigs based on extends relationship
The `ApiSurfacesConfig.apiSurfaceList` is in the order in which they
were defined in the configuration file. It may have cycles and a
surface may be referenced (in `extends`) before it is defined. That
makes it hard to work with.

This change adds an `orderedSurfaces` that applies a partial ordering
on the `ApiSurfaceConfig`s such that an `ApiSurfaceConfig` is always
before an `ApiSurfaceConfig` that extends it. If two
`ApiSurfaceConfig`s do not have an extends relationship (even
indirectly) then they are ordered based on the order in the
configuration.

Preparation for using this to create `ApiSurfaces` and `ApiSurface`
objects.

Bug: 391554590
Test: ./gradlew
Change-Id: I6cd97bf70a25ec6f8691cde07995dfbd97cf3e7b