Improve error handling of ConfigParser
There were a number of issues with the error handling of `ConfigParser`
that could lead to usability issues for developers:
1. It ignored fatal errors and carried on parsing and reading `Config`
   objects. At best that would lead to multiple errors for the same
   problem being reported, at worst it could lead to exceptions
   elsewhere in the code that would be meaningless to developers and
   could obscure the original exception.
2. Some exceptions were not caught and handle, e.g. `IOException` which
   would abort Metalava with a stack trace being dumped which is not
   very helpful.

This change fixes that by stopping using a `Reporter` to report issues
(and removing the `CONFIG_FILE_PROBLEM` issue) and instead just
collating errors found by the parser and throwing an exception with all
the errors. Instead of interleaving parsing a config file with reading
a `Config` object from it this parses all the config files and then if
there were no errors reads the `Config` objects. That simplifies the
error handling.

Bug: 391554590
Test: ./gradlew
Change-Id: I09142adb4f871a794e7c78266bb92aaf931d2b5c