Validate merged Config before returning from ConfigParser
Previously, `ConfigParser` would just return the merged `Config` object
and leave it up to properties, e.g. `ApiSurfacesConfig.byName` to
perform any validation that was required. However, that meant it was
possible for some code to access `Config` objects in an invalid state
through other properties, e.g. `apiSurfaceList`.

This change performs validation on the merged `Config` object before it
is returned by `ConfigParser`, ensuring that it is valid before any
other code access it.

Bug: 391554590
Test: ./gradlew
Change-Id: I4ed1eb76af294367559a232ed8fca10b716f58fc