Use log level INFO for hiddenapi messages if access is allowed.
WarnAboutAccess can also be called when access is allowed, e.g. when
the runtime is started with `-Xcompiler-option --debuggable`. INFO is a
more appropriate log level in that case since nothing is wrong.

This amends https://r.android.com/3326636. Before that CL, the log
level was always WARNING, regardless of result and enforcement policy.

Test: 674-hiddenapi 690-hiddenapi-same-name-methods 691-hiddenapi-proxy
      817-hiddenapi 822-hiddenapi-future 999-redefine-hiddenapi
      2038-hiddenapi-jvmti-ext 2270-mh-internal-hiddenapi-use
      on host with `testrunner.py --debuggable`
Bug: 392883970
Bug: 377676642
Change-Id: I9fc4abc1741c70eba981fefc4b701f92eb6209d5