Add --api-version-signature-pattern
Previously, when passing `--api-version-signature-files` the user also
had to pass `--api-version-names` to specify the versions associated
with each file in `--api-version-signature-files`. The values in the
two options had to be kept in sync.

This change adds `--api-version-signature-pattern` which leverages the
existing `PatternNode` mechanism used for `--android-jar-pattern` to
extract the version from the files in `--api-version-signature-files`.
That should simplify the user and Metalava (once `--api-version-names`
can be removed).

Bug: 388441272
Test: ./gradlew
Change-Id: I4ab8c8663b683dafbbe86c0e7861d0c70f49bfe0