arm64: Add qemu-desktop-arm64-test-debug
This patch adds makefiles for a new qemu desktop target. The makefile
structure is designed to match the existing generic qemu structure. By
adding this build, we will be able to run desktop specific storage
tests on host and eventually enable those tests to run in presubmit.
At the moment, there are some configuration differences between the
generic prebuilt Android image that qemu relies on and the desktop
configuration today. As a result, certain tests that rely on desktop
specific services may be difficult to run for now. Regardless, this
CL provides an initial platform for executing most tests.

Bug: 393181441
Test: `build.py qemu-desktop-arm64-test-debug`

Change-Id: I9dd821ed75a089b38cc3caa6e07f0d85a830b544