Add a function to log with a timestamp.
In order to avoid waking the entire Android OS, it might be better to
batch log messages in a buffer and send them later. When doing this,
the code wants to preserve the original timestamps and pass those in
when sending the messages. Add a new function to do this.

Add a new unit tests to verify this functionality.

Bug: 393191584

Test: All unit tests pass.
Test: Wrote a log message with a zero timestamp and verified the log
Test: contains this message with a timestamp.
Test: Verified logcat using a time filter, out of order messages are properly
Test: filtered out and other messages after are still printed.
Change-Id: Ia11cf0efc85717eefd3802da2774070cc0e1f531