Create VersionedSignatureApis from MatchedPatternFile
Previously, `VersionedSignatureApi`s were create by combining the
`pastApiVersions` (from `--api-version-signature-files`) and
`allVersions` (from `--api-version-names`) together. This adds an
intermediate step to create a list of `MatchedPatternFile`s first. That
is in preparation for using the `PatternNode` mechanism to extract the
version names from the `--api-version-signature-files` automatically
and remove the need for `--api-version-names`.

Bug: 388441272
Test: ./gradlew
Change-Id: I9564d984aba5525fb45d2f479c8e81e9a6639b14