Move creation of ApiPrinter into ApiGenerator
Previously, an `ApiPrinter` was created and stored in the
`GenerateApiHistoryConfig`. On one path the printer to use is dependent
on the output file extension, on the other it is hard-coded to
`ApiXmlPrinter`. This moves the latter into `generateApiHistory(...)`
ensuring consistency.

It does mean that if the output file passed to `--generate-api-levels`
did not have an `xml` extension that it would have worked previously
but will start to fail. If that does become an issue then it can be
fixed by defaulting to use xml instead of failing.

Bug: 383521289
Test: ./gradlew
Change-Id: I3af9f00beead1e0bf9395ccec9e941c7aceef230