Import Cronet 132.0.6808.3 into AOSP
This commit brings in Cronet from Chromium 132.0.6808.3 (plus cherry-picks) into
AOSP, along with merge conflict resolutions.

WHAT THIS COMMIT CONTAINS

The code in this commit was imported from:

    https://chromium.googlesource.com/experimental/chromium/src/+/refs/wip/cronet/aosp-132.0.6808.3
    Commit SHA: 7a951e79b67bfc399fcc587745ab9613bf567517

This commit is a descendant of Chromium src tag 132.0.6808.3 (06e4dca5), with
the following commits added on top:

 - 51 commits that were already backported from the Chromium src main branch
   in the previous import;
 - 5 commits newly backported from the Chromium src main branch;
 - 2 commits that are not yet on the Chromium src main branch, but will be soon;
 - 3 commits that are specific to 132.0.6808.3 and are not going to be submitted
   to the Chromium src main branch.

The only other change made in this CL on top of the imported code is resolution
of merge conflicts.

HOW TO REPRODUCE THIS COMMIT

The contents of this commit (minus the manual adjustments) can be reproduced by
running:

    gclient config --spec 'solutions = [
    {
        "name": "src",
        "url": "https://chromium.googlesource.com/experimental/chromium/src.git",
        "managed": False,
        "custom_deps": {},
        "custom_vars": {
            "checkout_copybara": True,
        },
    },
    ]
    target_os = ["android"]
    ' &&
    gclient sync --rev=refs/wip/cronet/aosp-132.0.6808.3 --shallow &&
    cd src &&
    vpython3 components/cronet/gn2bp/run_gn2bp.py

When running this command on top of this commit, the result is expected to be an
empty diff.

Fixes: 392052561
Change-Id: I3f0825a773e12865a33b49355f9b053be3dba116