Create text files in ApiGeneratorTest in root temporary dir
Previously, the `createTextFile(...)` method would create each file in
its own random directory. That is not a realistic representation of how
they are organized in AndroidX, where all the versions exist within the
same directory. This change creates them all relative to the root
temporary directory. It is a standalone change just to show that it
does not break any of the existing tests.

Bug: 388441272
Test: ./gradlew
Change-Id: Ie2e9dad0a6075d9ac8f77cfc1c720fabe214e838