[scudo] Clean up secondary tests. (#124999)
Change names to all begin with ScudoSecondary and change tests names
appropriately.

Move the cache option test to the cache test fixture.

Force the allocator test to use the no cached config so that all of
the allocations always fully exercise the allocator function and
don't skip this by using a previously cached element.

GitOrigin-RevId: c8f4189eeb927247ead2d8b082a6920d5b873aa8
Change-Id: If1f4260b979c1dca7d356fbacde05f26790e2a67