build: Use ML64 for x64 assembly compilation
The clang-cl compiler on Windows cannot directly compile .asm files.
This commit introduces the use of ML64 (the Microsoft Macro Assembler
64-bit) for compiling assembly files when using the x64 Windows clang-cl
toolchain.  ML64 is included with Visual Studio and is capable of
compiling .asm files into object files.

The following flags are now used on all platforms:

*   `compile_flags`: These flags are applied during C/C++ compilation
    using clang. They control aspects like target architecture,
    warning levels, and exception handling
    .
*   `assembler_flags`: These flags are passed to ML64/clang during assembly.
    They control the assembly process.

Windows only:

*   `compiler_defines_flags`: These flags define preprocessor macros
    that can be used to conditionally compile code. They're used
    for things like setting the target Windows version
    (`_WIN32_WINNT`), disabling specific warnings
    (`_CRT_SECURE_NO_WARNINGS`), and identifying the compiler
    (`COMPILER_MSVC`).

This commit ensures that assembly files are correctly compiled when
using the x64 Windows clang-cl toolchain and allows us to use different
feature sets for arm/x64 in the future.

Change-Id: I22a8aa3dd332d29e17ace781c9ebf94bdb7b984a