Add explicit extension property to MatchedPatternFile
Previously, differentiating between an SDK extension file and a dessert
release file was done by checking for the presence of an SDK extension
module name in the path as it was necessary to use `{module}` in a
pattern that matches an SDK extension file because they are stored in
module specific files. However, module specific files are also present
in the dessert release directories of `prebuilts/sdk` so it is not a
good differentiator. The presence of `{version:extension}` is better as
it is only used for extensions.

So, this change adds an explicit `extension` property and sets that to
`true` if the version is updated by a `{version:extension}`
placeholder.

As the code for handling extension APIs assumes that the
`MatchedPatternFile.module` property is set to a non-`null` value which
is only true if the pattern includes a `{module}` placeholder this adds
a check to verify that a pattern with a `{version:extension}`
placeholder also has a `{module}` placeholder.

Bug: 259115852
Test: ./gradlew
Change-Id: Ibc96c9d603ef25650c5f475fb088980dd6f9cbc6