pinweaver: Fix crash on boot
This patch addresses a pinweaver app crash on boot related to setting a
non-zero `UnbufferedService::max_message_length`. Although not
explicitly stated in the documentation, it seems a non-zero length
causes a data write fault in the underlying binder RpcServer class. It
is unclear at the moment if this is a bug in the Trusty `tipc` crate or
in `RpcServer`.

Bug: 388563293
Test: `build.py qemu-desktop-arm64-test-debug`

Change-Id: I0310dd9e42736499ae3feba993ce4ec39163698f