ThreadList::Dump() fix
Dump() can be called with and without the mutator lock. aosp/3111768
caused it to ask RunCheckpoint to repeatedly acquire the mutator lock
even when it already had it. Fix that.

This seemed to show up mostly after some other failures that try to
generate a thread dump.

Improve an obsolete comment.

Bug: 368234017
Test: Treehugger
Change-Id: I712d8d6f48fb60f7718a774a8379196cd945dfa6