Use ApiHistoryUpdater in VersionedSignatureApi
Previously, only the `VersionedJarApi` took an `ApiHistoryUpdater` as a
constructor parameter. The other `VersionedApi` subclasses took an
`ApiVersion` and used `ApiHistoryUpdater` internally. This change makes
`VersionedSignatureApi` also take an `ApiHistoryUpdater`. As
`VersionedSourceApi` is the only `VersionedApi` subclass that does not
take an `ApiHistoryUpdater` the `ApiHistoryUpdater` was pulled up into
`VersionedApi`.

Bug: 378479241
Test: ./gradlew
Change-Id: Ie0b8f658752c903145b8b28d2530ee9fa1e43d19