Replace apiVersion with updater in addApisFromCodebase(...)
Previously, the `addApisFromCodebase(...)` took an `ApiVersion` as a
parameter and used that to construct an `ApiHistoryUpdater` that was
used internally. This change replaces the `apiVersion` with an
`ApiHistoryUpdater`.

This is necessary because `ApiHistoryUpdater` handles updating the
history for both Android/AndroidX APIs and for Android SDK extensions.
Currently, `VersionedJarApi` is used for Android SDK extensions but in
future `VersionedSignatureApi` will also be used for that. A follow-up
change will allow `VersionedSignatureApi` to take an
`ApiHistoryUpdater`.

Bug: 378479241
Test: ./gradlew
Change-Id: I8d848067bf9be3aba44716d1d409a2021b943dce