dev: virtio: vsock: use hypercall based transport for protected x86 vm
In contrast to ARM, when Trusty is running as a protected VM on x86,
there is need for special care of emulated MMIO accesses: it is because
on x86 the host needs to read and decode guest instruction, which caused
the trap on MMIO access to find out which address and access type cause
it. Nevertheless the host, due to pKVM nature, can't read the guest
memory. To overcome this issue pKVM added support for handling pair of
hypercalls, which are aimed for IOREAD and IOWRITE and can be used by
guest.

The latest version of virtio-drivers introduced the new x86-specific
virtio transport based on mentioned hypercalls. Since the virtio-drivers
was updated to v.0.9.0, which contains mentioned support (aosp/3469618),
lets use it for x86 protected Trusty.

Test: run x86 Trusty as a protected VM and successfully run Trusty unit
      test which relies on virtio-vsock:
      /vendor/bin/trusty-ut-ctrl -D VSOCK:$CID:1 com.android.ipc-unittest.ctrl
Bug: 373536001

Change-Id: I7781c2f18febd3d0f7b7549a87d4a5db3f1711ab