[scudo] Refactor the secondary test (#125595)
Remove all redundant code and create a couple of structs to handle
automatic init and destruction. This replaces the test fixtures in
prepartion for passing in multiple configs for some of these tests. This
is necessary because not all of the gtest features are supported here,
and there is no easy way to create a test fixture with a template.

GitOrigin-RevId: 84fbed86ffcb97c24f9294a204c60da5444b8646
Change-Id: I8126b0685db43d44717c2c11c3d4fefe6d5bfaf6