Fix interface handling in addApisFromCodebase(...)
Previously, `addApisFromCodebase(...)` would not treat an interface
class as extending `java.lang.Object` like `readJar(...)` would so it
needed to fix up the `<extends name="java/lang/Object"/>` to try and
mimic the behavior of `readJar(...)`. While that worked when calling
`addApisFromCodebase(...)` after `readJar(...)` it did not work the
other way around, or when not using `readJar(...)` at all.

This change treats an interface class as if it extended
`java.lang.Object` allowing the code that tried to fix it up to be
removed. That ensures consistent behavior irrespective of the order
in which the methods were called.

Bug: 394259916
Test: ./gradlew
Change-Id: Ie32f833e4e82bc19826d257b3508d43e44527b6a