Autoupdate of 3 submodules
*external/TestParameterInjector updated to 009dc6c835edc3ca7d7303c91a3fa2986dc5d8b4
set sdk_version

This change set sdk_version on TestParameterInjector, since this target
doesn't use any implementation detail of the framework.

Test: presubmit
Bug: 345013072
Change-Id: I62d1ec8a9ca0793e697bdd3da9ac4dfd352f42c3

*packages/modules/Bluetooth updated to 7b8ba731e856a4bc02b183ca14b0862ead6aeaec
update the depedencies

This change removes unused dependencies, changes the library of
testlooper and removes the platform_api setting.

This change will stop the repackage rules of the framework propagating
to this test.

Test: atest GoogleBluetoothInstrumentationTests
Bug: 345013072
Change-Id: I81a9de602dd39e608646547857c260067aede1b6

*platform_testing updated to 5da9275c213e1afd10c823b5be116345dbb4cce6
set sdk_version

This change removes the platform_api, and set sdk_version to controller
the API usage of this target. Before this change this target will build
against the implementation detail of framework, which is not necessary.

Test: presbumit
Bug: 345013072
Change-Id: Ia1641e31e20267445867baeaa2cdd5ec98f90f0a