Changed empty flow to flow of empty list
This CL modifies the default implementation for creative category
interactor. In the older implementation, we were returning an empty
flow, as a result when we were trying to consume the value in the
section view model, it would immediately complete without emitting any
values. After changing to current implementation, it would emit an empty
list which is handled appropriately in the view layer.

Bug: 393631840
Flag: EXEMPT bugfix
Test: Tested by building and installing AOSP picker and checking if
categories do appear normally

Change-Id: Ie8a309d86831acda6a32acc0b9f82824c88b9d1a