Do not show the custom tab title
The custom tab invoked in the context of the captive portal login
app is likely perceived as being part of the system UI. Its title,
however, is just the title of the HTML page it shows. This would
give the page a chance to show arbitrary text in a surface that
the user thinks of as being owned by Android.

It's better not to show the title at all. In the future, possibly
use a partial custom tab and show custom UI at the top, which
would be actually under the control of the portal app.

See the bug for screenshots of what it looks like before/after.

Test: manual
Bug: 394723319
Change-Id: If362a552b9b7f4d12b4ab3a5d89f2b7bfcbd5383