Add test for undefined type parameters
There are currently some historical signature files in
`prebuilts/sdk` which do not define type parameters. That causes issues
when attempting to migrate them, one of which is they end up being
prefixed with `java.lang.`. This change adds a test to show that
incorrect behavior. Follow-up changes will fix the issue.

Bug: 394789173
Test: ./gradlew
Change-Id: I799ce9baf0720d166a0f103396a1897d71594838