Report recoverable errors in ApiFile
Previously, `ApiFile` could only report an issue by throwing an
exception which would prevent any further issues from being detected
making it tedious and time consuming to resolve parsing issues, e.g.
when parsing an older signature file that is syntactically readable
but has some other issues. This change adds support for reporting
recoverable errors to the `Codebase.Reporter`.

It uses the new mechanism for reporting mismatches in the
`kotlin-style-nulls` settings as while that is likely to cause other
issues it is not immediately fatal.

This is needed for a follow-up change to use to report issues with
unqualified types, e.g. if `Foo` was unqualified then it should not be
assumed to be in `java.lang` package because `Foo` is not a valid
class in that package.

Bug: 394789173
Test: ./gradlew
Change-Id: I60482a60a10050edfc9cc5a3093dae1631615977