Stop caching TypeItems that have errors
Previously, a `TypeItem` that reported errors during its creation was
cached which meant that all other occurrences of that identical type
string containing the same error would use the `TypeItem` from the
cache and not report an error. That breaks the main purpose for
reporting errors (rather than throwing an exception) which was to
report as many errors in a file as possible to make it easier to
fix them.

This change adds an `errorCount` field that is incremented every time
an error is reported and the `CacheEntry.getTypeItem(..)` method checks
to see if any were reported during parsing and if it was the `TypeItem`
is not cached.

An alternative implementation that cached the `TypeItem` along with the
errors that were reported was considered but that would increase the
overhead of the `CacheEntry` and `TextTypeParser` to store error lists
which would affect the parsing of all files, even those without errors.
This implementation only affects the performance of files with errors
which should be very rare as in normal operation Metalava only reads
signature files that it has written itself and so should be free of
errors.

Bug: 394789173
Test: ./gradlew
Change-Id: Iddb39eb077b4b882b6a6197d2de9003ddd249e47