Add test cases sending/receiving enums of incorrect values
@Backing(type="int")
enum Foo {
   a = 1 << 0,
   b = 1 << 1,
}
Send/receive 12 and (a | b).

This "works" (it's still wrong, please don't do this if you have the
option!) in Java backends without complaints and with casting in C++ and
Rust backends.

The suggested way to accomplish this is to use constants instead of an
enum to send the values back and forth.

Test: atest aidl_integration_test
Bug: 393455995
Change-Id: I5e185da34b1fdaf5cec45eeece493d0392fdf9bd