Avoid moving old-gen object to young-gen in generational CMC
There were cases (typically during a full-heap collection) that some
object was being moved from old-gen to young-gen. This was happening
in few corner cases. The cases were either when we skipped compaction
(as the objects were too densely packed), or due to aligning down the
mid-gen. Moving an object from old-gen to young-gen causes correctness
problems as this way we miss dirtying the corresponding card if it
contains native-roots to young-gen.

In this CL, we correctly adjust the generation boundaries whenever
compaction is skipped. Also, mid-gen boundary is aligned up (instead of
down). This ensures that once an object is promoted to old-gen, it may
get back to mid-gen, which again gets consumed by old-gen, but never
gets back to young gen.

Test: ART_USE_READ_BARRIER=false art/tools/run-libcore-tests.sh --mode=host --variant=X64 --debug --gcstress
Test: atest CtsPackageManagerTestCases (on aosp cuttlefish device)
Bug: 392794369
Change-Id: I4472b34af3cbf409b2e0478e337c62ffc45c89f5