Add TypeParameterList.toSource(TypeStringConfiguration)
Previously, `SignatureWriter` would use `TypeParameterList.toString()`
which gave it no control over the formatting of it, specifically the
types in the bounds. That meant that the formatting of its bound types
were inconsistent with the rest of the types in the signature files.
This change adds a `toSource(TypeStringConfiguration)` to allow the
`SignatureWriter` to control the formatting. The supplied config is
passed down to `TypeParameterItem.toSource(TypeStringConfiguration)`
which also has a new `TypeStringConfiguration` parameter. They both
default to the existing `SOURCE_TYPE_STRING_CONFIGURATION` value.

A follow-up change will use that to control the formatting of type
bounds.

Bug: 395639482
Test: ./gradlew
Change-Id: I7e8744f25bf4228639a8ef1413c10e3222e63bbe