BUILD: add 'notap' tag to exclude 'perfetto_android_library' from TAP builds in google3.
This is a follow-up to aosp/3474111.
'perfetto_android_library' is by default build on TAP. If it depends on
'perfetto_android_jni_library' the build fails, because the jni dependency can't be built,
unless the root target is 'perfetto_android_binary'.

To fix the TAP build if exclude all 'perfetto_android_library'
as well as 'perfetto_android_jni_library'.

Bug: b/386353531
Change-Id: Ifb5bff4e0d1671f2fa8977e864f824f8c51cc674