Add ccdec integration tests
This CL defines simple integration tests with ccdec. The
tests validate the md5 checksum of the decoded bitstreams. As of now,
the tests use local test bitstreams under the project to run against.

The tests run by passing the command line arguments to be executed
with the main() function from the ccdec binary.

As there is no built-in parameterized tests available yet in stable
Rust (go/parameterized-testing), each input format has its own test in
the change.

Bug: 394638011
Test: Run test ccdec_test
Change-Id: I3969cb12505d38732e60da794b996b9dbe2a0f11