[scudo] Modify header corrupption error message (#126812)
Update the error message to be explicit that this is likely due to
memory corruption.

In addition, check if the chunk header is all zero, which could mean
corruption or an attempt to free a pointer after the memory has been
released to the kernel. This case results in a slightly different error
message to also indicate this could still be a double free.

GitOrigin-RevId: 9db0f91ceb6c5065fa7b119888e0fe15702300d3
Change-Id: I5bf7047ff60bb4a028e346358ccbcc9ad568abe7