Gather all historical api files for a version
Previously, `ConvertJarsToSignatureFiles` would use `PatternNode` to
find all historical jar files of the relevant surfaces and then map
from that to the corresponding `android.txt` file. However, it would
treat each surface separately. That meant it was not possible to
update `system` API files that were deltas on a `public` API file as
it did not have the `public` API file to subtract from the `system`
one.

This change adds `HistoricalApiVersionInfo` which encapsulates all the
API related files for a specific version, including the files for each
surface. This will be used in `ConvertJarsToSignatureFiles` in
follow-up changes to support generating API deltas.

Bug: 396431439
Test: ./gradlew
Change-Id: I05e2e7f0aea98235114a6677d6e773c00e86fc4b