Check hashes on extension AOSP interfaces
If an AOSP interface is attached as an extention to another interface,
we should still expect that implementation to using a known version of
that interface.

Test: atest vts_treble_vintf_vendor_test
Test: implement and attach a modified version of
android.hardware.authsecret to android.hardware.biometrics.fingeprint
service and verify vts_treble_vintf_vendor_test fails that service with
the expected error with mismatched hashes.
Bug: 396424232

Change-Id: I896cd943c312fd2053bb3e2d10763801f84c41e9