Load all contributing signature files for old API
Previously, when attempting to extract deprecation information from an
existing signature file it would only load the signature file for the
surface, e.g. for `public` it would load the `public` file and for
`system` it would load the `system` file. That causes problems when
generating a delta between the system API and the current API as
deprecated classes would be treated as deprecated in the public API
but not-deprecated in the system API.

This change avoids that by loading all files that contribute to an API
surface, e.g. for the system API it will load the public and system
files so it can extract the deprecation status for the public API as
well as the system API.

This changes the behavior of a test and makes it slightly more correct
but it is still broken. It will be fixed in a follow-up change.

Bug: 396431439
Test: ./gradlew
Change-Id: I33cd5d5ef920714609c8f0cc66408896f3db1200