Stop relativizing MatchedPatternFile.file
Previously, the `File` that was stored in `MatchedPatternFile.file`
was made relative to the `ScanConfig.dir` directory. The main reason
for that was to simplify testing but it makes `MatchedPatternFile` more
fragile and difficult to use as `file` may not be relative to the
current directory. That is illustrated in `ConvertJarsToSignatureFiles`
as it has to resolve the relative files against the directory it passed
to `ScanConfig.dir`.

This change stop relativizing the file, fixes up the broken tests, and
removes the unnecessary resolving from `ConvertJarsToSignatureFiles`.

Bug: 396431439
Test: ./gradlew
Change-Id: I05f381508fcdec43f58b1019f5cbb3fd83dbe794