Make EmittableDelegatingVisitor filter constructors
Previously, `NonFilteringDelegateVisitor` overrode the
`BaseItemVisitor.visit(ConstructorItem)` method instead of
`BaseItemVisitor.visitConstructor(ConstructorItem)`. That meant the
`skip(Item)` method was never called for `ConstructorItem`s. That in
turn meant that `EmittableDelegatingVisitor`, which overrides
`skip(Item)` to check the value of `SelectableItem.emit` would never
skip a `ConstructorItem` even if its `emit` property was false.

This was tested in the Android build using the
`gather-android-metalava-artifacts.py` script and did not seem to cause
any issues or differences in the current behavior. It was found while
working on a follow-up change.

Bug: 369078254
Test: ./gradlew
Change-Id: Ife72c0af71ebb8097ceefd568e83b7602a26840b