Add GBL Fastboot customization protocol
Add definition and implementation for the GBL Fastboot Customization
protocol.

Most methods are stubbed out and return EFI_UNSUPPORTED, and the
methods that are not stubbed out have simple implementations that are
mostly useful for testing against.

Includes selftest to verify the local session methods do basic
parameter checking and can provide a facsimile of a local session.

Bug: b/397442490

Change-Id: I236c6ee15218a2d230287649b3cb2bb24519e06c