Add SnapshotDeltaMaker to create delta between Codebases
The `ApiFile` can combine multiple signature files for a hierarchy of
surfaces. It consists of a file for the standalone surface at the top
of the hierarchy which does not extend another surface, followed by a
file for the surface that extends it and so on. It has some special
handling for classes to merge their members, annotations and super
class type.

This change adds a new `SnapshotDeltaMaker` which given a base
`Codebase` and a `CodebaseFragment` will create a snapshot
`CodebaseFragment` that will output the difference. It is basically
the reverse of `ApiFile` and as such can handle different super types
and annotations correctly.

As part of this it was necessary to modify `BaseItemVisitor` to allow
the hard-coded skipping of packages with `PackageItem.emit == false` to
be avoided. This was necessary because the delta was created by marking
everything with `emit = false` first, including the packages.

This will be used to replace `computeDelta(...)` in
`SignatureToJDiffCommand` and in `AndroidJarsToSignaturesCommand` to
generate signature files for API surfaces that extend another surface
as a delta on that surface.

Bug: 369078254
Test: ./gradlew
Change-Id: I8d86e005481a9260aaf29d340c210f0c948dfc88