Add normalize-final-modifier
While the Psi model does remove the `final` modifier from methods and
properties that are members of a `final` class the other models do not
and there are some inconsistencies in the model, e.g. when duplicating
a final method from a non-final class to a final class. They cannot be
fixed in the model because it would change Android/AndroidX signatures
and updating the former is very expensive.

This change adds a `FileFormat` property that controls whether the
`final` modifier is normalized when writing the signature files. This
will make it easier to migrate existing signature files. At the moment
it only affects methods not properties but a follow-up change will also
fix it for properties.

Bug: 396431439
Test: ./gradlew
Change-Id: I7db2a072e23a156a60eb4d697643fc7d20c0ea18