Add AuthMgr BE storage traits.
This CL introduces the storage traits used by AuthMgr BE logic and the
data structures associated with the traits.
We plan to separate out the storage functionality into a super storage
trait, persistent and non-persistent and keep the common functionality
across the two storage  types in the super trait.
We do not add the non-persistent storage in Android 16. However, we plan
for it for the completeness of the AuthMgr BE logic.

Test: will be added in the upcoming CLs
Bug: 369145328
Change-Id: Ie9c1df6f68e8e44184354d09b2ec829a621a5fb2