Add @Deprecated annotation when setting deprecated
Previously, when `ConvertJarsToSignatureFiles` read the existing
signature file to copy any deprecated status over to the jar `Codebase`
it called `setDeprecated(true)` but did not add a corresponding
`@Deprecated`. That lead to an inconsistency in the annotations between
the base and jar `Codebase`s when computing the delta.

This changes forces it to add an `@Deprecated` annotation as well which
fixes the delta in the test.

Bug: 396431439
Test: ./gradlew
Change-Id: I5840aea6bfb0d1654e4bd5c743e8e55a66e2d166