Check native-roots (to young-gen) of all objects during marking
In aosp/3489378 we (rightly) changed the logic to align-up (instead of
aligning down) the mid-gen-end boudary to page-size. However, this
opened a window towards the end of mid-gen where some young objects get
pre-maturely promoted to mid-gen in PrepareForCompaction(). This is fine
in general, but means that our 'native-roots pointing to young-gen'
logic during marking can't be restricted all objects outside young-gen.
We have to now consider all objects being visited.

Bug: 343220989
Bug: 397381582
Test: ATP test infrastructure
Change-Id: I083dac9a7396b152138987254f0934e3d1a1b3b8