Do not check fence state for regular waitfences calls
We started waiting for the correct fence state with recent
changes, but that doesn't support the case where only some of
the fences were signalled and we're getting some test failures.
Until the underlying issue is identified this CL will limit
the newly added cpu side cv wait only to external waitForFence()
calls to restore the previous behavior.

Bug: 397501277
Test: CI && run apps on TV api level 31
Change-Id: I349b51a3f9c7929802a9c1d62ea96cf5df50cc6a