Add test to show deprecated copied from existing signature file
The test improves the test coverage of `AndroidJarsToSignatureCommand`
and in the process revealed that it was broken by the addition of a
call to `Codebase.freezeClasses()` in `JarCodebaseLoader`. This fixes
that by passing `freezeCodebase = true` to `loadFromJarFile(...)`.

The test also reveals some issues with the handling of deprecated
status which will be fixed in follow-up changes.

Bug: 396431439
Test: ./gradlew
Change-Id: Ic4a6b75e20ee370b35d170857b1b5035dc145d56