smptest: Add a check for active cpus.
build-config-kerneltests currently specifies that this test needs 4 CPUs
and should only be run on systems that have at least 4 CPUs. The test
used to fail if fewer CPUs were online, but when the test was modified
to test all active CPUs it lost the ability to detect that CPUs that are
meant to be active never activated. Add a new build flag to specify how
many CPUs are expected to be online, with a default value of 4 to match
build-config-kerneltests, and a test to check that at least this many
CPUs are active.

Bug: None
Change-Id: I2d1e17339d83126c49fd948635aaba6d216b1889