kernel_test: disable the TestUser64Bit test
The test's bitness appears to not correlate
to the device userspace bitness... thus sometimes fails,
even when it shouldn't.

We don't actually lose any real test coverage since NetBpfLoad
will already fail to init and thus device boot on such setups.

Maybe we should just delete this?
For now leave it as is as documentation of intent.

Bug: 397352461
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I19596455d983643b5b0638055353b32fadfdf2c3