Move the excluded tests from the TEST_MAPPING to the XML configuration
Coverage run disregards the TEST_MAPPING completely as it is not meant
to be integrated with it which means that if we disable a test through
the TEST_MAPPING then it will still keep running through coverage runs.
By moving to the AndroidNetTest.xml then the excluded filter will be
respected regardless of where the tests are executed.

Bug: 398135623
Change-Id: I02ffed524fb760ccf7720d3989d0dd04459e6601