Fix the crash in testCaptivePortalUsingCustomTabs_multiNetworkNotSupported.
We call DeviceConfig.getDeviceConfigPropertyBoolean in the production
code to read a boolan value which indicates whether we can use any other
browser in the system if the default browser doesn't support custom tabs
and multi-network feature. However, the test doesn't have the proper
READ permission, which causes the crash.

Mock the getDeviceConfigPropertyBoolean in the test code and return the
default value directly instead of reading the real device config.

Bug: 397658274
Test: atest CaptivePortalLoginTests
Change-Id: I19dcfbb99a6a81db8e3739a5ce3970236ceb3fba