Stop hiding android.annotation classes
Previously, it was necessary to hide `android.annotation` classes in
tests that used them as otherwise their definitions would be emitted to
signature files and stub files introducing noise, making it difficult
to see the true purpose of the test. A preceding test avoids that by
marking them as `emit = false` by including the package in the default
`skipEmitPackages` parameter. That makes the tests more realistic as in
practice the annotations will be provided on the classpath, in which
case Metalava will mark them as `emit = false`.

Bug: 398837120
Test: ./gradlew
Change-Id: I1ceb7bf6760a3ec0c81b269cd73f62938aafa570