Do not emit android(x).annotation classes in tests
Previously, each test that did not want to emit classes from the
`android.annotation` and/or `androidx.annotation` packages had to
specify that explicitly (by using `androidAnnotationHide` and/or
`androidxAnnotationHide` respectively. This change stops them from
having to do that by adding `android.annotation` and
`androidx.annotation` to the `skipEmitPackages` default. A couple of
tests needed to emit the packages so overrode the default and one
test emitted it but did not need to so the test was fixed to avoid
that.

Follow-up changes will remove `androidAnnotationHide` and
`androidXAnnotationHide`.

Bug: 398837120
Test: ./gradlew
Change-Id: Ic787fae3c1e15a73432bf846ae529639749c98b6