Dedup skipEmitPackages default value
Previously, the default value of `skipEmitPackages` was duplicated in
`DriverTest` and `AbstractStubsTest`. This changes the type to make it
nullable and sets the default in both places to `null`. The
`DriverTest.check(...)` method falls back to the default if the
parameter is `null`.

This will make the follow-up change which modifies the default simpler
as it will only have to make the change in one location.

Bug: 398837120
Test: ./gradlew
Change-Id: Id5ce1fa6850130b45d898a5b68c4d8a06e6e4312