Fix mid-gen-end alignment when deciding to skip compaction
Ensure that mid-gen never goes lower than its previous level when
skipping compaction. This is necessary for correctness as the previous
mid-gen-end is during marking to decide what the young-gen range is.
If later in PrepareForCompaction() we decide to lower it, then it runs the risk that we may miss dirtying the objects containing references to the diff between new and previous mid-gen. Effectively, this is same as
moving some objects back to young-gen, which can cause memory corruption
due to missing card marks.

Bug: 343220989
Bug: 397381582
Test: ATP test-infra
Change-Id: Iabb6a261f36a5ef46d390c97d5933689191bf8da