Preserve annotations when method is parameterized
Currently when JunitParamsRunner parameterize a method, the original
annotions are not passed down to the child methods. This causes
filtering by annotation for parameterized tests to not work.

This change creates a Description object for the paremeterized methods
with annotations included. createSuitDescription has been used instead
of createTestDescription to properly pass the annotations. However the
resulting Description will be identital as before.

Bug: 394914027
Test: atest MicrodroidHostTestCases.CTS, presubmit
Change-Id: I51d52abf55b40021cd7ea3811183e660e23421cd