Update gather-android-metalava-artifacts.py to work with latest build
Some recent changes to the Android build broke the script as various
artifacts were no longer copied to make directories by default and
were copied into `dist` instead. The script does not want to use the
`dist` files as that would require building all files for the `dist`
goal, which would take much longer. So, this script is updated to
pick more of the files up from `out/soong/.intermediates/`.

One problem with using the files from Soong is that Soong tends to use
the same basename for each artifact type, e.g. `api-versions.xml`, for
the API history files created by Metalava. They would collide if they
were just copied into the output directory. The rules for copying them
into make directories already dealt with that.

So, this change replaces the simple build file string with a `Target`
class that allows for an alternative file name to be specified in the
output directory.

Bug: 398805908
Test: run in Android.
Change-Id: I247a0c89780fca546cc1b94bfc08c5a234d2b7ce