XML parser for ARA rule engine for OMAPI server.
The Global Platform Open Mobile API (OMAPI) server requires an
implementation of the Global Platorm Access Rules Application, which
reads access control rules from the secure element and enforces them.
The rules from the SE are ASN.1 BER encoded but it's useful to augment
these rules with additional rules provided by the Android platform, to
allow device makers to specify rules for platform components such as
KeyMint and Weaver.  This CL adds a module that parses XML-encoded
rules that conform to the ARA rule data structure.

The platform-provided XML-encoded rules will be combined with the rules
retrieved from the SE in most cases, but during early boot the platform
rules will be used alone, before rules are retrieved from the SE to
avoid delaying the operation of the platform components.

Note that the build flag mentioned below has not been created yet and
is not referenced by the code.  That flag will eventually control whether
this code is built and included in Android, so it will be used primarily
in device.mk files.

Bug: 380331467
Test: atest libse_service_test
Flag: build.RELEASE_NATIVE_OMAPI
Change-Id: If2aa7aad239c0ca62cc317f3e57324f08664b4ed