Add ApiFlags and wire into AnnotationInfo
Previously, Metalava had no concept of the flags and just relied on the
`--revert-annotation` to handle flagged APIs. However, that is no
longer sufficient so this adds basic `ApiFlags` support and wires it
into the `AnnotationInfo`.

Follow-up changes will use this to replace `--revert-annotation` with
the ability to configure flags via `--config-file`.

Bug: 399846595
Test: ./gradlew
Change-Id: I2990b521042f36cd3bcc5dc23b235024efe5149a