Only create snapshot with reverts when required
Previously, `snapshotIncludingRevertedItems(...)` was called if any
`--revert-annotation` options were specified on the command line, even
if they did not actually match any annotations. This changes that to
be called only when a `Codebase` contains at least one reverted item.

Preparation for when items can be reverted via configured flags and
not just `--revert-annotation` options.

Bug: 399846595
Test: ./gradlew
Change-Id: I0c56fc541608e5e32f37d84a7357dda9099d5e48