Add DefaultAnnotationManager.couldRevertItems()
Previously, a couple of places in `DefaultAnnotationManager` checked
whether `revertAnnotations` was empty (or not). That was necessary to
ensure correct behavior when reverting newly added items or items that
have moved from a more restricted to more general API surface. However,
it was not the presence of `--revert-annotation` options that caused
that but rather that an item can be reverted. So, this change abstracts
the check of `revertAnnotations` behind a `couldRevertItems()` to make
that clearer.

Preparation for when items can be reverted via configured flags and
not just `--revert-annotation` options.

Bug: 399846595
Test: ./gradlew
Change-Id: I9f7213512b5fead711ca49cfe98b27874525899e