Create ApiFlags from ApiFlagsConfig
Previously, `ApiFlags` was only created from `--revert-annotation`
values. This change adds support for also creating them from
`ApiFlagsConfig`. It is an error if both are provided.

Some additional test configurations were added to the
`ParameterizedFlaggedApiTest` to run the same test as the existing ones
but using flags instead of `--revert-annotation`s. There was no need to
provide a flags version of `KEEP_ALL` as that is the default behavior
when no flags or `--revert-annotation` options are provided.

Bug: 399846595
Test: ./gradlew
Change-Id: I98fb5039dfcaad5fa7f5a4d73e7fc2ec454c8c95