feat: Add device registration event support
This commit adds the ability to register events when a QEMU device
becomes available in the guest. This allows for triggering behavior when
a device is registered.

It introduces the `DeviceRegistrationListener` class,
which allows you to register callbacks that are invoked when a new
device of a specific type is registered with the `ConnectorRegistry`.

Change-Id: Id4f59cb0b0d3b773837cf13b2d92382312244cd4