Fix problem with scripts/refresh-baselines.py metalava
Running `scripts/refresh-baselines.py metalava` caused failures in
tests for commands due to inserting colors in the help they printed.
That lead to lots of additional tests being added to the baseline
files. After bisecting it turns out that it was caused by
https://r.android.com/3519674 which replaced a call to
`Test.setEnvironment(...)` with `Test.environment.putAll(...)`. The
difference is that the former cleared the existing environment but the
latter did not. While that did not affect normal tests it broke the
refresh script which clearly runs with slightly different environment.

This change fixes it by explicitly calling `setEnvironment()` to clear
it before adding in any custom environment variables. This is done by
the `MetalavaBuildPlugin` as that runs first, before the code in
`metalava/build.gradle.kts`.

Bug: 399846595
Test: ./gradlew
Change-Id: Ied235e91bf124d06c028db816e5b6d981d8c461b