Create TurbineValue to own the value handling in Turbine
Turbine has two representations of a value, a `Const` which Turbine has
evaluated (or retrieved from a jar) and the optional `Expression` which
is available when the value is from the source. Metalava needs both
when available and there are quite a few places where they are passed
around together. Unfortunately, that leads to the code for handling
them to be spread around too.

This change adds a `TurbineValue` class to encapsulate them and moves
the `extractAnnotationDefaultValue(...)` method into it. Future changes
will move other related code in there.

Bug: 400664604
Test: ./gradlew
Change-Id: Ia46717fa25e9554da368520318795f842016d611