Build test_module_config modules' module_info_json in soong
To be able to run these tests in soong-only builds.

This cl produces the following diff:
https://paste.googleplex.com/4917659043627008

The diff is enitrely additional host dependencies. Without the line
adding the ExtraHostRequired modules it becomes a diff of missing
host dependencies. Of the two I chose the extras. It's difficult to get
the required list exactly matching make.

Fixes: 400731860
Test: The diff + atest CtsWifiTestCases_NoNonMainlineTest
Change-Id: Idbbfbf7f74eb2eeaf68d64c0bcfbc97ddae3e2ea