Autoupdate of 2 submodules
*art updated to a7f2b44e843cdf9033a22bdb1b30f8f25a4ea6f5
Add a new API getExecutableMethodFileOffsets

This new API overloads the existing API of the same name and accepts
"Executable" as the argument type (rather than "Method" for the existing
one). This allows it to handle things like a constructor. The API reuses
the existing implementation which already handles an Executable
properly.

Bug: 400457896
Test: TH
Change-Id: I754e798da5b041edc3f9dcfe4a3476fea2a234bf

*libcore updated to 5a0520c088b5c30faae7da892744fda688c692cc
Merge "Add a new API getExecutableMethodFileOffsets" into main