Symbolically link existing files into the test directory
Previously, `ExistingFile.createFile(targetDir)` would just return the
existing file. That breaks its contract as it expects the file to be
created in the `targetDir`. This change fixes that by creating a
symbolic link in `targetDir` to the existing file.

Avoids problems in follow-up changes that make more use of this.

Bug: 400664604
Test: ./gradlew
Change-Id: I90701f09a952600848b66b2086bee8e7c7f18911