ui: Use SliceRef in debug/query tracks' details panels
This allows the SQL query representing that slice to be copied to the
clipboard, which makes it more consistent with most other details
panels.

Acknowledge this query has little use outside of the UI as debug/
query tracks usually reference temporary UI only tables, but the
consistency argument still holds.

Change-Id: Ib2fa9837ad141b012f529e13fa40751890098af7