ui: Add support for uniondataset type widening
Given a union of {foo: NUM} and {foo: NUM_NULL}, currently the foo
field will be removed as NUM !== NUM_NULL.

This CL adds automatic combination of nullable and non-nullable types,
widening the resulting type to the nullable version.

E.g. the resulting type of the union will now be {foo: NUM_NULL}.

Change-Id: Ic238fc0a21e94957a6a8d2e7e33123124a87442d