Autoupdate of 2 submodules
*external/mesa3d updated to 6f036c5c303805aa06df5398d78f31330cf3c5e7
Merge "Wrap vkEnumerateInstanceExtensionProperties for host" into main
*hardware/google/gfxstream updated to 849854f7ece4530898231c3624af51448901c851
Serialize vkEnumerateInstanceExtensionProperties on linux

This function crashes randomly on XR emulator on linux with
AMD GPUs and avoiding multithreaded operations seems to be a
good workaround until the underlying issue, or the driver bug
is fixed. Since the call is expected to be called rarely, we
don't limit this to aemu or amd.

Bug: 401005629
Bug: 373601997
Test: run XR emulator 20 times, play videos on chrome
Change-Id: Id11eff65b8a2ef33618d903e4499b4856a05209f