ui: Fix binder plugin parsing strings for SQL
We previously used parseFloat to determine if a string
is a number. Unfortunately, it makes a best effort at
number conversion. So if you have a string
(in this case a thread name) that started with a number e.g
5-Thread. parseFloat will return 5 and we'll fall-through to
a BigInt conversion which will fail.

Used a more robust conversion that first checks if the numeric
conversion matches the original string conversion.

Test: Manual
Change-Id: I9650b791ad7989609851d44e1f3c8701b412bd16