Simplify computation of command line classes
Previously, `createAllClasses(...)` would be supplied with a map from
`ClassSymbol` to `SourceTypeBoundClass` for all source classes provided
on the command line. It would then filter out nested and inaccessible
classes and create `ClassItem` instances from them and their nested
classes.

The creation of that map would take the map of all source classes and
remove those that are not in the `extraUnits` that were loaded from the
source path. However, that was hard to reason about and required the
ability to find the `CompUnit` from `TyDecl`.

This change simplifies and optimizes the two methods. It creates a map
of top level, accessible, command line classes, by simply excluding any
nested, inaccessible classes whose source path is not in the paths
provided on the command line.

Bug: 400664604
Test: ./gradlew
Change-Id: I3c1b01be1eaf0c244835f8c0ed8cf590aa3990e1