Preserve duplicate methods when loaded from a single file
Previously, `ApiFile` would replace a method that was added from one
signature file with a duplicate method added in a following signature
file. That was necessary in Android to allow a system API to modify a
method that was added in the public API. However, it causes problems
when trying to reformat some AndroidX signature files as they contain
duplicate methods within the same class.

The reason they do that is because in Kotlin it is possible to create
am inline value class `V` which "holds" a single property of some type
`T`. In Kotlin types `V` and `T` are distinct, you cannot cast one to
the other. However, you can "create" an instance of `V` from `T` and
retrieve `T` from its property in `V` just as you would a normal class.

In Java however, `V` is represented as `T`.

So, in Kotlin you can have the following:
```
   fun method(p: V) {}
   fun method(p: T) {}
```

However, if both types are mapped down to `T` in Java those methods
would collide. Kotlin works around that by mangling the name of the
first method. The consequence of that is Java cannot easily call that
method.

Currently, when those methods are written to a signature file they use
the Java representation so they both end up with identical signatures
When they were loaded in from signature files the duplicates were
removed.

This change will keep the duplicate method when loading the `Codebase`
from the first file it loads but discard them if they are replaced by
a method in a following file. That maintains the existing behavior
needed by Android.

Longer term we will need a away to represent inline value classes in
the signature files, e.g. `public inline value class ....`. However,
that will require using the Kotlin view of them not the Psi view.

It adds tests for loading duplicate methods from a single file and
multiple files and also a round trip test to ensure that they are
preserved when reading and writing.

Bug: 402049952
Test: ./gradlew
Change-Id: I1bfe1250df2dd48a282143e3b3079575fa9aacc8