Convert concise-default-values to include-default-parameter-values
Previously, the `concise-default-values` was used to differentiate the
version "3.0" representation of default parameter values as arbitrary
expressions and the version "4.0" representation of just using the
pseudo-modifier `optional` to indicate a parameter that has a value.
That is no longer necessary as version "3.0" is no longer supported.
However, there is still a need to differentiate between a file format
that can contain default parameter values for Kotlin and one that
cannot, or at least does not have to, for Java.

Rather than remove `concise-default-values` this just co-opts it to
serve that purpose; renaming it to `include-default-parameter-values`.
The `SignatureWriter` was modified to not write out the full default
parameter expression as that is no longer needed.

Bug: 402049952
Test: ./gradlew
Change-Id: Ie8147dd78898bc4a64647b0a45af58c3fedca4cd