Simplify ensureCompanionJvmStatic
Removes some old commented out code (it was to check @JvmField on properties, which is done with ensureCompanionJvmField).

Removing the commented code allowed collapsing the nested ifs into one.

Removes the is public check because the API lint runner will handle only reporting for public API.

Removes the is kotlin check because ensureCompanionJvmStatic is only called when the method is kotlin.

Bug: 401569415
Test: KotlinInteropsChecksTest
Change-Id: Iff838434f9894576eb69c6669140800850534c08