Dump references when found invalid without type-of
Once we know that some reference in a given object is invalid, calling
PrettyTypeOf() on all references is bound to fail. OTOH, it is useful to
know at what offset was a certain reference was. So dump that info for
all references, without type-of, when an object is found to have invalid
references in it.

Test: art/test/testrunner/testrunner.py --host
Change-Id: I898d86ad13e12545ad483f0c2e6de0c85cbf76f5