[ot-daemon] resets all runtime states after leave()
Most likely the ot-daemon runtime states (e.g. mState) should have been
set to the correct value with the otThreadDetachGracefully() API, but
there are no guanratee on the timing and it's risky to depend on it.

So to be safe, this commit explictly clears all ot-daemon runtime states
right after resetting the OT instance at the end of the leave() API.

Bug: 323301831
Bug: 402563812

Change-Id: I8ecfeb8de36f984eccf74dc6e2eaf222612eb51d