Autoupdate of 3 submodules
*external/openthread updated to e92530a144208402b9603a29b8f8dd8b1cc8ba2b
[srp-server] fix fast-start mode disruption when device role changes (#11330)

When in fast-start mode, the SRP server will be disabled and then
immediately enabled again if the device transfers from Child to Router.
This commit fixes this issue by also matching the MLE-ID with the SRP
server address in the Network Data.

(cherry picked from commit 1dd7f11ae7c2688050496202729e7de1f703672c)
Change-Id: I235324e36482bcb59f359971221e2ae6009f7372

*external/ot-br-posix updated to a4d7bfe2079f1374f12d5fe15824c856601260a2
[epskc] fix inconsistent ePSKc states update

The current the code has an issue that when the
activateEphemeralKeyMode() callback is invoked, there are no gurantee
that the ephemeral key state and lifetime are updated to the corrected
values.

Bug: 401859577
Change-Id: I3092292d45c34d4540352a9684ee2efc62f59c74

*packages/modules/Connectivity updated to 7cb5078ba1824210f777105c242fb67790c0bd48
Merge changes from topic "thread-router" into main

* changes:
  [thread] refactor the settings module
  [thread] Support non-BR mode