Autoupdate of 3 submodules
*external/abseil-cpp updated to 2a1813e03aa8ca730d0eb6c76aa2e56e929b5cea
Expose abseil visibility to more external projects

private-join-and-compute and anonymous-counting-tokens package in forks
of abseil to depend on. The forks cause problems when updating protobuf
because protobuf now depends on this abseil. Expose visibility so
they can depend on this canonical abseil directly.

Bug: 329747255
Test: Presubmits
Change-Id: I138c7af7b50206d4549c8890bd7d72d059cf84d6

*external/anonymous-counting-tokens updated to 38af3dd2e1ccf5b0bcbce592e6de9ba191056375
Use canonical ABSL

The forked version of absl causes problems when updating protobuf, as
protobuf now links against the canonical version of absl.

Bug: 329747255
Test: Presubmits
Change-Id: I915f935a24f60f7c2bd5bd79582f5eb25676988e

*external/private-join-and-compute updated to 19b20109795a393f972fa1eb132597d4794f1da4
Use canonical ABSL and remove local fork.

This fork causes problems when updating protobuf, as protobuf now
links against the canonical version of absl. Remove it.

Bug: 329747255
Test: Presubmits
Change-Id: I9c14df489f2e1f182433712090d4734d365b2349