Rewrite CommonParameterizedValueTest to make it more exhaustive
After further investigation of the handling of values within the model
it was discovered that the inconsistencies are worse than expected. The
previous incarnation of the test did not exhaustively check the all the
value expressions across all the sites where values could be used
making it difficult to map out where the inconsistencies lie.

This change rewrites the test to rectify that. It adds a list of the
different `ValueExample`s to test and replaces the hand written classes
that were tested with classes generated from the `ValueExample`s. It
creates tests for all `ValueExample`s across all suitable places where
a value can be provided.

The test was set up so that they would all pass on Psi as that is the
standard, even when its behavior is inconsistent. Follow-up changes
will try and address those inconsistencies.

At the moment this only tests on java but it will be expanded to test
Kotlin too at some point although the focus of the work for now is java
because the immediate goal is to try and improve Turbine.

This adds type parameter to `TestCase` and uses it in `Expectation` as
follow-up changes will test expectations of types other than `String`.

Bug: 354633349
Test: ./gradlew
Change-Id: I32dc8658d93ac78b699d005d638db6b3f021c055