Autoupdate of 2 submodules
*external/compose-hero-benchmarks/poxedex-compose updated to e777ba83c95e103a5be4f2b24511ac84deb74159
Updates navigation dependency to avoid a lint crash

We are enabling the ObsoleteLintCustomCheck lint error in AndroidX - older versions of the navigation dependency have an issue that will result in this error. Update the dependency to use a fixed version that does not have this issue.

Test: lint
Change-Id: Ib522f6ccc8eb60665d869b5d16eb473de7fde282

*frameworks/support updated to efcc0337cc186d780585353355c067b7686e93c5
Merge "Revert "Support for common source sets in the `CmoposableStateFlowValueDetector`."" into androidx-main