Autoupdate of 2 submodules
*trusty/device/arm/generic-arm64 updated to bd2d37d443739bde4bd9786060c2dc5995902964
project: qemu: Add hostcmds directory

Add a new directory for scripts that run on the host
and interact with android running in qemu. This directory
gets copied to the build output:
   build-root/<build>/hostcmds

Bug: 395975104
Change-Id: Ia1d19b46d7fa0149ff67e100b2b5ddd8ffe97bbd
Signed-off-by: Jay Monkman <jmonkman@google.com>

*trusty/vendor/google/aosp updated to d760ad7a56e323047444113f1ccaf5b15c98acb8
trusty: scripts: Add hostcommandtest

hostcommandtest allows running a program on the host that can
interact with android runnning in qemu. This is similar to
androidtest, except that the test to run runs on the host
instead of inside the emulator.

Bug: 395975104
Change-Id: Ie649e3b9fa4ce95e9b2c1d66ffd8840ca243d889
Signed-off-by: Jay Monkman <jmonkman@google.com>