Allow APEX versions in the OAT header to be overriden.
APEX versions are stringified mtime's of the APEXes installed on device,
for detecting samegrade placebos (go/samegrade-placebos). In the case of
Cloud Compilation, we cannot put the APEX versions in the OAT header
because the OAT file is generated on the server. Instead, we put the
APEX versions in a separate file (SDC file) when the SDM file is being
installed on device. Therefore, we need a way to override the APEX
versions in the OAT header.

This is a no-op change.

Bug: 377474232
Test: m test-art-host-gtest-art_runtime_tests
Change-Id: I69658b61942979ff7f56e4908ce3204bd0a2116e