Fix build when the global default is _FORTIFY_SOURCE=3.
The config.h files define _FORTIFY_SOURCE, so Android.bp needs to
undef it to avoid a "macro redefined" error. This seemed to make more
sense than changing the config.h files to 3 because (a) the host libcs
don't support it anyway and (b) the config.h files might get regenerated
on update.

Bug: http://b/291762537
Change-Id: I4e1b36e5cef27ef15442225499e6935437ebb996