Support creating LiteralValues in Turbine
Adds support for creating `LiteralValue`s in Turbine.

The `Value` tests are quite lax and set up to fail only when a non-null
`Value` is provided and does not match the expected `Value` (or a
non-`ValueProviderException` exception is thrown). Otherwise, they
throw an `AssumptionViolatedException` which causes the tests to be
recorded as ignored. That meant it was necessary to manually verify
that the literal tests are not ignored on Turbine. They are not and
they pass.

Bug: 354633349
Test: ./gradlew
Change-Id: I47063755d3ed115603e25bb05d22a2ea7db35e8a