Add support for creating Codebase dependent expectations
Previously, the `Expectation`s produced could only return values that
could be created during test setup. However, that is not sufficient for
testing expectations that require a `Codebase`. Follow-up changes will
add `Value` objects that reference a `TypeItem`. A `ClassTypeItem`
requires a `Codebase` to be provided on creation.

This change adds a `CodebaseExpectationProducer<T>` lambda type and a
`codebaseExpecations` builder which allow for creating an `Expection`
that will invoke the lambda with a `Codebase` and return that as the
expected value.

This is not currently used but will be used in follow-up changes.

Bug: 354633349
Test: ./gradlew
Change-Id: I1b7bcb1fce96951c728ed6ab7a724c6bf5ad2dd5