unittest-rust: Support `E: !Display` in `assert_ok`
Widen the types `assert_ok!` supports to include `Result<_, E>` for any
`E`. It previously only supported `E: Display`.

Through some abuse of autoderef, our error messages will still use the
`Display` implementation if it exists, the `Debug` implementation if it
doesn't, or show the `type_name` if neither exist. (So long as the point
where the macro is invoked isn't overly generic, which is unlikely in
tests.)

Bug: 369143019
Test: See comments
Change-Id: Id60e6ca0f722e17409402da3f9896e6bc90a86a3