dev: virtio: vsock: Move init_vsock out of PCI-specific TrustyHal and rename it
This moves the init_vsock function to vsock.rs renames it to
vsock_init. This commit also adds type annotations in parts
of the PCI vsock driver where rustc can no longer infer the
types. It also makes the changes TrustyHal in init_vsock to
a generic `H: Hal` to allow reusing it for the virtio-msg
driver it.

Bug: 379677575
Change-Id: If2b60447a5786d404a5e431f4cfe8fd68acab370