Autoupdate of 2 submodules
*build/make updated to 70487557f703397c46495586986e7124635dd046
Merge "aconfig: add finalized list in fake class" into main
*platform_testing updated to c7296cda4bc77a7caa5a0973cc137138ab8c1e4f
SetFlagsRule: check whether flag is finalized

If a flag is finalized, then the SetFlagsRule cannot change the value of
the flag. Thus in this case if the test requests to set the flag value
as false, then this test should be skipped, since the flag is finalized
on the current device.

Test: atest FlagJUnitTests FlagJUnitHostTests
Bug: 402588368
Change-Id: I7aceb849343e8169ce92b88894b2d5793d6189fe