Add tests for AnnotationItem.toSource() value handling
The Psi implementation of `AnnotationItem.toSource()` uses its own code
to format values which differs from the format used by its
implementation of `AnnotationAttributeValue`. This change adds tests
for that formatting. It does so by formatting the whole annotation and
then extracting the value from that.

It necessitated adding another `ValueUseSite`.

Bug: 354633349
Test: ./gradlew
Change-Id: I12a06ad880de975af861c6e4f8f70cfd13436366