Autoupdate of 2 submodules
*frameworks/native updated to 0fa903349f312d84af88a944e25c595a1f293c06
Merge "Add get() method to declare_binder_enum" into main
*system/tools/aidl updated to 8590b191172df0e72c365d07c46e36c28c8dc98f
Test the use of get() for enums

Users want to be able to shift bits with enumerators in Rust like they
do in other AIDL backend libraries. This provides a way to do that by
first returning the value of the backing type (integers).

Test: atest aidl_integration_test
Bug: 393455995
Change-Id: I2c901abe4c787b58814e5bd73a082a32a3b87930