Break CommonParameterizedValueTest.kt into separate classes
Previously, all `TestCase`s were run in sequence which meant that the
test run length was larger than any other test making it part of the
critical path. That was especially true when run in the
`metalava-model-snapshot-testing` project which runs all tests for all
models together. It would also only get worse when the tests are run
for Kotlin and signature files too.

This change, makes `CommonParameterizedValueTest` the base class for
5 separate classes, one for each `ValueUseSite`. That allowed the
`TestClass.toString()` method to be simplified as it no longer needed
to differentiate between the different `ValueUseSite`s, which it did in
both the `testCaseName` and `testClass.className`. This replaces both
of them with `valueExample.name`.

Bug: 354633349
Test: ./gradlew
Change-Id: I004c065e1aa680195568eb4158e668132f4bba4b