Remove min_sdk_verison property
r-launched-apex-module already sets the min_sdk_version value, hence
there is no need to override it.

Changing apex defaults from s-launched-apex-module to r-launched-apex module as statsd was launched in r, which keeps min_sdk_version to 30.

It does remove compressible:true property, but as discussed in the comments, it is fine as this is just a fake of statsd.

Flag: EXEMPT removing overriding property is not flagged
Bug: 361092707
Test: m art_fake_com.android.os.statsd
Change-Id: Iee5e64561d491104c976d2ee7bf804d72c671299