Handle non-constant field expectations in ValueExample
Previously, it was the responsibility of a `TestCase` that access a
field's value to ensure that it correctly expected no value to be
provided for a non-constant expression. This change moves that into
`ValueExample`. Simplifying the `TestCase`s and making them more
consistent.

Preparation for deduping the logic in the different `TestCase`
subclasses.

Bug: 354633349
Test: ./gradlew
Change-Id: Ide2775a8b5f5c023516b8f53d4335276e9922bbb