Add TestClasses to TestCase
Previously, the `TestClass`es were all created when creating the
`TestCase`s which meant the cost was paid even if a test was not run
and it also coupled the `TestCase` to a `ValueUseSite` specific
`TestClass`. This adds the `TestClasses` class which creates the
`TestClass` appropriate for a `ValueUseSite` on demand. It creates
a single `TestClasses` for each `ValueExample` passing it into to
each `TestCase` for that `ValueExample.

Follow-up changes will create a single `TestCase` for each
`ValueExample` instead of one `TestCase` for each
`ValueExample`/`ValueUseSite` pair.

Bug: 354633349
Test: ./gradlew
Change-Id: I0be0a8f3caa368e2d8d84c03d7cf999c150b6fd5