Add support for filtering arguments for parameterized test
Previously, there was no way to filter the list of combined argument
sets for a parameterized test once it had been created by combining the
implicit arguments from the model providers with any additional ones
provided by the test using `@Parameterized.Parameters`. That meant it
was not possible to ignore certain combinations that did not make
sense. Instead that logic had to be done in the test method which
wasted time creating and running a test that was never going to do
anything.

This change adds the `@ParameterFilter` annotation which can be used to
annotate a public static method for a test run by a subclass of the
`BaseModelProviderRunner` that will be called with each set of
arguments, including the implicit `CodebaseCreatorConfig`. That is used
in `BaseCommonParameterizedValueTest` to ignore anything except
`InputFormat.JAVA`.

In future, once other `InputFormat`s are supported it will be tweaked
to ignore jar based tests for anything other than `InputFormat.JAVA`.

Bug: 354633349
Test: ./gradlew
Change-Id: Ibe600f1399d2e9626d9e5415fc51048f0e4463e1