Allow Configurables in select providers
Providers are generally forbidden to contain Configurables, as
Configurables are expected to be evaluated in the context of the
current module, and Providers could send it to another module where
it could be evaluated in a different context.

However, overrides modules contain properties that are propagated
to the base module and are expected to be evaluated in the context
of the base module.  Relax the ContainsConfigurables check by adding
a struct tag to identify fields in Providers that are allowed to
contain Configurables.

Test: TestContainsConfigurable
Change-Id: Ia6278ca31e38ac2f209bce40df9e68305cb96252