Move suitableFor filtering to testCasesForValueUseSite(...)
Previously, `ValueExample`s which were not suitable for a specific
`ValueUseSite` were filtered out when creating `ValueUseSite` specific
`TestCase`s. This change moves it into `testCasesForValueUseSite(...)`.
It simplifies the creation of `TestCase`s in this and also simplifies
the follow-up change.

The main `ValueUseSite` that needs filtering are the fields as it is
not allowed to use a literal annotation in a field. Previously, the
filtering used `FIELD_VALUE` to filter both it and
`FIELD_WRITE_WITH_SEMICOLON`. However, now each is used separately.
As a result it was necessary to remove `FIELD_WRITE_WITH_SEMICOLON`
from `allValueUseSitesExceptFields` to avoid more test failures.

Bug: 354633349
Test: ./gradlew
Change-Id: Idcb4c478e9a829b17ac13194a2f537b449ac0a4a