Fix P0 test_concurrent_avds by preventing asyncio exception.
The default value of the _wait_for_started argument in the Application
class is inconsistent with the asyncio.wait_for function, which expects
an awaitable. This inconsistency triggers the exception when the start
method is called, causing any assertion to fail.

This change replaces the default "lambda: True" argument with an
equivalent asynchronous function that returns True after sleeping for 0
seconds.

BUG: b/371195354
Change-Id: I8bc3755b00def060de87b0e682ed24c014b3e558