Make Assertions companion implement Assertions
Previously, in order to use `Assertions` classes had to implement it.
That was not always possible. In order to work around it the
`CallableParameterDescriptorUsingDotsTest` had its `companion object`
implement `Assertions` allowing them to be imported and used in places
where the `Assertions` interface was not implemented. Unfortunately,
auto-completion lead to other classes importing `Assertions` method
from `CallableParameterDescriptorUsingDotsTest`'s companion.

This change makes `Assertions`' `companion object` implement
`Assertions` itself allowing it to be imported from there and removes
it from `CallableParameterDescriptorUsingDotsTest`'s `companion object`
and replaces imports of that with imports from `Assertions.Companion`.

Bug: 354633349
Test: ./gradlew
Change-Id: Iccd664bfa110dca6fec64e1bc8506d4fc1a27240