Clean up GetRootContainingLibartbase.
It worked by looking up the path to the DSO containing itself, which
assumes libartbase is in its own DSO. That's no longer the case for
most (all?) gtests since they started linking ART libs statically in
https://r.android.com/2623410.

From that point on, the DSO was the test binary instead, so it always
returned the path two levels up from its directory. Do the same by
looking at the command line instead, to make this a no-op change.

Test: presubmits
Bug: 404306250
Change-Id: Ifa574cb7f7f490af8a8be7d0a03c9180c3dd2685