Ignore TestCases that are not valid for the current InputFormat
Some `ValueExample`s that work in Java are not valid in Kotlin, e.g.
Kotlin does not support using a lower case `l` suffix for longs. This
provides support for ignoring those on Kotlin, and also ignoring any
future tests that may make be valid for Kotlin but not Java.

This will be used in a follow-up change that adds support for Kotlin.

Bug: 354633349
Test: ./gradlew
Change-Id: I76b632ce3cd45ad1dabbf2b2a2efe566fdc1d1c3