Add partialExpectations(...) to fall back to another Expectation
Previously, the `NotConstantFieldExpectation` would check to see if it
was being requested expected values for a field and if it was then
return `NO_INITIAL_FIELD_VALUE` to indicate that the field should not
have an initial value as it was not constant. If it was not then it
would delegate to another `Expectation` to provide the remaining
values.

That behavior of checking a set of partial expectations before falling
back to another set of expectations will be useful when adding Kotlin
tests as while Kotlin shares a lot of expected behavior with Java it
has some differences.

This change generalizes the `NotConstantFieldExpectation` to support
creating partial `Expectation<T?>` (that return `null` when the value
could not be found) that can then `fallBackTo(...)` an
`Expectation<T>` that is either complete or will error if it cannot
find an expected value.

`Expectation<T>` was changed to `Expectation<out T>` so that an
`Expectation<T>` can be assigned to an `Expectation<T?>`, i.e. so
that a complete set of expectations can be treated as a partial set
of expectations.

Bug: 354633349
Test: ./gradlew
Change-Id: I2ba733adc050c756051467058a74350b693be086