Add basic support for the new Value model
Adds the `Value` root type of the model hierarchy along with the
`DefaultValue` implementation class root. They are wired into the
model at the following points:
* AnnotationAttribute.value
* FieldItem.initialValue
* MethodItem.defaultValue

The `Value`s are supplied by the model implementations using a
`ValueProvider` (or `OptionalValueProvider`) to defer creation. They
are used consistently across all the model implementations.

Currently, the model implementations just use temporary placeholders
that return `null` (where possible) or throw an exception but they will
be replaced in follow-up changes.

This also adds tests for the three new properties. At the moment the
test ignores exceptions and `null` values when a non-`null` value is
expected to avoid having to keep updating the baseline files as
otherwise, they would change a lot which would be a source of
conflicts. Over time, the tests will become more strict.

Although there are currently 5 subclasses of
`BaseCommonParameterizedValueTest` only three of them have test methods
for the new properties. That is because while annotation attributes and
fields have multiple legacy string representations they only have a
single `Value`.

Bug: 354633349
Test: ./gradlew
Change-Id: Ia714259d15d6a169aac78697e712c3de0cba5fc5