Supply AnnotationItem when creating AnnotationAttributes
Previously, when creating the `AnnotationAttribute`s for an
`AnnotationItem` they did not have any connection to the containing
`DefaultAnnotationItem` or `Codebase`. That meant it was impossible to
correctly handle values, references to fields or classes could not be
resolved, the type of annotation attributes could not be determined.
This change passes the `AnnotationItem` as the first parameter to the
`attributesGetter` lambda which is responsible for creating the
`AnnotationAttribute`s. Follow-up changes will make use of that.

Bug: 354633349
Test: ./gradlew
Change-Id: I0a5eda556d2aa9dbec34731a2e91262ce0c11987