Improve IDE debugging of parameterized tests
Parameterized tests are great for running the same test across lots of
different sets of data and ensuring comprehensive test coverage of
different combinations of data. However, debugging can be a little
tedious as in the event of a test failure the IDE will help you
navigate quickly to the code that fails but often it is the data that
is at fault and navigating to that can be tedious.

This aims to solve that by manipulating any `Throwable` thrown during a
test so that the top stack trace element will refer directly to the
test data creation code so the IDE can navigate there with a simple
click on the stack element.

This manipulation will only be done if the `Throwable` has no init
cause set as that is where its original stack trace will be moved to
ensure that it is still available if needed. Preserving an accurate
stack trace for general debugging, is more important than convenience
in the IDE.

The main issue this has to address is how to identify which stack trace
element should be at the top of the stack to avoid the developer having
to find the correct location for themselves. The basic idea is that the
most relevant stack trace is the one that calls into the test
infrastructure code to create an individual piece of test data.

e.g. the `EntryPointCallerTracker.<init>` constructor should not be at
the top of the stack as it is the same for all pieces of test data.
Similarly, if that is constructed in the initializer of say `TestData`
then its constructor should not be at the top of the stack either.

This works by partitioning the code into test infrastructure code
(which is used for lots of pieces of test data) and other code which
calls that to create a single piece of test data. The stack element
that is at the top is the call into the test infrastructure that
creates a single piece of test data.

e.g. simple test data may be created by calling `TestData` constructor
directly. In that case the calls to the constructor should be at the
top of the stack. In the following, if there is an issue with "data1"
then the call to `TestData("data1", ...)" should be at the top of the
stack.
```
   val testData = listOf(
     TestData("data1", ...),
   )
```

With more complicate test data creation, e.g. using helper methods,
builders and lambdas, the set up becomes more complex but the basic
idea is the same.

Follow-up changes will integrate this into a number of parameterized
tests.

Bug: 354633349
Test: ./gradlew
Change-Id: I269862bef74d06963f8a5792ed2e4e20f9a1ae50