Add comprehensive equality tests
Previously, the `Value.equals(Any)` and `Value.hashCode()` methods were
not tested. This adds comprehensive tests for them which revealed an
issue with `NaN` (for both `Float` and `Double`). The primitive `NaN`
is not equal to itself but the `java.lang` wrapper object for `NaN` is
equal to itself. The `DoubleValue` and `FloatValue` both compare
primitives.

This change adds special case for checking equality of `NaN` to fix
the test.

Bug: 354633349
Test: ./gradlew
Change-Id: I2851dcd08ae790ab5c6ce1dfc8fffd4c06c36256