Track creation of TestCase in ParameterizedValueStringTest
Uses `EntryPointCallerRule` to improve IDE debugging of test failures
in `ParameterizedValueStringTest` by rewriting the stack trace of test
failures to refer to where the `TestCase` being tested was created.

After adding `TestCase.entryPointCallerTracker` and
`EntryPointCallerRule`. Some calls to following methods were modified
to cause them to fail:
* `testCasesForValue(...)`
* `verifyConfigMatchesDefault(...)`
* `verifyConfigChangesOutput(...)`

`EntryPoint` and `ExitPoint` annotations were then added as needed to
nsure that the top stack trace element navigated to the callers of
those methods. Any private method annotated with `EntryPoint` or
`ExitPoint` were made internal (as Kotlin appears to drop annotations
from private methods).

As `testCase(...)` was both an `EntryPoint` and an `ExitPoint` the
latter part was split out into its own method as they cannot be used on
the same method.

Bug: 354633349
Test: ./gradlew
Change-Id: I1dc857d7bd2ae7440e7b1a77d40bfafed3f84037