Support creating LiteralValues in Psi
Adds support for creating `LiteralValue`s in Psi from both Java source
(using `PsiAnnotationMemberValue`) and Kotlin source (using
`UExpression`).

The `Value` tests are quite lax and set up to fail only when a non-null
`Value` is provided and does not match the expected `Value` (or a
non-`ValueProviderException` exception is thrown). Otherwise, they
throw an `AssumptionViolatedException` which causes the tests to be
recorded as ignored. That meant it was necessary to manually verify
that the literal tests are not ignored on Psi for both Java and Kotlin.
They are not and they pass.

Bug: 354633349
Test: ./gradlew
Change-Id: I92d2ac7392fdd37299f3ae68680ea0e929dcae8c