Autoupdate of 2 submodules
*trusty/user/app/authmgr updated to 8087dcc6cfe81d6396f991c2f5bbde9c4f06478b
authmgr-be: move AIDL generation

The build rule is moved to trusty/user/base/interface and the crate name
is renamed to align with Android so that core code can be shared between
Android and trusty.

Bug: 377906851
Test: atest VtsSeeHalTargetTest
Change-Id: I8415333c07474a4ec0fcdbb0c9db27db1d9242cc

*trusty/user/base updated to 28d0d724c82b07330948d66af559ae92c92706c7
Move AuthMgr AIDL rust build rule

AIDL backend code generation for authmgr was previusly in
the authmgr-be TA source tree. This change moves it to a central
location and renames the crate to match the Android crate. This allows
trusty code to use authmgr-fe core code in system/security/authmgr
that depends on the authmgr interface.

Bug: 377906851
Test: atest VtsSeeHalTargetTest
Change-Id: I4fa7104fa35d7c7641a50355720b8cbbd3abf26c