Remember if an int was used for a double/float/long value
Sometimes an int literal, e.g. `1` or `0`, will be used in the source
as the value of a `double`, `float` or `long`. Ideally, it should not
matter and it should be normalized to the correct type but
unfortunately, the source representation is preserved for some uses of
the value. This does affect some checked in signature files.

This change preserves enough knowledge about that source representation
to maintain the legacy behavior when needed. It adds a
`wasOriginallySpecifiedAsInt` to the associated `Value`s and a
`treatAsIntIfOriginallySpecifiedAsInt` option to the
`ValueStringConfiguration` to provide control over whether that affects
the string representation.

Bug: 354633349
Test: ./gradlew
Change-Id: I7f3873e38d5f7e4612be38aa735418cfae334428