Push testLegacySource() into BaseCommonParameterizedValueTest subclasses
The `FieldItem.initialValue(Boolean)` boolean method does not provide a
source representation of the value it provides the value itself. That
is why its representation is often different to the others, e.g.
strings and characters without quotes.

This change pushes the `testLegacySource()` method down into each
subclass in preparation for stopping running it against
`FieldItem.initialValue(Boolean)` after adding tests for the legacy
value. The `AnnotationAttributeValue` also provides the underlying
value as well as the source so it will also need to add a test for the
legacy value.

Bug: 354633349
Test: ./gradlew
Change-Id: I16f5d9b302c7ed6cdb999a7727015a1a1432377b