Revert "Add support for creating Codebase dependent expectations"
This reverts commit 241137ad1eef85dfc7ab1ce9bcfb09800d3a63ff.

The `codebaseExpectations {..}` was added as it was needed during
prototyping when `Value` included a `TypeItem` however, that was found
to be unnecessary so it is not actually needed. At least not at the
moment.

If in future `TypeItem`s are needed then it will most likely not
require a complete `Codebase` as changes https://r.android.com/3556511
and https://r.android.com/3556512 have replaced `Codebase` in
`ClassTypeItem` and `AnnotationItem` with `ClassResolver` and
`AnnotationContext` interfaces which provide the functionality they
need.

Bug: 354633349
Test: ./gradlew
Change-Id: Ib47fe2c44c26e10bfdd647256e3b4ff45f9d9e4e