Make ParameterizedPrimitiveValueConversionTest support strings
Previously, it would only support primitives. This change adds support
for strings in preparation for providing some conversions between
strings and chars which are needed for parsing signature files, where
chars are sometimes represented using single quotes and sometimes as
single character strings using double quotes.

This change removes the class to make it clear that it supports all
literals and removes some `ValueFactoryTest` methods for strings as
they are now covered by this test or are already covered by
`ParameterizedCreateLiteralNoTypeItemTest`.

Bug: 354633349
Test: ./gradlew
Change-Id: I3fcba1cd02d3bd76c221169ad1173c81aaf42578