Fix P0 test_app_launchafter_snapshot_load by using App subclass.
Use the Application subclass to fix the test app launch. Change the test
app from the animation app to the Messaging app, which is less prone to
launch errors in an attempt to reduce test flakiness. Add a new fixture,
messaging_app, to launch the Messaging application.

This change reverts a temp change in the snapshot_service that prevented
the default snapshot from being deleted.

BUG: b/401345242
Change-Id: Iaff6c2be473258ea0345a77ed6de8efab6e3a4b7