Add testLegacyValue() to CommonParameterizedAttributeValueTest
Requires adding some new expectations and tweaking the existing ones
where needed. Two points to note:
* The `enum` test does not have `expectedLegacyValue` set as the value
  it returns is a Psi specific objects, e.g. `PsiField`.
* The kotlin class literal tests are baselined because they throw a
  class cast exception because Metalava does not handle class literals
  in Kotlin properly.

As is done for the other expectations the behavior of the Psi model is
taken as the standard and any deviations from that by other models are
baselined, even when a model's behavior makes more sense.

Bug: 354633349
Test: ./gradlew
Change-Id: I824f82a4c89ac758c62b410972ee2f8f271cb0a0