Autoupdate of 2 submodules
*trusty/user/base updated to f592a4e96a85b1912537df3674524a897bd29a64
include/user: Add `HSET_MOD_WITH_COOKIE`

Adds a variant of `HSET_MOD` that requires the provided cookie to
match the cookie that was previously stored with the handle.

Bug: 382291660
Test: tipc.test && authmgr_be_lib.test
Change-Id: I8437ba5c5df4305221a5b06b7359c38dafe97af7

*trusty/kernel updated to fdaa1d7c1c82e5c80d152ab530399e359412a3f0
lib: trusty: Add `HSET_MOD_WITH_COOKIE`

Adds a variant of `HSET_MOD` that requires the provided cookie to
match the cookie that was previously stored with the handle.

Bug: 382291660
Test: tipc.test && authmgr_be_lib.test
Change-Id: I04947f83b7698cbfde897c1127e419ef1207af1d