Add initial LegacyValueFormatter
The legacy source representations are affected by a number of factors:
* Use site.
* Jar or sources.
* Java or Kotlin

Rather than add that to the `Value` classes this adds the
`LegacyValueFormatter` to handle formatting a `Value` in the same way
as a use site, taking into account the other factors. The intent is
the existing legary source providing code will be implemented in
terms of the corresponding `Value` and a use site specific instance of
this.

This adds an initial implementation that supports replacing specific
values with a custom string representation.

Bug: 354633349
Test: ./gradlew
Change-Id: I36ba406a9ed917f3f37aa0a41fd3cb4796266227