pvmfw: Validate input DT PSCI node
Make pvmfw parse the PSCI node of the input DT and validate that the
values of the 2 properties we expect (<compatible> and <method>) are the
same as in the platform.dts as it would previously have silently
replaced unexpected values with the hard-coded ones. Now, pvmfw will
abort the boot instead, ensure a more stable integration (fail early).

This prepares for a change allowing PSCI v0.2.

Test: m pvmfw
Bug: 396136321
Change-Id: I4eae3c4aad7d2841263bba2b8307637d1b372590