Prepare for adding legacy state to Value classes
Currently, `Value` classes only contain the normalized form of the
underlying value that only depends on the logical value of the original
expression from which it was created, not the specific form. e.g. the
following all produce the identical value.
* `longValue = 3000`
* `longValue = 3000L`
* `longValue = 3_000L`

However, in order to be able to use the `Value`s to produce the same
string representation as the current legacy string generating code it
is necessary to include some information in the `Value` about the
original form of the source expression. That introduces issues about
what equality means. This change clarifies that and documents how
that additional "legacy" state will be dealt with in general.

Bug: 354633349
Test: ./gradlew
Change-Id: I329fe3c6c3c6ebee82b3547e3f0f12602e517c49