Support ClassObjectValue in ValueParser
The changes to `ValueParser` parsing code are quite simple. The rest of
the changes are needed to integrate `TypeItemParser` with
`ValueParser`.

It was necessary to modify `ParameterizedValueParserTest` to ignore
some test cases that either cannot work (e.g. unqualified types) or do
not currently work but will be fixed in future (e.g. Kotlin syntax for
class literals).

Bug: 354633349
Test: ./gradlew
Change-Id: Ifed07468e0d0cbaacfbfdda882d5d69a262e7a09