Add support for FieldReferenceValues
A `FieldReferenceValue` is either an `EnumConstantValue` or a
`ConstantFieldValue`. It was necessary to normalize some fields to
their constant value to avoid differences between Java and Kotlin.
e.g. `Double.NaN` is represented as the `java.lang.Double.NaN` field in
Java and `kotlin.jvm.internal.DoubleCompanionObject.NaN` in Kotlin.

Adding the `expectedValue` broke the "String using constant" tests
because the existing code evaluates the `Constants.STRING_CONSTANT` to
its constant value of `"constant"`. They have been baselined as part of
this change. The follow-up changes that add support for
`ConstantFieldValue`s will remove them from the baseline files.

Bug: 354633349
Test: ./gradlew
Change-Id: I6a111f618b4dab1c18770e69b6a6e7b6cc799ff4