Use Value in JDiffXmlWriter
Running this on Android did actually reveal a few difference. Every
enum constant called `<X>` had `value="PsiField:<X>"` in the JDiff
file, which is clearly wrong. That is because an enum constant has
itself as its `value`. After this change they had no `value`. It's
possible that a consumer of a JDiff file relied on that in some way but
it is unlikely given it is exposing an implementation detail of the Psi
library/Metalava. If they are it should be easy for them to fix it by
using the `metalava:enumConstant` extension.

Bug: 354633349
Test: ./gradlew
      # Before applying this change run:
      tools/metalava/scripts/gather-android-metalava-artifacts.py before
      # After applying this change run:
      tools/metalava/scripts/gather-android-metalava-artifacts.py after
      # Make sure that there are no differences.
      meld before after
Change-Id: I0853d05902770afebc4a21e6ef9eee131e1402c4