Fix legacy value parsing in DefaultDefaultAnnotationValue.create(...)
Previously, it would not unescape strings and characters or handle the
float and long suffixes of `f`/`F` and `l`/`L` properly. This change
fixes it. That makes it easier to add additional tests for those types
as they will not require baseline updates. It should not cause problems
with existing uses, e.g. `signature-cat`, as signature file just passes
through the source representation.

Bug: 354633349
Test: ./gradlew
Change-Id: I969eafa2973dd48d047409c0532b8111a99cc373