Allow PsiValueFactory to return null Value for fields
Previously, if Psi encountered an expression that it could not convert
to a `Value` it would throw an exception. While that is a valid option
for annotations, where every valid expression should be convertible to
a `Value` it is not valid for fields. A field initializer could be an
expression that does not produce a constant value at compile time, e.g.
a method call. In that case it should be mapped to a `null` `Value`.

This change refactors `PsiValueFactory` to return `null` for unknown
expressions for fields (but not for annotation values).

Bug: 354633349
Test: ./gradlew
Change-Id: I63746f22fd251cafc0190d0f00a06b6660a7b1cc