Add ValueUseSite and pass to model value factories
Previously, `PsiValueFactory`, `TurbineValueFactory` and `ValueParser`
did not differentiate between `Value`s produced for annotations or field
initializers. However, fields can only support constant values. This
change adds a `ValueUseSite` enumeration and passes it through to the
model value factories for them to use, if necessary.

Follow-up changes will use this information. It will not be needed in
`TurbineValueFactory` as Turbine will return a `null` `TurbineValue`
for a field if it was not a constant value.

Bug: 354633349
Test: ./gradlew
Change-Id: Ic3eb3dadb19b1f5b2324233568de7e6a0b75945c