Make all expectations support returning null
Previously, `Expectation`s were expected to return non-null and had to
use special value of `NO_INITIAL_FIELD_VALUE` to pretend to be null.
That was necessary for `ChainedExpectation` which assumed that a `null`
expectation meant that there was no expectation available.

This change allows `Expectation`s to use `null` and removes the need to
use `NO_INITIAL_FIELD_VALUE` which was replaced with `null`. The
`ChainedExpectation` was fixed by adding a `hasExpectationFor(...)`
method to `Expectation` which `ChainedExpectation` uses to determin if
there is an expectation before retrieving it.

This will simplify follow-up changes which need to return `null` for
`FieldItem.initialValue`.

Bug: 354633349
Test: ./gradlew
Change-Id: I49107b1afb51adbc2b0745d304d36b4f4dd7ac90