Support testing null Values for fields
Previously, `checkExpectedValue(...)` ignored any `null` `Value`s. This
change makes it check them. That revealed a number of issues with the
handling of `Value`s on fields:
1. Fields of constant type should always have a `ConstantValue`, but
   sometimes they had `ConstantFieldValue` (which is not a
   `ConstantValue`).
2. Fields of non-constant type should always have a `null` `Value` but
   sometimes they were producing a non-null `Value` and sometimes the
   test was expecting a non-null `Value.

This fixes those inconsistencies by:
1. Forcing the `ValueProvider` to only return `ConstantValue`s for
   fields.
2. Forcing the expectations to only expect `ConstantValue`s for fields.

This change had to temporarily disable the `String with constant` tests
in signature files as `ValueParser` does not yet support field
references.

Bug: 354633349
Test: ./gradlew
Change-Id: I26e87e533fc5e8fbff97203401b3b4d0353fec23