Switch FieldReferenceValue.qualifiedClassName to classTypeItem
Previously, `FieldReferenceValue` specified the class that contained
the field using `qualifiedClassName: String`. However, switching it
to `ClassTypeItem` will allow it to take advantage of the support for
handling types in signature files, i.e.:
* Removing `java.lang` prefixes when writing.
* Replacing `java.lang` prefixes when reading.
* Detecting other unqualified types and reporting issues with them.

This makes it consistent with `ClassObjectValue`.

The latter is quite important as there are many unqualified fields
and class literals written into signature files. Using the same
mechanisms for both will ensure consistent handling.

Bug: 354633349
Test: ./gradlew
Change-Id: I6cbceddbe15eeaae7d5d9b1cb50515a85c2f4b7e