Fix issue with fields on generic classes
Previously, Psi would fail if attempting to process a field from a
generic class. `ValueFactory.createConstantFieldValue(...)` requires
that the `ClassTypeItem` has no arguments however, when called from
`ValueFactory.createConstantFieldValue(FieldItem,...)` when `FieldItem`
was from a generic class it would pass in a `ClassTypeItem` with
arguments. This change adds a test for that and fixes it.

Bug: 354633349
Test: ./gradlew
Change-Id: Ia608f473c0dacc4381c021a6cc6425478323c82b