Test behavior of enum constants in JDiffXmlWriter
Previously, `JarToJDiffCommand` did not have a test for enums. That was
primarily because testing jar based code was difficult as it had to be
manually created and built outside Metalava and then made available to
Metalava tests. However, jars can now be built in tests so this change
adds a test for enums which shows that it mistakenly uses
`value="PsiField:CONSTANT"`. A follow-up change will fix that.

Bug: 354633349
Test: ./gradlew
Change-Id: I24676f7fbc89c64ba8c9a0b13bbe3781c3c2d508