Simplify creation of FieldReferenceValue from FieldItem
Previously, it was the responsibility of the caller to choose whether
to create an `EnumConstantValue` or `ConstantFieldValue` from the
`FieldItem` and, if the latter, to obtain the field's `ConstantValue`,
if any.

This change moves all that into the `ValueFactory` and uses the field's
`initialValue` as the `ConstantValue`, if possible. The latter will
avoid repeatedly recomputing the value of each field.

Bug: 354633349
Test: ./gradlew
Change-Id: I65a2606b335d9d6768c93a22f60a339b7a7b13d0