Allow Tokenizer to return a token without finding a separator
Previously, `Tokenizer.getToken(...)` would throw an exception if the
text being scanned did not end with a separator. That made it confusing
when adding tests as it would fail for no obvious reason.

This change refactors `getToken(...)` to only throw an exception when
it reaches the end of the file if there is an incomplete token, i.e. a
a string with no close quote or a type parameter list that was not
closed.

This does not detect every possible incomplete token but it can easily
be improved in future.

Bug: 354633349
Test: ./gradlew
Change-Id: I88018831a8e76ae01fe2d060252a82b4c47c9a22