Use FieldItem.constantValue in compareFieldItems(...)
Previously, it used `legacyInitialValue()`, for which `constantValue`
is the equivalent replacement. This change also updated
`hasSameValue(...)`, renaming it to `hasSameConstantValue(...)` to
clarify its behavior.

This change broke a `BinaryCompatibilityClassFieldsTest` that was
invalid. It worked previously because it used an invalid signature file
as the new API. That was invalid because it had a field which was not
both `static` and `final` with a value, which is not something that can
happen. The text model also has a bug in that it just returns whatever
value is specified for all calls to `legacyInitialValue(...)`. So, the
value check compared `0` (constant) against `0` (not constant) and
decided that they matched.

This change made text model return `null` from `constantValue` as it is
not a constant. So, the value check now compares `0` (constant) against
`null` (not constant) and reports an issue. Some tweaks were made to
the error reporting to produce more meaningful error messages.

e.g. The `!old.modifiers.isFinal()` was removed as it was unnecessary
and would just lead to `null` being written out in a message but `null`
is not constant and so is not written out to signature files anyway.

Bug: 354633349
Test: ./gradlew
Change-Id: I4b1376395aba4128a96abda2cd20ce392864c68b