Improve testing of FieldItem.constantValue/initialValue
For every test for `legacyInitialValue(...)` this adds an equivalent
test for `constantValue` and `initialValue`. There are some differences
in the set of values which are handled by them which have been
explained and `TODO`s added for those differences that will need
revisiting in future, either to explain why it is ok, or to fix the
differences.

Bug: 354633349
Test: ./gradlew
Change-Id: I899eb03fb8543f1d6ad149f43ddc21a4f1870fa1