Use initializer blocks in enum stubs
Previously, when generating stubs an `enum` class was treated like
interfaces and annotations in how it handled setting unknown values for
`static final` fields, i.e. they were set to the zero value of the
type which was wrong.

This change makes them behave like classes and use initializer blocks.

Bug: 354633349
Test: ./gradlew
Change-Id: I440f045c8a00402033853ea32c30178f33c4f000