Use FieldItem.constantValue in writeValueWithSemicolon()
Previously, `writeValueWithSemicolon()` used `legacyInitialValue()`.
This change switches it to use `constantValue`. The check for
`static final` was removed as `constantValue` is only set to non-null
for `static final` fields anyway (or `const` fields in Kotlin).

This necessitated some changes to the `ValueExample`s as it fixes some
inconsistencies between the jar and sources (there seemed to be a bug
in the `legacyInitialValue()`'s handling of references to fields in
jar files) and between Java and Kotlin.

Generally, changing the behavior of the legacy code such that the
expectations need to be changed is something to be avoided as it could
lead to changes in signature files downstream. However, in this case
it appears that it does not affect AndroidX and only has a mild impact
on Android (fixes a bug with one field that was missing its constant
value).

This fixed a number of signature and Turbine tests. It also broke a
couple of tests for Psi `writeValueWithSemicolon()` that were already
broken for `FieldItem.constantValue` as they now use the same
underlying code.

Bug: 354633349
Test: ./gradlew
Change-Id: Ie2cbda99694c74e1a65abe57f4da0e43f559aecc