Ensure that ConstantValues are of a constant type
Previously, `Psi` assumed the underlying value of a
`PsiLiteralExpression` or `ULiteralExpression` was always a
`ConstantValue`. However, in order for it to be a `ConstantValue` it
must also be a suitable constant type, i.e. `String` or primitive type.

e.g. the initial value of the following field is a `ConstantValue`
because the type and value are both constant.
```
    public static final String CONSTANT = "constant";
```

However, the initial value of the following field is not a
`ConstantValue` because while the type is a constant the type is not.
```
    public static final CharSequence NOT_CONSTANT = "not constant";
```

This change adds a test for it. That breaks in `PsiValueFactory` as it
calls `createLiteralValue(...)` with a type of `CharSequence` which is
incompatible with the `String` value.

Usually, changes like this would separate the fix from the test to show
the effect of the change. However, as the test fails with an exception
rather than the wrong value there is no real benefit to doing that so
this change also fixes the problem by adding a check to verify that the
type is constant.

Bug: 354633349
Test: ./gradlew
Change-Id: I3a54b63d01d7fb47ccb59b2932b2ba8a2bebb5cb