Add tests to show misbehavior of ConstantEvaluator
Metalava uses `ConstantEvaluator` from Android Lint to evaluate what
should be constant expressions as defined by Java 21 LS 15.29.
Unfortunately, it evaluates more expressions than that, e.g. it will
evaluate `new int[] {1}` or `arrayOf(1)`. Its behavior is also
inconsistent between Java and Kotlin, e.g. it will not evaluate, i.e.
return `null`, for `(new int[] {1})[0]` but it will evaluate the
equivalent Kotlin expression ot `arrayOf(1)[0]`, returning `1`.

This adds a test to show that inconsistent behavior. The Java behavior
is correct so that is used as the expected result of the test which
causes the Kotlin tests to fail. They are baselined until the issue is
fixed in `ConstantEvaluator`.

Bug: 354633349
Test: ./gradlew
Change-Id: I7c3731abe67598c2d9b28fcf0b443cec79e49fea