Fix fields initialized with a new array expression
Previously, expressions like `new int[0]` would cause problems in
`PsiValueFactory` because `ConstantEvaluator` should have evaluated it
to `null` as it is not a constant. Unfortunately, it is evaluated to
either an actual array, or an Android Lint specific class  if the array
is above a certain size. That ends up being passed to
`createLiteralValue(...)` along with the `int` type and it fails as an
array of ints is not a valid `int`.

This avoids that issue by checking for `PsiNewExpression` early and
immediately returning `null`.

Bug: 354633349
Test: ./gradlew
Change-Id: Ie3c42007dc93d874d5a308b588e8f9d37273497d