Improve ParameterizedValueParserTest
Previously, `ParameterizedValueParserTest` would assume that every
legacy source expectation of a `ValueExample` would, if parsed, produce
the exact same `Value` as the `ValueExample.signatureExpression` would.
However, that is not true. Some of them have lost information that is
necessary to create that `Value`. That assumption meant that some tests
had to be baselined.

This change corrects that assumption and instead only requires that the
`TestCase`s created from expectations instead of `signatureExpression`
can be parsed.

A future change will try and strengthen that check by ensuring that
when given a suitable `LegacyValueFormatter`, the `Value`'s string
representation will match the string from which it was parsed.

Bug: 354633349
Test: ./gradlew
Change-Id: Ieb28564bc349ffb2bc3f6049835f2852fb6cbe32