Fix handling of <type>::class in Kotlin
Previously, there was a bug in the code which was supposed to convert a
`<type>::class.java` into a `<type>::class` for use in annotations
which meant that the tests for it were invalid. Fixing that revealed a
missing `import kotlin.reflect.KClass`. Fixing that revealed a hole in
the `PsiValueFactory` as while it could handle `<type>::class.java` it
could not handle `<type>::class`. This change fixed that by extracting
the code that handled `<type>::class` from the code that handled
`<type>::class.java`.

Bug: 354633349
Test: ./gradlew
Change-Id: I1757262de3a27e572a9d88a77eb2ce9aba833285