Process RequiresPermission annotations consistently
Previously, `AndroidApiChecks`, `ApiAnalyzer` and `DocAnalyzer` each
had their own code for extracting information from a
`RequiresPermission` which was similar but not the same. This change
adds `RequiresPermissionInfo` to dedup the accessing of that data and
abstract it away using the new `Value` objects. It switches all the
existing code to use that instead.

One slight issue was that this caused some previously unreported issues
with `RequiresPermission` annotations from `tools/metalava/manual`. It
temporarily works around them by downgrading the severity of the issues
to error when new and they will be fixed in bug 414336151.

Bug: 354633349
Test: ./gradlew
Change-Id: Ic93803152c1e64d753c96f16913686893f7aa88e