Only create ConstantValues for fields
Previously, `PsiValueFactory`, `TurbineValueFactory` and `ValueParser`
would all create non-`ConstantValue`s for fields, e.g. arrays, that
would then be ignored by the `FieldItem.constantValue` field. This
change makes them more efficient by only creating `ConstantValue`s for
fields and ignoring other values.

The `PsiValueFactory` and `TurbineValueFactory` rely on their model's
specific support for evaluating constant expressions to inline constant
fields and will ignore any non-constant values, returning `null`.

The `ValueParser` is strict and if a field has an initializer and it is
not a constant then it will throw an exception. That is because
signature files should already only use constant values as field
initializers so detecting any that are not should help flush out any
bugs in the parsing. If it turns out that there are lots of
non-constant values used as field initializers in signature files then
the decision to be strict can be revisited.

Having `ValueParser` be strict is not an issue because the preceding
change prevented running `testFieldValue()` with non-constant
expressions. Strictly speaking the same tests are invalid in Kotlin too
(as `const` vals can only have constant expressions) but `Psi` does not
seem to care.

For Java code it is important to test that non-constant expressions are
handled correctly and return `null` as they are valid field
initializers, they just are not treated as constant.

Bug: 354633349
Test: ./gradlew
Change-Id: I854193831429ac6c0adbf5a961e2a68c72cd0d58