Remove AnnotationFilter's specialized parsing of annotations
Previously, `AnnotationFilter.fromOption(...)` has its own custom code
for parsing an annotation specified on the command line into a separate
qualified name and `List<AnnotationAttribute>`. The latter was created
using `DefaultAnnotationAttribute.createList(...)` which is really an
implementation detail of `DefaultAnnotationItem.createFromSource(...)`.

This change replaces that with a call to `createFromSource(...)` and
then extracts the `qualifiedName` and `attributes` from the returned
`AnnotationItem`. Although that can return `null` it only does so if
the `AnnotationManager` rejects the annotation class but that will not
happen in this case as it uses the `NoopAnnotationManager`.

Bug: 354633349
Test: ./gradlew
Change-Id: I8d56ddcfe9cb12e102d86218257dd4fffc4090b9