Adjust filter criteria in ParameterizedValueParserTest
Removes the check for empty string as that is unnecessary and the check
for no imports (which is not needed as `signatureExpression` for those
`ValueExample`s is fully qualified. That did cause a breakage for class
literals of the form `Array<X>::class` as `Array<X>` should be
converted into `X[]` but it does not appear that form is currently used
in any signature file.

Bug: 354633349
Test: ./gradlew
Change-Id: Ic2a494f54039bfdd9c6a3e5c8259726d4fb0f8cd