Stop taking nullable strings in normalizeInput/OutputName
Previously, `AnnotationManager` methods `normalizeInputName(...)` and
`normalizeOutputName(...)` would take a nullable `qualifiedName`
parameter. That was a hold over from code that was written for Psi
which allows `PsiAnnotation` to have a `null` name. As that is not
something that makes sense for any other model, this change moves
the handling of `null` name to Psi and makes `qualifiedName` non-null.
That also allows `normalizeOutputName(...)`'s return type to be
non-null too as it would only return `null` if it was supplied with
`null`.

Bug: 412564465
Test: ./gradlew
Change-Id: I5a6602372c4bbe772c75aee1544c17c970408490