Add SystemApiType to select the system API to test, if any
Previously, `includeSystemApiAnnotations` was a boolean that would add
`--show-annotation ...SystemApi` and `--show-annotation ...TestApi` if
not already supplied. That was unrealistic. This change adds
`SystemApiType` to allow the specific system API that is being tested
to be selected and replace the type of `includeSystemApiAnnotations`.

At the moment the `SystemApiType` is not realistic as it only adds
`--show-annotation android.annotation.SystemApi` (with no `client`).
That is because some tests fail when that is provided. Follow-up
changes will fix that.

Bug: 412581637
Test: ./gradlew
Change-Id: I51b7310b5ed77c53c5c091a3463f9312e348a08b