Make system API related tests more realistic
Previously, the tests used
`--show-annotation android.annotation.SystemApi` to select system APIs.
However, that is not realistic as Android never specifies an annotation
filter for `SystemApi` without specifying a `client` attribute.

This change fixes that. It caused a number of tests to start breaking
on Turbine. That is because Turbine's `AnnotationItem.toSource(...)`
does not support `showDefaultAttrs` and without that `@SystemApi` will
not be treated the same as `@SystemApi(client=PRIVILEGED_APPS)`. That
will be fixed in future as part of the work of b/354633349.

The `SystemServerCheckTest` had to be reverted back to using
`--show-annotation android.annotation.SystemApi` due to bug 412743564.

Bug: 412581637
Test: ./gradlew
Change-Id: If95218116291252b831bc9c3fd00c1588bc9724f