Do not show default values by default in AnnotationItem.toSource(...)
Previously, `AnnotationItem.toSource(...)` would default to setting
`showDefaultAttrs = true`. That was a waste of time as there is only
one place where it is required and that is in `AnnotationFilterEntry`'s
`fromAnnotationItem(...)` method. This change inverts the default.

This revealed that `ANNOTATION_TO_SOURCE` was unnecessarily including
default values. They will be cleaned up in a follow-up change.

Bug: 354633349
Test: ./gradlew
Change-Id: Ia6e287137371455d685280a172c6ec1c3d10dec0