Use Values in AnnotationFilterEntry
Previously, `AnnotationFilterEntry` only worked properly in Psi because
it relied on `toSource(showDefaultAttrs = true)` to add default values
so that `@SystemApi` would be treated the same as
`@SystemApi(client=PRIVILEGED_APPS)`. That was only supported on Psi.

This change replaces that with code that uses the new `Value` objects
from `AnnotationAttribute.value` and `MethodItem.defaultValue`. That
works on Psi and Turbine and should be slightly quicker too.

Bug: 354633349
Test: ./gradlew
Change-Id: Ib3b8dc62b7d85e567068041f8b698d7ef0f938f8