Add test for using show annotations with internal
Shows that classes with `internal` modifiers are treated are as
accessible if they have a show annotation.

This is added to test the behavior of `BaseModifierList.isAccessible`
which access `showability` during `Codebase` construction. Computing
`showability` in that case requires accessing the value of
`RestrictTo.value`. Currently, it uses `legacyValue` but in future it
will use the new `Value` objects so this will make sure that accessing
a `Value` during `Codebase` construction works correctly.

Bug: 354633349
Test: ./gradlew
Change-Id: Ieb749f2b37b7362938509cc6df7ebb5f6d818cab