Avoid resolving annotation class in Turbine
In order to compute a `Value` for an annotation attribute it is
necessary to obtain the `TypeItem` for its attribute in order to
perform any necessary conversions. While Turbine already does a lot of
that itself it does not do it all.

Previously, Turbine would use `CachingAnnotationValueProvider` to do
that by resolving the `AnnotationItem` to a `ClassItem` and searching
for the attribute's method and using its return type. That causes
problems if the `Value` is accessed during construction as it may end
up resolving a class which has not been created yet.

This change replaces the use of `CachingAnnotationValueProvider` with
a Turbine specific class which does the equivalent job.

Bug: 354633349
Test: ./gradlew
Change-Id: Ia5fd47de574a4992353e80bfaba92a9aa9924fc8