Replace mapping for special values with one for fields
In Java `Double.NaN` is represented by the field `java.lang.Double.NaN`
but in Kotlin it is represented by
`kotlin.jvm.internal.DoubleCompanionObject.NaN`.

Previously, to avoid having different values for Kotlin and Java the
`createFieldReferenceValue(...ConstantValue)` method would map the
`ConstantValue` to the Java field. However, that relied on the field
having been evaluated to a `ConstantValue` and passed in to the
factory. Future changes will defer the creation of the `ConstantValue`
until it is needed at which point it will no longer be available in the
factory.

In preparation for that this change stops replacing fields that use one
of the special values with their `ConstantValue`. Instead it adds
mappings from the special Java and Kotlin fields to their
`ConstantValue`.

Bug: 354633349
Test: ./gradlew
Change-Id: Iebbb2f85adc17ae298525dad90c9b5625f09ac84