Replace FieldReferenceValue.classTypeItem with qualifiedClassName
Change https://r.android.com/3575774 replaced `qualifiedClassName` to
`classTypeItem` as that would allow it to reuse code for parsing
class types, e.g. detecting unqualified types. However, using
`classTypeItem` in the `FieldReferenceValue` complicates lots of other
code.

This change switches back to `qualifiedClassName` and leaves the use
of `ClassTypeItem` for reading and writing signature files as an
implementation detail of the text model.

Bug: 354633349
Test: ./gradlew
Change-Id: Ie7d8646dfd1143487334b29e3bf59fa6f587c36c