Support unresolvable fields in Psi and Text models
Previously, using unresolvable fields in an annotation would break in
the Psi model and using unresolvable fields which are also unqualified
would break in the Text model. This change adds a test for that and
makes it work in Psi and Text models.

The test is baselined in Turbine as it is stricter and does not allow
unresolvable fields in annotations.

Bug: 354633349
Test: ./gradlew
Change-Id: I4e7d25e91e63d4c94468cdf563e5f6521f8f560b