Override defaultsForAnnotationClass(...) in Psi
The Psi model needs to access `AnnotationDefaults` during `Codebase`
construction and it cannot resolve an annotation's `ClassItem` to do
that as it causes problems if the `ClassItem` has not yet been created.
This change overrides the `defaultsForAnnotationClass(...)` method in
`PsiBasedCodebase` to create the `AnnotationDefaults` from the
`PsiClass` if the `ClassItem` is not yet available.

This prevents https://r.android.com/3602054 from breaking some tests.

Bug: 354633349
Test: ./gradlew
Change-Id: I7363e490b6310509eda8ae4f82cd209e5b244b09