Clarify misleading naming of CodePrinter.inlineFieldValues
Previously, the `CodePrinter.inlineFieldValues` did the opposite of
what was expected by its name. Setting `true` would cause it to not
inline field values and setting it to `false` would cause it to inline
field values. That confusion leaked into `ExtractAnnotations` too
where it used `inlineConstants` to mean not `inlineConstants`.

This change inverts the behavior of `inlineFieldValues` and renames
`inlineConstants` to `keepFieldReferences`, clarifying the behavior.

Bug: 331752084
Test: ./gradlew
Change-Id: Ia02c2a76b476c89f8e7283046131cc2ec3efe70a