Support formatting single element arrays like the source
Previously, `ArrayValue`s containing a single element would only be
formatted based on their content, i.e. always wrapped or always
unwrapped. However, legacy use sites sometimes formatted based on the
source representation, e.g. unwrapped if the source was unwrapped and
wrapped if the source was wrapped.

This change adds a `wasUnwrappedInSource` to `DefaultArrayValue` and
a `SingleArrayElementFormat.SOURCE` to format like the original source.
The `SOURCE` enum is deprecated as it should only be used for
replication legacy behavior.

Bug: 354633349
Test: ./gradlew
Change-Id: I76f94190c7233668b62b12ab08c6f68bf8733a8e