Add test for array containing single item
The existing tests only cover a single item that is not syntactically
in an array but which is wrapped in one because of the type. This adds
a test for a single item syntactically in an array. It is needed as the
behavior of the two is different in some legacy uses.

The test currently fails in Turbine as `ATTRIBUTE_DEFAULT_FORMATTER`
does not handle it correctly. A follow-up change will fix that.

Bug: 354633349
Test: ./gradlew
Change-Id: I66c1e55a5f30c29e13a858ac96d18ec15fee5952