Return name of unresolvable field in AnnotationItem.optionalFlagName
Previously, change https://r.android.com/3605699 switched
`optionalFlagName` from `legacyValue()` to `value`. Unfortunately, that
changed its behavior when the flag was an unresolvable field. Before it
would return the unresolvable field name but after it returned `null`.
That in turn affected the behavior of `getFlagForAnnotation()` which
would return `REVERT_FLAGGED_API` before and `null` after. That in turn
affected `showability`.

Bug: 354633349
Test: ./gradlew
Change-Id: I511c26fcbfafad16288786f4afc89a3304a11062