Drop suffix from long and float for Kotlin legacyDefaultValue()
`CodePrinter.toSourceString(UExpression?)` does not add a type suffix
to long and floats, i.e. `L` and `f` respectively. This change emulates
that behavior by delegating to `Value.appendValueStringTo(...)` and
then removing any type suffix added. It does this rather than just
formatting the value itself as `treatAsIntIfOriginallySpecifiedAsInt`
may be set in which case the long or float may actually be treated as
an int.

Bug: 354633349
Test: ./gradlew
      # Tested in Android and AndroidX and does not affect signature
      # files in either or stubs in the former.
Change-Id: I1722525b06bfcd22afa20f1985c34dd312275b82