Remove PsiMethodItem.legacyDefaultValue()
Previously, `PsiMethodItem` provided its own implementation of
`legacyDefaultValue()`. This change removes that so it will use the
`DefaultMethodItem.legacyDefaultValue()` implementation.

There was a slight difference between the two as the former would drop
annotation attribute name if an annotation had only a single attribute.
That is ok if the name is `value` but it did it for any attribute.
Fortunately, that behavior is not relied on downstream so this change
fixes the test.

Bug: 354633349
Test: ./gradlew
Change-Id: I931f0d738724d512a63b69ed1db61863b71cb658