Add test to show issue with extending annotated classes
When parsing multiple signature files into a single `Codebase`
`ApiFile` has to detect when a class has already been registered in a
previous signature file, make sure it is consistent and then merge any
new annotations into the existing class. Merging annotations requires
comparing them and their values. Comparing their values may require
resolving their class to find the type of the attributes.

Currently, the merging is done while parsing the file which means it is
possible for an annotation to be resolved before it is seen. If that
happens then a stub class is created but it has a `ClassKind.CLASS`.
Unfortunately, when the annotation class definition is processed it is
created with a `ClassKind.ANNOTATION`. As a `ClassItem` for it already
exists it compares the stub class and actual class and fails because
they have different `ClassKind`s.

This adds a test for that which currently breaks. Follow-up changes
will fix it.

Bug: 354633349
Test: ./gradlew
Change-Id: Iafec7b390dbc3681d4941474b5d8fe0344e684ce