Add API history test to show incorrect deprecation
Currently, every extension version behaves as if its API was part of
the next SDK version and mark it as being part of the next SDK version.
However, that is not true as releases of SDK extension versions
were/are interleaved with releases of SDK versions. That causes
incorrect history to be recorded. e.g. Imagine the history of class
`Foo` was as follows:
* Added in SDK version 39
* Extended in SDK extension version 102
* Deprecated in SDK version 40
* Still deprecated in SDK extension version 103.

It should be marked as deprecated in SDK version 40.

However, the history is processed as follows (assuming the next SDK
version was 41):
1. SDK version 39 - `Foo` is not deprecated.
2. SDK version 40 - `Foo` is marked as deprecated in 40.
3. SDK extension version 102/next SDK version 41 - `Foo` has
   deprecated status cleared as it is not deprecated in SDK extension
   version 102 and so is treated as not being deprecated in SDK version
   41, overriding the previous deprecated state.
4. SDK extension version 103/next SDK version 41 - `Foo` is deprecated
   in SDK extension version 103 and so marked as deprecated in SDK
   version 41.

So, the result is it is incorrectly marked as deprecated in 41 when in
fact it was deprecated in 40. The cause of the misbehavior is step #3.
It clears the disabled status allowing it to be reset. Step #3 should
not be being performed at all.

If step #3 was not performed then it would behave as follows:

1. SDK version 39 - `Foo` is not deprecated.
2. SDK version 40 - `Foo` is marked as deprecated in 40.
3. Not performed
4. SDK extension version 103/next SDK version 41 - `Foo` is
   deprecated in SDK extension version 103 and so marked as deprecated
   in SDK version 41 but because it is already deprecated it has no
   effect.

The result is that `Foo` is marked as deprecated in 40, as required.
This will be fixed in a follow-up change.

Bug: 409725916
Test: ./gradlew
Change-Id: Id3f346d788962d4c5dfb7acfbd7e532ef07c92bb