Differentiate formatting of value and non-value annotations
Value annotations, i.e. `AnnotationItem`s that are referenced from
`AnnotationValue` are formatted differently in Kotlin than they are in
Java and that non-value annotations are formatted. In Kotlin they do
not have a leading `@` and require parentheses but the other two uses
do have a leading `@` and parentheses are optional. This change adds
support for differentiating between the two uses.

Bug: 354633349
Test: ./gradlew
Change-Id: I033e4231955179dc63836e46f94d6f59d96b417f