Move AnnotationValue formatting code to AnnotationItem
Previously, `AnnotationValue` did its own formatting of an
`AnnotationItem` to avoid changing the behavior of the existing
formatting done in `toSource()`. This change moves that code to a new
`appendAnnotationStringTo(...)` method that sits alongside `toSource()`
in preparation for being used to implement it.

It uses `appendAnnotationStringTo(...)` to replace the call to
`toSource()` in `toString()` because `toString()` just needs a valid
string representation of the annotation for use in error messages and
a couple of temporary work arounds. It does not need all the complex,
and currently model specific, behavior of `toSource()`.

Bug: 354633349
Test: ./gradlew
Change-Id: I7939521cd2b3a57f8f3f2bdf021bd20692bb9258