Show deprecation inconsistency in compiler generated enum methods
The enum compiler generated methods are not tracked by Metalava at all.
They are filtered out of the `Codebase` whether it is created from
sources, or signature files. That also means they are not part of the
signature files. Instead their presence in the API is implied by the
class being an `enum`. That means that their history cannot be tracked
properly when generating from signature files.

Previously, they were still being added to the history in a couple of
ways:
1. When including historical API from jars the custom ObjectWeb ASM
   based code was not filtering them out like we do for the `Codebase`.
2. When including the API from the source `Codebase` they were being
   added explicitly (because they were not present in the `Codebase`).

The latter explicitly marked the methods as not being deprecated even
when the class was. That leads to inconsistencies in the handling of the
deprecated status for these methods.

This change adds a test to show that inconsistency.

Bug: 409725916
Test: ./gradlew
Change-Id: I1815b55561f98af0ac49e03d4402e59cf1902e48