Make compiler generated enum methods use class history
Previously, the history of compiler generated enum methods, i.e.
`values()` and `valueOf(String)` was tracked separately and was
inconsistent with the containing class. This change removes that
tracking and instead adds them at the end with the exact same history
as the containing class.

An alternative solution of just dropping them altogether was debated
debated but that could impact downstream consumers of the
`api-versions.xml` file.

Bug: 409725916
Test: ./gradlew
Change-Id: I0fd82003dada288051657116f48711e17ad41a2c