Autoupdate of 2 submodules
*external/compose-hero-benchmarks/pokedex-views updated to dc749b98c8f0395f27e956e425249b271b8af4fd
Disable buildConfig generation for pokedex-views

This project does not use buildConfig so generating it is not needed.
It also fixes the issue where despite this project having no tests,
our build tries to run them, this happened because when we check whether
this project has test sources, buildconfig file was found.

Bug: 416000929
Test: ./gradlew :compose-hero-benchmarks:pokedex-views:app:GenerateTestConfigurationreleaseAndroidTest
Change-Id: Ib1162c93254ff1d71a11423f5d19532cfe0bf0d8

*external/compose-hero-benchmarks/poxedex-compose updated to 5006b3b68fc074bab9eca68e8a579c78fbd16dbc
Disable buildConfig generation for poxedex-compose

This project does not use buildConfig so generating it is not needed.
It also fixes the issue where despite this project having no tests,
our build tries to run them, this happened because when we check whether
this project has test sources, buildconfig file was found.

Bug: 416000929
Test: ./gradlew :compose-hero-benchmarks:poxedex-compose:app:GenerateTestConfigurationreleaseAndroidTest
Change-Id: Ifb311e61aaeff4b53794e83692cdcced7f2cb5c2