Add ValueStringConfiguration.specialValues
Previously, `Settings.stringReplacement` was added to allow
`LegacyValueFormatter` to select different special values to use for
each legacy value use site. This change adds a similar `specialValues`
property to `ValueStringConfiguration` and moves most of the string
replacement uses to use it instead.

This is done for a couple of reasons:
1. It slightly simplifies the formatting of `FloatValue` and
   `DoubleValue` which each had to handle the special cases explicitly.
2. The `AnnotationItem.toSource()` will use different string
   representations for the values depending on the source expression
   used so it will need something similar for that so using it for
   both allows the code to be reused.
3. It is possible that in future we will need to have slightly
   different representations for the special values, albeit more
   consistent and logical.

Bug: 399365344
Test: ./gradlew
Change-Id: I9e1ae03d629172549170486d7960532b24fd41fa