Sync XrExtensions JARs to Impress-0.0.4.
The extensions are now synced to http://mhcr/374826,
which is also the live version in g3.

This version matches what was used to build impress-0.0.4
in http://aosp/3626752.

This is a backwards sync, but:
a) is necessary in order so we can do an AndroidX drop
b) is ok because we already confirmed JXR is not using any
XrExtensions methods that are scheduled for removal next week.

Change-Id: I771de7778aeddcc33ee408fa916dbdd616f024bc