Replace ToValueStringDependsOnSourceForm with DefaultNumericValue
Previously, only a couple of `Value` implementations needed to track
whether they were specified as an int in the source. Now, all of the
numeric values, apart from `DefaultIntValue`, need to do that. Rather
than treat the latter specially this just adds a `DefaultNumericValue`
super class and moves the functionality into there.

Bug: 399365344
Test: ./gradlew
Change-Id: Idc2920bf32901b693f0664b1339cbb8dc968d46d