Track wasOriginallySpecifiedAsInt for byte/short/int
Previously, `wasOriginallySpecifiedAsInt` was only needed for `double`,
`float` and `long` as those were the only types affected by it when
formatting a field and/or method default value. However, the other
types need it for `AnnotationItem.toSource()`. Doing it for `int` does
seem a little redundant but keeping things consistent helps.

This change updates the factories for the types to pass in the
`wasOriginallySpecifiedAsInt`. That changed the expected values. That
in turn required some changes to the `PsiValueFactory` to handle Kotlin
in the same way as Java and some changes in `TurbineValueFactory` to
handle byte and short in the same way as Psi.

Bug: 399365344
Test: ./gradlew
Change-Id: Ic11b6b50b95b4a9cfa248a7b2185cb2be4051ff6