Allow spaces around = in annotation representation
The `LegacyValueUseSite.ANNOTATION_TO_SOURCE` uses
`AnnotationItem.toSource()` to format an annotation with a single
attribute and then extracts the value representation from that.
Previously, it assumed that there would be no space between the `=`
and the value but that will not be true in future. So, this change
tweaks the code to trim any such spaces.

Bug: 399365344
Test: ./gradlew
Change-Id: I712b7daef52d2d5b3132a04764b482a7b1ef6708