Fix CommonAnnotationItemTest for text model
There were a number of differences between the signature files and the
corresponding source files that caused the signature files to fail as
well as some errors. This change fixes them all.

One effect of fixing the test is that one of them that was passing
before started failing but that will be fixed in a follow-up change.

Bug: 354633349
Test: ./gradlew
Change-Id: I2152b80121590708445650f1cd31b5b0c3306314