Track whether a FieldReferenceValue is for a companion field
The legacy formatting of `AnnotationItem.toSource()` will include the
name of the `Companion` class for a companion object field. However,
the legacy formatting of `MethodItem.legacyDefaultValue()` does not.
In order to support them both from the same `FieldReferenceValue` it
is necessary for it to track whether it is for a companion object or
not.

This change adds `BaseFieldReferenceValue.kotlinCompanionClass` to
track that information and `showKotlinCompanionClass` to
`ValueStringConfiguration` to control whether it is included in the
formatting or not. A follow-up change will integrate that into the
`PsiValueFactory`.

Bug: 354633349
Test: ./gradlew
Change-Id: I81d407a4365efd344e1ff0a2efc3b9fb7e843554