Stop PsiValueFactory from optimizing field values
Previously, `PsiValueFactory` would optimize the handling of field
values by only checking for `ConstantValue`s. Unfortunately, that lost
information that is needed to handle some edge cases. This change makes
`PsiValueFactory` treat field values similarly to annotation values
(with the exception that the former can be `null`).

While this change does not fix any existing behavior it is needed for
follow-up changes that add support for explicit conversion functions in
Kotlin.

Bug: 354633349
Test: ./gradlew
Change-Id: Ie5492d15313eb898ead9ff3fe1b8d85e8fca1279