Allow CodebaseSnapshotTaker to resolve hidden class
Previously, `CodebaseSnapshotTaker` would throw an exception if it was
asked to resolve a class that was part of the original `Codebase` but
which was not part of the snapshot, e.g. because it was hidden. This
change adds a test to show that behavior and changes the behavior to
return `null` just like a normal `Codebase` that cannot resolve a
class.

This is needed because there can be `FieldReferenceValue`s which
reference classes which are not part of the API but which do need to
be written to the stubs, e.g. the `SdkConstant` annotation is a source
retention only annotation that needs to be written to the stubs as
explained in https://r.android.com/1830332.

The test added by this change specifically tests the behavior of
`SdkConstant` as it was that which caused an issue in the Android
build. As part of that this marks `SdkConstant` as `@hide` to match
how it is defined in Android.

This was not previously an issue because `PsiAnnotationItem` overrides
`snapshot(Codebase)` to essentially do nothing and so references to
`SdkConstant` from those annotations were resolved in the original
`Codebase` not the snapshot. However, a follow-up change will remove
the `PsiAnnotationItem` subclass and use `DefaultAnnotationItem`
instead. That implements `snapshot(Codebase)` correctly and so the
`SdkConstant` reference will be resolved against the snapshot.

Bug: 354633349
Test: ./gradlew
Change-Id: Ib82cb98191e7cbfea6acd30e9a8d000f8b36c12f