Remove defaults from ModifierListWriter constructor
Previously, the constructor had defaults for two of the parameters.
However, the constructor is private and only called from two factory
methods, each of which sets the defaulted parameters to different
values. So, the defaults do not save much code and makes it harder to
reason about so this change removed the defaults.

Bug: 354633349
Test: ./gradlew
Change-Id: If98e7270cc27cfc534a1cde4e6f455f65aa2793d