Add AnnotationItem.createMarkerAnnotation(...)
Quite a few places create marker annotations, i.e. annotations without
any attributes. Previously, they used `Codebase.createAnnotation(...)`
which takes a source representation of the annotation and parses it. In
order to use that they had to create the source representation (i.e.
add an `@` prefix to the annotation class name). All of that work is
unnecessary as they can just create the `AnnotationItem` directly from
the annotation class.

Note: It used to be necessary to do it that way to create a model
specific implementation of `AnnotationItem` but they no longer exist.

This change adds an `AnnotationItem.createMarkerAnnotation(...)` to use
instead.

Bug: 354633349
Test: ./gradlew
Change-Id: Icd607cb4669f585f15603333f43b14d8eae09f43