Provide non-lazy AnnotationItem implementation
Previously, `DefaultAnnotationItem` was the only `AnnotationItem`
implementation and it obtained its `attributes` lazily. However, there
are quite a few places where the attributes do not need to be obtained
lazily, e.g. either because there are none, or the values are already
known.

This change makes `DefaultAnnotationItem` a non-lazy implementation and
adds a separate `LazyAttributesAnnotationItem`.

Bug: 354633349
Test: ./gradlew
Change-Id: Id38d1c187993441bdd75fe9d9f52246ab2d1cd32