Add --api-version-range
Previously, the range of historical versions that could be included in
the API version history was determined by the `--first-version`,
`--current-version` and `--current-codename` options. The latter two
options also affected many other aspects of the API version history
which made it very difficult and sometimes impossible to control the
historical versions that were included without having unwanted side
effects on other aspects of the history.

This change adds an optional `--api-version-history` option that falls
back to the existing behavior if not provided. It adds some new tests
and also modifies one existing one to make sure it works as expected.
In that case it was previously specifying:
* `--first-version 30`
* `--current-version 32`
* `--current-codename Current`

So, the effective range was from `30:33`. The `33` comes from the
`--current-version + 1`. The `+ 1` is needed because
`--current-codename` was specified to a value that was not `REL`.

Bug: 424440294
Test: ./gradlew
Change-Id: I04769ee9e0afa810406524c1140807b653efd04f