Add --api-version-label
Previously, the `--current-codename` could be specified when generating
doc stubs (or when enhancing the documentation of normal stubs) and it
would be used as the label in `@apiSince` and `@deprecatedSince` for
any version that was equal to or greater than `--current-version`. The
purpose of that was to allow documentation to be generated for a
development release for which the numeric version was not yet know.
Unfortunately, `--current-codename` also had an effect on the
generation of the `api-versions.xml` file so it made it difficult, if
not impossible, to specify it correctly when doing both.

This adds an `--api-version-label` option that can be used multiple
times to provide mappings from an `ApiVersion` to a label. If
specified, then it will be checked first before falling back to use
the `--current-codename`.

Bug: 424440294
Test: ./gradlew
Change-Id: Ia332f5d8849f78ebe55fdf64c0698eb6ebad7f29