Add --api-version-for-sdk-extension
Previously the API version to use as the next SDK version into which
newly added SDK extension APIs would be assumed to appear was set using
a combination of `--current-version` and `--current-codename`. This
change adds a separate API to use instead to provide more control over
the `api-versions.xml` generation, updating the test added in the
preceding change to use the new option.

This also adds tests to show how to use it to generate
`api-versions.xml` for a standalone SDK extension, i.e. one released
separately to a dessert release and one that is released as part of a
dessert release.

The former test reveals an issue when generating an `api-versions.xml`
file from historical API versions and historical SDK extensions and the
current source. In that case APIs that are added in the current source
but are not present in the most recent historical SDK extension can be
mistakenly marked as having been removed. A TODO was added to explain
the issue and it will be fixed in a follow-up change.

Bug: 424440294
Test: ./gradlew
Change-Id: I8759e09b4ad1ee088271efdc05911f06ca48d8ed