Revert "Fix handling of flagged super method with no previously released API"
This reverts commit 05fe8a66865f5de8a4202e70555e4f0131501a4c.

The change was added to work around an issue caused by trying to revert
a flagged API without having a previously released API to which it
could be reverted. That was due to incorrect `Android.bp` files in the
Android build. Those have now been fixed and the following change will
cause Metalava to fail immediately if it is asked to handle flagged
APIs without a previously released API. That makes the work around
pointless so this change removes it.

Bug: 379940628
Bug: 425309457
Test: ./gradlew
Change-Id: I413589a11ab4e20ffdd574af3b465aebfcf25314