Hide @ExtensionFunctionType in java docs for integration tests
This annotation is no longer present if we switch dackka to k2, and the dokka team has said the issue is low priority since the annotation is internal and only used by the compiler. It won't really be useful to java developers anyway, so remove it from the docs now to minimize the diffs from k2.

Parallel change to aosp/3672614 which updates the actual androidx docs.

Bug: 392589580
Test: ./gradlew test
Change-Id: I50c23e2a46d88448a39a139a389dbbbeee6f22f5