Move isSamCompatibleOrKotlinLambda from ParameterItem to TypeIten
With overrides for specific subclasses of TypeItem

This adds tests for various cases of SAM compatibility, handling the variable type case which wasn't previously tested. The trailing lambda syntax can be used in Kotlin if a variable type has a Kotlin lambda type as its bound, but not if it has a different SAM compatible type as its bound.

Bug: 354889186
Test: ./gradlew test
Change-Id: I1c62b64a4591293dfaa8e1e2a1a39e29c84de742