Update how dackka determines if a class is synthetic
Previously dackka relied checking file-level annotations to tell if a DClasslike was synthetic (checking if `file:JvmName` is used). However, K2 changes how file-level annotations are handled, so non-synthetic classes defined in a file will have the file-level annotations attached.

This attaches an extra property to synthetic DClasses when dackka creates them, so that to tell if a class is synthetic all that's needed is to check whether the property is present.

Bug: 392589580
Test: ./gradlew test
Change-Id: I708fdeb22aa78a5c534b07fe73797e554cb7b5c3