Only create implicit default constructors for Java source classes
For Kotlin, if there is an implicit no-args constructor, it gets included in the psi for the class. For Java, it does not. This makes it possible to simplify `hasImplicitDefaultConstructor` by checking whether a class is Java source.

The followup changes start generating some constructors with the kotlin analysis API and from bytecode after the initial generation of the codebase from psi occurs, so this change prevents a no-args constructor from being added before the real constructors are added to a class.

The only test changed here is for an inline class, which is a deprecated feature in favor of value classes, and the removed constructor should not have been present.

Bug: 417740481
Bug: 427753622
Test: ./gradlew test ("Basic Kotlin class" in ApiFileTest covers this case, among many other tests)
Change-Id: I75fcbf615c18be7d17862c901fe85c5f74d9e560