Autoupdate of 2 submodules
*frameworks/support updated to 2bb6aeb8565c1ad2800dbb81d27e026dc3ca292e
Merge "Temporarily disable androidx_with_metalava" into androidx-main
*tools/metalava updated to b4859067fff0caaa9dadb5ac8c880e6302433d8d
Track bytecode constructors with DefaultConstructorMarker when needed

The compiler generates extra overloads of kotlin source constructors with a DefaultConstructorMarker parameter added (if a constructor has an optional parameter, there is also an additional overload with both an int and DefaultConstructorMarker parameter added). In general, these aren't important for API tracking since we can just do compatibility checks with the regular version without the DefaultConstructorMarker parameter.

If a constructor uses a value class type, it can only be used from Kotlin source. The only version of the constructor that will exist in bytecode is the DefaultConstructorMarker one, so in that case it makes sense to track it for binary compatibility checks.

This adds DefaultConstructorMarker constructors to the codebase when there isn't already a constructor present with a matching signature excluding the DefaultConstructorMarker parameter, since these will be the cases where the DefaultConstructorMarker constructor is needed for binary compatibility checks.

Bug: 417740481
Test: CommonTargetLanguageTest

Change-Id: I65f9487c4b0bc82ccac8f6fa121aac180c4291a9