Restrict Language annotation to parameter
When the `Language` annotation is supplied on a constructor property it
currently only applies to the parameter but an experimental feature
will cause it to also be applied to the field. Without enabling that
experimental feature it produces a warning that breaks the build.

Rather than enable the experimental feature this explicitly restricts
it to just the parameter.

Bug: 427971908
Test: ./gradlew (before and after switching to Kotlin 2.2)
Change-Id: Iff997899599e4940f99ed450d0f15ed6a2fac9f3