Update display of varargs in Java docs
Currently, vararg parameters from Kotlin source are not dislayed in the Java docs with any indication that they are varargs. This adds "..." after the parameter type in the Java type to match Java varargs syntax.

Note that this only works for Kotlin source varargs. Java source varargs are displayed as a regular array "[]" type (b/293340652)

Bug: 422715275
Test: DefaultParameterComponentTest, updated integration tests
Change-Id: I44299dc77d09b836a5504395318d9c1182056e1d