Restructure how stub-annotations are added to tests
Previous mechanism relied on project(":stub:annotation") to be
configured before every other project as it was poking deep into the
task graph of a different project. This is not Gradle project isolation
safe. https://docs.gradle.org/current/userguide/isolated_projects.html

Instead, move to using a configuration and resolve it as a
FileCollection in a lazy way to get the jar. This set up makes sure that
the jar is built only when needed. This uses a similar mechanism to
https://www.liutikas.net/2025/01/08/The-Task-Addendum.html

This change also cleans up stub-annotations to have no warnings when it
builds

Test: ./gradlew  -> still passes, but now with fewer violations of
                    Project isolation
Bug: None
Change-Id: I5747b652573b621b2c49843007afe04df4ab116d