Add basic ANTLR4 grammar for Javadoc
In preparation for supporting parsing Javadoc this adds a basic ANTLR4
grammar from the ANTLR github grammar examples from
https://github.com/antlr/grammars-v4/tree/753536777d827ccc0c9b108531ea67375c2039ac/javadoc.

The files have slightly different names but are otherwise unchanged
from the original version and are currently unused. They are added
unchanged and unused to make it easier to see what local changes have
been applied to it.

Bug: 429965593
Test: ./gradlew
Change-Id: I2d26e912382ce93e58d1c86a6128a4bc2c030d56