Autoupdate of 2 submodules
*frameworks/support updated to 92d5ee1d3a724ecadbed479d35c8e46d6e903138
Merge "Temporarily disable androidx_with_metalava" into androidx-main
*tools/metalava updated to fcc2aa08f691f28bf9b8198ade0137d3fe06fb4c
Update creation of reified inline methods

Reified inline methods can't be used from Java. When they are called from Kotlin, the usages are inlined, so while they are present in bytecode, they are never called in a binary (it is an error if they are).

This change stops reified inline methods from being created from psi, and instead creates them with the analysis API in KaCodebaseAssembler. Then, in KotlinBytecodeApis, it prevents any reified inline signatures from being added (the binary signature might not be present if the function uses a value class type and the bytecode name is mangled), since the bytecode forms cannot be used so they don't need to be tracked for compatibility.

Bug: 421201575
Test: CommonTargetLanguageTest
Change-Id: I4ea070ef9b9a1a25a8827237a5c3f5c1fa13af12