Autoupdate of 2 submodules
*prebuilts/androidx/internal updated to f4bee1b142e7804396a236aea422eab27707fee0
Merge "Update compose foundation prebuilts to remove extra () after @sample" into androidx-main
*tools/dokka-devsite-plugin updated to bfb04c7d9db4899ed2f2d400a1ce3abd5d5f151e
Support interleaving samples tags with other documentation for K2

When switching to K2, the parser stop being able to handle the case where a @sample tag is interleaved with other documentation and links are used between the tags. To work around this, we can apply a preprocessing step of replacing `@sample` tags with `@author #@sample`, which will make the parser interpret it as an author tag, which dackka doesn't support. Then, dackka can inspect author tags to find the transformed samples.

This adds the preprocessing step to the dackka integration tests. It will also need to be added to the androidx build when androidx switches to K2 dackka.

Test: Test output changes are minor -- the compose change removes a `<p>` tag around some text which is in the kdoc inline with the sample tag, which won't really change the rendering. The sampleAnnotation change is to start processing a java-style sample that previously wasn't being handled. Also tested this using K2, and found the only other change to test outputs was to correct the broken links that occur without this change.
Bug: 427708573
Change-Id: I5e9f04c8410f378f78135157726da8755d16ee22