Fix snapshot handling of PackageItem documentation
Previously, `CodebaseSnapshotTaker` would create a `PackageDocs` object
that only contained a single `PackageDoc` object for the `PackageItem`
being snapshot. It did so because it expected any ancestors of that
`PackageItem` to have already been snapshot. However, that was not
true.

e.g. A `Codebase` containing a single `test.pkg.Foo` class would have
the following `PackageItem`s:
* `` - i.e. the root package.
* `test`
* `test.pkg`

Of those, the first two would have `emit = false` which means that they
would not be visited by `CodebaseSnapshotTaker` and it would only see
`test.pkg`. That meant that the documentation of empty packages, i.e.
`test` was not snapshotted correctly.

This change fixes that by creating a `PackageDocs` containing a
`PackageDoc` for the `PackageItem` being snapshot and any of its
ancestors which have not already been snapshot.

Bug: 429965593
Test: ./gradlew
Change-Id: Ic7904181a1f47f86438b8974a66e2f4f4523bcbe