Handle full version codes in RequiresApi
Previously, `DocAnalyzer.handleRequiresApi(...)` assumed that the value
supplied to `RequiresApi` was an API version level, i.e. a major
version. However, that is not completely true. It can also be a full
version code which encodes both major and minor. This change detects
that and decodes into a `major.minor` version.

Bug: 424435764
Test: ./gradlew
Change-Id: I7693dc06523c283978f3ecbf9cd47f4019a15a67