Stop treating @suppress/@pending as @hide
Previously, `@suppress` and `@pending` doc tags would be treated the
same as `@hide`. However, neither of those is really necessary and
adds work and complexity to Metalava.

AndroidX does not use `@suppress` to control the API surface. in fact
it does not process the documentation at all. There are only two uses
of `@pending` in Android that affect the API surface and they can
easily be replaced with `@hide pending`.

This change removes support for `@suppress` and `@pending` and fixes
up tests that use it and removes a test that just checks `@suppress`.

Bug: 429965593
Test: ./gradlew
Change-Id: Idd4d305e7c1fe814942e55738bb88c5fce8c712f