DocCommentParser: Add special case for @hide required by Android
While `@hide` is mostly treated as a block tag in Android there are
some other forms that it takes. That is because the current
implementation just looks for `@hide` in the text without regard to the
surrounding text and so developers have used different forms of it
that are not block tags as defined by Javadoc/KDoc.

These different forms are:
* `....some text... {@hide} - i.e. an inline doc tag.
* `....some text.... @hide` - i.e. a block tag at the end of some
  text.
* `* {@hide} - i.e. an inline doc tag being used where a block tag
  would be.

This change adds tests for those cases and updates `DocCommentParser`
to support them.

Bug: 429965593
Test: ./gradlew
Change-Id: I788f0fa4847314b6479823d36cab34a34d1ec386