Combine PsiItemDocumentation.javadoc(...) and initializeText()
Previously, `javadoc(PsiElement)` would return the text of the
documentation comment to `initializeText()` which would optionally add
`extraDocs` and then store it in the `_text` field. This change
combines `javadoc(PsiElement)` and `initializeText()` into a single
`initializeFromPsiElement(PsiElement)` which stores the value in the
`_text` field directly. The caller will append the `extraDocs` if
needed.

This is preparation for a follow-up change that will record the
`PsiComment` from which the text was retrieved so that it can be used
to provide a more accurate location for any issues that are reported in
the documentation comment.

Bug: 429965593
Test: ./gradlew
Change-Id: I9426400a629d5c6053e477abb6a09911ee8a51c1