Add ItemDocumentation.fileLocation
Previously, `ItemDocumentation` did not track the location of the
document comment making it impossible to accurately report the location
of an issue in the comment. This change adds `fileLocation` to it and
implements that in `PsiItemDocumentation`. It also adds some tests to
verify the behavior, baselining them for Turbine which does not yet
support that.

Bug: 429965593
Test: ./gradlew
Change-Id: I790c56866f789aee6d4f7b72a44a0a5fff9e10d0