Disable kernel ltp test with hugetlb mmap
Tests are creating a mount point on the tmpfs
and performs mmap on the mount point.
creating mount point on the tmpfs to copy hugepages
on a device running with AOS/AAOS not needed.
during vts test, test fails with EACCES(13) to create a
mount point on tmpfs.
No usecase identified with such a request.

Bug: 436306632
Test: Run below test module, it's passing.
run vts -m vts_ltp_test_x86_64

Change-Id: I9d9633281c4143c42114df9d51f00b39f8e5a0ef
Signed-off-by: V,Anilkumar <anilkumar.v@intel.com>