Add --treat-as-error command line option
Previously, the only way to treat a `Severity` like `WARNING` as a
different `Severity`, e.g. `ERROR` was to use `--warnings-as-errors`.
However, that treats both `WARNING` and `WARNING_ERROR_WHEN_NEW` as
`ERROR` which defeats the whole purpose of the latter which is to not
break the build. That behavior is necessary for AndroidX which does not
actually support the `WARNING_ERROR_WHEN_NEW` behavior anyway but it
does not work for Android which does support `WARNING_ERROR_WHEN_NEW`.

This change adds another option `--treat-as-error` which specifies the
`Severity` to be treated as an `ERROR`. That will be used on Android
instead of `--warnings-as-errors`.

Bug: 438246528
Test: ./gradlew
Change-Id: I4114826296fbea451895fc34564a5a6283b3d16f