Remove unused ANTLR Javadoc lexer and grammar rules
Initially the ANTLR Javadoc lexer and grammar was created to parse the
whole Javadoc comment. However, some of that work was moved into
`DocCommentParser` for performance reasons and the preceding change
make `JavadocParser` only handle the `description` rule. This change
removes the rules which are no longer used.

Specifically, it removes rules which are related to the start and end
of the comment and rules that handle the main description and block
tags. The remaining rules are related to handling a description block,
i.e. a mixture of text and inline tags.

Bug: 429965593
Test: ./gradlew
Change-Id: I1b723055357802d9ddca092ac94c855396470cd7