Skip testCallbacks when unrestricted interfaces are present
The old assumption check ran the test on devices with unrestricted
interfaces which were configured in the overlay. Apparently, this is
a common configuration on automotive devices.

There are likely other tests that are broken which need to be moved to
the new assumption check.

Test: TH
Flag: EXEMPT mainline
Merged-In: I504534920ea2c083a05bf217bec17a084a829184
Change-Id: I504534920ea2c083a05bf217bec17a084a829184