Improve error reporting when patching SdkExtensions history
The `android/os/ext/SdkExtensions` class has its history patched to
retroactively make it appear available in versions where it was present
on the platform but not yet ready for public use. That is to make it
easier for application developers to target those versions. The
patching code does some consistency checks to make sure that the actual
history is as expected. Previously, the error message was not very
informative, this change adds some more explanation and details to make
it easier to understand the problem.

Bug: 259115852
Test: ./gradlew
Change-Id: Ieeca65543a4f45d43a64c6ecf33d037b28835a0a