Add test to show InvalidJavadoc issues are suppressed by baselining
If multiple `InvalidJavadoc` issues are reported in the same file only
one is reported when baseline is enabled. That is because the issue is
reported without a `Reportable` and only a `FileLocation` and the
baseline key for the latter only uses the path not the line number to
prevent the baseline from becoming out of date just because some lines
were added or removed.

This change adds a test to show that behavior and updates an existing
test to show the behavior without baselining.

Bug: 429965593
Test: ./gradlew
Change-Id: I77d9205fb122b53387b3efbb2729dac13e71fb96