Avoid searching documentation text for @deprecated
Previously, `Item.documentationContainsDeprecated()` would search the
documentation text for `@deprecated`. That is inaccurate as it will
match documentation that contains `@deprecated` that is not a block
tag or something like `contact@deprecated.com`. This change switches it
to use `hasBlockTagOfType("deprecated")` instead and reorganizes the
remaining code to avoid getting `text` unless needed.

Bug: 429965593
Test: ./gradlew
Change-Id: I1cc72a51d4796478abcad253d854abec58e6d83e