Autoupdate of 2 submodules
*frameworks/support updated to da70d9cab3e3172827b3cd12a739e372b1b68eec
Merge "Temporarily disable androidx_with_metalava" into androidx-main
*tools/metalava updated to 90e8ddb0794645e7b72b129f6bd00eec65e319c4
Update creation of APIs from delegates

Previously these APIs were marked as InaccessibleFromJava, even though they can be used from Java. Update their creation to include all target languages.

It could be possible for there to be a method from a delegate that is not accessible from Java, for instance if it used a value class type. However, it has been difficult to find a reliable way of telling if the delegate method can be used from Java without special casing certain situations (it should be possible to do by looking at the psi of the kotlin analysis function or by checking the super methods metalava has created, but those aren't working when using mapped kotlin collections types). Currently, the number of delegate APIs in AndroidX is very small (see aosp/3763424) and this isn't an issue, but this could be revisited in the future if that changes.

Test: ApiFileTest, CompatibilityCheckTest
Bug: 323223405
Bug: 215389034
Change-Id: Id1a0eeef3a610670c76b1cde316e20a6d7c7a518