Separate inlineTag rule from braceText rule
When building a model from the parser rules it will be necessary to
treat `inlineTag` differently to the rest of `braceText` (which is just
different forms of textual content). Previously, `inlineTag` was mixed
in with `braceText` which would make that difficult.

This change moves `inlineTag` from `braceText` to `braceContent` to
make that easier.

Bug: 429965593
Test: ./gradlew
Change-Id: I9c4e6ca50c17c3d715aa8beaf0e47a3e5b062eca