Separate main description and block tag section if needed
Previously, the main description and block tag section were not
separated by a blank line. That would cause a lot of insignificant
changes to the Android stubs. Just adding a blank line between them
will also cause a lot of differences in the Android stubs. This
change minimizes the differences by adding a blank line only when
the block tag section itself requires multiple lines.

So, the following does not have a blank line:
```
/**
 * Main description.
 * @deprecated deprecation reason.
 */
```

But the following does:
```
/**
 * Main description.
 *
 * @deprecated deprecation reason.
 * @see OtherClass
 */
```

Bug: 429965593
Test: ./gradlew
Change-Id: I517e4451b1f5bf2f07bf2cb84c06db52746bbbdf