Print javadoc model as a Javadoc comment
Preparation for a future change that will use this to print javadoc
comments to documentation stubs. This change prints the Javadoc comment
using the `JavadocContent` model and not the source text.

The initial printing is quite simple but will result in a lot of
insignificant changes in the documentation stubs. That would make it
difficult to see significant differences that indicate problems in the
handling of the Javadoc. Follow-up changes will improve the printing
to reduce the number of insignificant differences.

Bug: 429965593
Test: ./gradlew
Change-Id: I1230f242b327251bbea82cb157ad124c84abf1c5