Correct the line number and character offset in JavadocParse errors
Previously, errors that were reported in a block tag description had
the wrong line number of character positions as the `JavadocParser` did
not correctly compensate for the fact that the `JavadocParser` was
being called on text that had been extracted from another block of
text.

This change fixes them both by correcting the offsets and in the
process removes some unnecessary parameters from the
`JavadocParser.parse(...)` method.

Bug: 429965593
Test: ./gradlew
Change-Id: I5246887d351b724ae0ad03bf18d2825b4d930d4f