Autoupdate of 2 submodules
*frameworks/support updated to 3b263ecbeeeebbfb9a9bb001234fe3a548fb3975
Merge "Update baselines for Kotlin-only functions which are no longer linked to the previous definition" into androidx-main
*tools/metalava updated to dd647c0a594d75b8a490efe375747474b122717e
Update when parameter types are considered the same in CallableItem.matches

If a callable exists in bytecode, it makes sense to treat methods as matching if they have the same erased parameter types. If the callable is kotlin-only, erased parameter type equality isn't enough.

However, requiring an exact type match means that a lot of cases where signature entries differ between `? extends T` and just `T` would no longer be considered equal. To allow for that case to match, this adds a transformer to flatten wildcards for comparing types.

There are still some cases where this update means that a type from an old signature file and a new signature file don't match in AndroidX, due to differences in how non-final varargs are listed (b/447153877). Updated ignore files for those cases are in this topic.

Bug: 445490139
Test: CommonSuperMethodsTest
Change-Id: I3778d44b55c86c649c65f1a1b03d7fe4134c4c7b