Do not indent with ` *` if first char is `/`
Previously, if a javadoc comment contained a line that started with a
`/` then it would be indented by ` *` resulting in ` */` in the middle
of a comment which would end the comment prematurely causing compiler
errors when compiling a stubs class that contained that comment. This
change detects that case and does not indent with ` *`.

Bug: 429965593
Test: ./gradlew
Change-Id: I5408f30616dc6aba350ed0dd1953ef9ee87c5981