Add DocCommentParserTest to show behavior of multi-line inline tag
When an inline tag is split across multiple lines such that its inline
tag type is on one line and the nested content starts on the following
line the current printing code leaves a trailing whitespace after the
inline tag type and before the end of the line, e.g.
```
 ...
 * {@code <-- trailing whitespace
 * the nested content}
 ...
```

This change adds a test to show that behavior.

Bug: 429965593
Test: ./gradlew
Change-Id: I8915e7f9de3c624189f4ccd70b729a2dcfc53009