Fix handling of NEWLINE in inlineTag
The previous change added a test to show the incorrect handling of a
`NEWLINE` token in an `inlineTag` rule. This change fixes that by
reusing the `descriptionNewline` rule and its special processing as
part of the `inlineTag` rule. It renames the `descriptionNewline` rule
to `newline` to avoid it from being considered to just be part of the
`description` rule.

Bug: 429965593
Test: ./gradlew
Change-Id: I15512a2144282e50774861e0c55071b66b2ff91e