Add JavadocParserTest to show behavior of newline in an inline tag
The `NEWLINE` token matches a newline character and then an optional
indent consisting of some spaces followed by an `*`. The purpose of
that is to make it easy to remove the indent (which is not considered
part of the documentation) while preserving the actual documentation
when constructing the model.

A `NEWLINE` token is wrapped in a `descriptionNewline` rule which is
processed specially by the `JavadocParser` to replace the whole token
with a single newline character. However, a `NEWLINE` in an `inlineTag`
rule is just treated as a piece of text so the whole token, including
the indent is copied into the model.

This change adds a test to show that behavior. A follow-up change will
correct that.

Bug: 429965593
Test: ./gradlew
Change-Id: Ifc9d9fcdaa7595a506c277c1d90dded4d5437892