Use FlaggedApi annotation location in FlaggedApiLiteral messages
Previously, `FlaggedApiLiteral` issues were reported on the annotated
item rather than the `FlaggedApi` annotation. That meant that a change
which just added a `FlaggedApi` annotation with a literal string would
report an error on a different line to the actual change which meant
that the Build Warnings Aye Aye Analyzer would not associate the
warning-error-when-new with the change and so not treat it as an error.

Bug: 442395516
Test: ./gradlew
Change-Id: Ic916202772d43d016630afac46b6529f774af4d8