Fix handling of leading asterisks/whitespace in trimDocIndent(...)
Previously, leading asterisks/whitespace was not handled correctly in
`trimDocIndent(...)`. e.g. it:
* removed common indentation, possibly removing necessary indentation
  in `<pre`> blocks.
* added a leading space when unnecessary, changing relative indentation
  in `<pre>` blocks.
* did not handle multiple leading asterisks correctly, should be
  treated as one asterisk.

This change treats each line separately, adding no extra indentation.
That does mean that lines inside a `<pre>` block which did not use a
leading asterisk could be indented differently compared to how it
previously behaved. However, the behavior will be consistent and easily
fixed in the source documentation if necessary.

Bug: 448617351
Test: ./gradlew
Change-Id: I1af8597c22dfb898553d22a40624db7db0e2277c