Add characterPosition to FileLocation
Previously, it only included `Path` and `line`. While that was fine for
identifying a specific API element which is always on its own line(s),
or (in the case of parameters) has a unique name on the line it is not
sufficient for Javadoc issues which could have multiple similar issues
on the same line. e.g. both `@link` tags are invalid as they do not
have a closing `}`, they have a closing `)` instead.

```
* ...
* Use {@code SomeClass) when dry. You can also use {@code SomeClass) if
* it is only raining a little bit in April.
* ...
```

The only thing that differentiates them is their character position.

This change adds `characterPosition` to `FileLocation` but only
implements it in `PsiFileLocation`. This does not implement it in
Turbine because Turbine does not currently track the `FileLocation` of
`ItemDocumentation` instances so adding `characterPosition` would not
have any benefit.

The `characterPosition` is only initialized for `PsiComment`s because
as mentioned above it is not necessary for other uses. It would also
cause lots of churn in tests that check for reported issues and fixing
those tests in Psi would mean that they would not work in Turbine.

Bug: 449688896
Test: ./gradlew
Change-Id: I49c7df7a6bf9b915d7677eafe8d2a14921dd4f9c