Dedup ItemDocumentation.text implementations
Previously, `DefaultItemDocumentation` and `PsiItemDocumentation` had
their own implementations of the `ItemDocumentation.text` mutable
property. That required some duplication of the implementation of the
abstract property `text`, specifically its handling of calling
`textChanged()` when it was set.

This change moves the functionality into `AbstractItemDocumentation`
deduping the `text` setter. That allows `textChanged()` to become
private.

Preparation for allowing `docComment` to be mutated.

Bug: 450228132
Test: ./gradlew
Change-Id: Ifba3696a37c57f257b2bb461120aad8967e0d201