Switch DocAnalyzer.handleInliningDocs(...) to DocContent/Owner model
Previously, it manipulate the `String` representation of the comments.
This change switches it to use the `DocContent/Owner` model.

This fixes a lot of Turbine tests that were broken because Turbine did
not implement some of the `ItemDocumentation` methods. This change
removes the last uses of them and they will be deleted in the following
change.

Bug: 450228132
Test: ./gradlew
Change-Id: I69d92443a8950b893895384ee9fc75c5f52405d9