Refactor DocAnalyzer.handleInliningDocs()
Previously, when checking to see if a `CallableItem` documented the
`null` policy it would concatenate the main description and return
description together before searching. That is ineffecient and makes
follow-up changes that switch from using `String` to
`DocDescription` more complicated as it requires `DocDescription` to
support being combined together in some way.

This change refactors the code to avoid that concatenation.

Bug: 450228132
Test: ./gradlew
Change-Id: Idc065e8e0a7f502a764fd19e5ecfb28c10986fd7