Add tests for Javadoc first sentence containing "e.g. "
Currently, `DocAnalyzer.tweakGrammar()` works around an issue with
Javadoc where it ends the summary line at the first `. ` even if that
is in `e.g. example`. This adds tests to show that workaround is not
applied when printing from `DocComment`. This is preparation for a
future change that will move the functionality from `DocAnalyzer` to
`DefaultDocDescription.printAsJavadocComment(...)`.

Bug: 450228132
Test: ./gradlew
Change-Id: I186b63c9aab3762324a72e0888969e5e5be80694