Sort block tags by standard-ish order
Previously, block tags were kept in the order they were found in the
and added by Metalava. That made it unstable and sensitive to changes
of the order in the sources or Metalava. This change orders the tag
types using a standard-ish ordering based on standard style guides but
tweaked to better suit Android sources.

Bug: 452699372
Test: ./gradlew
Change-Id: Iac084da296c90fee8a91efb4fe790196db5db669