Add test to show ordering of block tags
At the moment block tags are left in the order in which they were found
in the source and/or added by Metalava. That makes tests and stubs
prone to changing due to insignificant changes in the sources or
Metalava. This change adds a test to show that ordering, follow-up
changes will make a more stable order that is less sensitive to changes
in the source or Metalava.

Bug: 452699372
Test: ./gradlew
Change-Id: Iab8045cfe3c2d1170e4e06bdfd73b96c4cb02b51