Add DocCommentContext to provide context for DocComment
A `DocComment` contains information that is context dependent, where
the context is the item to which the comment applies. That can be the
in-scope symbols (including imports) for resolving `{@link}` and
`@see` tags, the order of parameters to ensure correct ordering of
`@param` tags to give two examples.

This change adds `DocCommentContext` to encapsulate that information
and pass it through to where it is needed. It is empty at the moment
but will be extended in future.

Bug: 450228132
Test: ./gradlew
Change-Id: I6275ea45fef3a519563060907d3a134639d1e946