Refactor DocDescriptionPrinter to JavadocContentPrinter
Previously, it would print `DocDescription`s but really they are just a
container for `JavadocContent`. This change renames it to match and
replaces `print(DocDescription)` with `print(JavadocContent?)`. It also
uses the latter to print the content of a `JavadocInlineTag` so now
there is one entry point for printing every block of `JavadocContent`.
That will make follow-up changes to add `TagType` specific printing
simpler as they will only need to modify one place.

Bug: 450228132
Test: ./gradlew
Change-Id: I7c2e2cc32028a3f5fbf2ceb2ed93e56a695c091f