Remove methods from DocDescription
The `DocDescription` class is really just a wrapper around and a
factory for creating `JavadocContent` objects with a few additional
methods (which are implemented in terms of `JavadocContent`). This
change removes those methods and replaces uses of `DocDescription` with
direct use of `JavadocContent` instead.

This is preparation for cleaning up `DocDescription` and how it is
used.

Bug: 450228132
Test: ./gradlew
Change-Id: I6aec5704c315aea650ea86faf8d27fae919c7842