Allow TagType.extractData(String) to consume some content
Previously, when `TagExtractor` extracted `TagData` from some
`JavadocContent` the latter was unaffacted. While that was ok when
just using the data for sorting or searching it does not work well
when mutating.

e.g. in `@param p` the `p` is treated as part of the `description`
of the `@param` tag. That causes problems when appending to it as
`description` is not empty and so it adds a `<br>` separator
resulting in something like:
```
* @param p
* <br>
* Appended text.
```

When it should be `@param p Appended text`.

This change allows a `TagType` to consume text from the content from
which it created the `TagData` by returning the number of characters
to remove.

Removing the content from the start will also remove it from the
printed output which would result in invalid Javadoc being generated.
The following change will address that by allowing `TagData` to
contribute to the printed output.

Bug: 450228132
Test: ./gradlew
Change-Id: Ic06b84ac63381c229d04aaef3edfab0cf276e530