Allow TagData to contribute to the printed output
When `TagType.extractData(...)` consumes some of the content when
creating `TagData` that content will no longer be printed as part of
the Javadoc resulting in invalid Javadoc.

This change rectifies that by adding `TagData.printAfterTagType(...)`
which allows the `TagData` to print the content that was consumed
when it was created.

Bug: 450228132
Test: ./gradlew
Change-Id: Id09e9c63d92331f73042a1fbacd7d79d11f58d71