Add concatJavadocContent {...}
While it is not required, it is helpful when constructing a
`JavadocContent` hierarchy to avoid it becoming too deep as that makes
it more difficult to understand and visualize. It can also make
processing slightly more expensive with the potential to produce stack
overflows if it becomes extremely deep.

This change adds a `concatJavadocContent {...}` method that makes it
easy to construct a `JavadocContent` by concatenating other instances
together.

Bug: 450228132
Test: ./gradlew
Change-Id: I3f51efc383bb24451d11c12df6e37b7b9ab3b0f8