Make blockTagDescriptionOwner(...) non-nullable
Previously, `blockTagDescriptionOwner(...)` would return `null` if
called for a block tag that was not present in the comment. That means
it is not possible to then use `DocContentOwner.append(...)` to append
some content to it, creating the block tag in the process. That differs
from the behavior of the `appendDocumentation(...)` method which will
create the block tag if it does not exist.

This change adds `DocComment.pendingBlockTagSection(...)` to fix that.
It returns a `DocContentOwner` implementation that when used to append
content will first add a `BlockTagSection` and then append content to
it.

Bug: 450228132
Test: ./gradlew
Change-Id: I95029e0e6b04668833f8fd364c77cc48feb935bc