Hide setter method for DefaultDocComment.blockTagSections
Previously, `DefaultDocComment.blockTagSections` could be mutated from
outside the class as its setter method was not private. That could
cause problems as changes to that property should invoke the
`DocCommentMutationListener`.

This change makes the setter private to prevent that but creates a new
`addBlockTagSection(BlockTagSection)` method that can be used by follow
up changes to safely add `BlockTagSection`s.

Note: The `blockTagSections` property definition had to be moved out of
the constructor as it is not possible to use `private set` in that case
to make the setter private.

Bug: 450228132
Test: ./gradlew
Change-Id: I52e080ff5519af9aa356cffe2ada6b8fcba5fcbe