Allow JavadocContent.accept(...) to return a value
Previously, `JavadocContent.accept(...)` returned `Unit`. That limited
the uses of it, necessitating the addition of `rewrite(...)`
and methods like `isMultiLine()` and `startsWithNewline()`.

This changes `accept(...)` to take and return a type parameter `<R>`.
The existing implementations of `JavadocContentVisitor` use `<Unit>`.
Follow-up changes will replace `rewrite(...)` with `accept(...)`
with a type parameter of `<JavadocContent?>`.

Bug: 450228132
Test: ./gradlew
Change-Id: Ibc8fb8c4ed6272c8ee6ce315ef603be223ec9c19