Autoupdate of 2 submodules
*frameworks/support updated to c9ba0f35fb200e563b16dacbda7387bbe73e1f2f
Merge "Change `VariantFailureIcon` to `GenericFailureIcon` and `FailureIcon` to `ConnectionFailureIcon` as suggested by the API council." into androidx-main
*golden updated to 3a0274476d923094e13606f54a25f2d12d34bafa
Change `VariantFailureIcon` to `GenericFailureIcon` and `FailureIcon` to `ConnectionFailureIcon` as suggested by the API council.

Bug: 453730430
Test: screenshots test
Relnote: "Add a new default `GenericFailureIcon` to be used in `FailureConfirmationDialog`, which shows a generic error icon. Rename existing FailureIcon to ConnectionFailureIcon."
Change-Id: I8e965334a1d419481debcf5481fde0b4e560678b