Remove @throws Throwable class type from the block tag content
Previously, `ThrowsTagType` extracted the `Throwable` class type from
the `@throws` block tag content but left the content unaffected. This
change removes the class type from the block tag content in preparation
for future changes which will resolve it to the fully qualified type.

Removing it from the content means that the `ThrowsTagData` is now
responsible for printing the content. That does result in white space
between the class type and the rest of the content being normalized to
a single white space necessitating a test change.

Bug: 447588621
Test: ./gradlew
Change-Id: I236cc056701b6f9bf9727dd180be4678bed9aa55