Add ImportResolver
Needed for follow-up changes that will use it to support resolving
names against imports in `AbstractSourceFile`.

Note: The test fails on snapshot because it does not create the nested
`Entry` class when it resolves `java.util.Map`. That means it cannot
find it in `nestedClasses()`. This should not affect the behavior in
practice as this should only be used when resolving documentation
references which should happen before snapshot not after.

Bug: 447588621
Test: ./gradlew
Change-Id: I254aa1699ca7c5605c7112ce239d25aafa532799