Add support for resolving type names relative to PackageItem
Currently, references to types within documentation are resolved using
Psi specific functionality. It is slow and inefficient and obviously
does not work for Turbine. This change is the first of a set of changes
that will add model independent support for that.

This adds a new `ReferencableNameScope` that will provide support for
resolving names that reference a part of the API to the specific part.
It will eventually support resolving packages, classes, type
parameters, constructors, methods and fields. This initial work just
supports resolving packages and classes from a package. Follow-up
changes will extend that.

Bug: 447588621
Test: ./gradlew
Change-Id: I8525c8de644b7312396f4fc9a6e15f80493edf01