Snapshot SourceFile
Previously, `CodebaseSnapshotTaker` would just reuse the `SourceFile`
from the original `Codebase` being snapshot. That worked because while
`classes()` did contain `clasreferences to `ClassItem`s in the original
`Codebase` that was not used in tests or the main Metalava code.

This change adds support for snapshotting `SourceFile`s (caching them
to match the behavior of Psi and Turbine).

Bug: 447588621
Test: ./gradlew
Change-Id: I340957a87bc10766b78b6f98c216c02d3d4748bf