Add test for SourceFile.getHeaderComments()
Shows the behavior is inconsistent in Java, Kotlin and Turbine.
Follow-up changes will correct it.

Kotlin is broken because it includes all top level comments in the file
not just the ones before package.

Turbine is broken because it includes everything up to the first
occurrence of `package` which is supposed to be the start of the
`package` statement but actually ends up truncating the following just
before `package`:

```
    // Inline comment before package
```

Bug: 447588621
Test: ./gradlew
Change-Id: I6847ad19db1e36208a51ebee27a12c385f2e0636