Cache PsiSourceFile instances across PsiBasedCodebase
Previously, `PsiSourceFile` instances were created on every call to
`PsiClassItem.sourceFile()` for top level, source classes. That meant
that `PsiClassItem` had to implement `equals(Object)` in order to pass
the tests.

This change removed the implementation of `equals(Object)` (and
`hashCode()`) to find the tests that were relying on that. They were
then switched to use `assertSame(...)` instead of `assertEquals(...)`
to make it clear that the same instance should be used.

The modified tests pass for Turbine as it already caches its
`TurbineSourceFile` instances. This change fixes the tests for Psi by
adding a similar caching mechanism for `PsiSourceFile`s.

Preparation to reduce the performance hit of some future work that
will add more functionality into `AbstractSourceFile` that increases
the cost of creation and the frequency they are accessed.

Bug: 447588621
Test: ./gradlew
Change-Id: I76838689a36942d70e671830c223db191053e2e0