Remove SourceFile.snapshot(Codebase)
Previously, `CodebaseSnapshotTaker` used `SourceFile.snapshot(...)` to
create the `SourceFileSnapshot` instance. However, because
`FilteringSourceFile.snapshot(...)` delegated to the `SourceFile` it
was filtering the snapshot bypassed the filtering resulting in imports
being kept that should have been removed.

This change removes the `snapshot(...)` method and creates the
`SourceFileSnapshot` instance directly to avoid that.

Bug: 447588621
Test: ./gradlew
Change-Id: Ie3cd02752d86769daca6fe0666b5962c061909ac