Fix TagData extraction for inline tags
Change https://r.android.com/3812648 added the ability for a `TagType`
to remove some content from the start a tag. Unfortunately, it did not
implement it correctly for inline tags. This change fixes it in
preparation for adding special support for `@link` which needs to
remove the reference to replace it with a fully qualified one.

Bug: 447588621
Test: ./gradlew
Change-Id: If05e42962f8cb0d5d93e1d40a1167f9f89579f78