Add tests for type name collisions between imports and package
While testing type resolution in `{@link}` tags in Android an edge case
was discovered that the `ReferencableNameScope` mechanism did not
handle. That case was when there was a class in the package whose name
collided with an on demand import (in that case the implicit on demand
import of `java.lang.*`).

The package class should have won but the type resolution chose the
imported class. That is because the order of processing was:
1. Named imports.
2. On demand imports (including implicit ones).
3. Containing package.

But it should have been:
1. Named imports.
2. Containing package.
3. On demand imports (including implicit ones).

This change adds tests to show the incorrect behavior. As the rules are
complex the test relies on the underlying model's type resolution
process to confirm the expected behavior is correct.

A follow-up change will fix the issue.

Bug: 447588621
Test: ./gradlew
Change-Id: I1a6e9ec3862efd4e362bdf1c14f5adedf79ccafd