Fix extractCustomLinkText(...) dropping parts of the label
Previously, it assumed that the label was encompassed in a single
`PsiDocToken` but that is not true when it is split across multiple
lines. In that case it would drop everything before the last line.

This change fixes it. The code is quite complicated for what it is
trying to do because it has to handle inconsistencies in the
`PsiElement` classes and complexity caused by leading asterisks at the
start of the line.

The model does not have this issue as it has good abstractions for
handling content and removes leading asterisks from the model
completely and replaces them only when printing.

So, this code will be very short lived as it will be removed once the
model supports fully resolving links. However, despite that it was
decided to fix it now as it reduces the number of differences between
the current behavior and the model behavior making it easier to see any
significant changes.

Bug: 447588621
Test: ./gradlew
Change-Id: If2aa1fb8f483c93c0f607248beb3cfb0134fd2a3