Move workaround for Java summary truncation to printing
Previously, it was applied as part of `DocAnalyzer` using a separate
`ApiVisitor` across the whole `Codebase`. That was very expensive and
inefficient. This change moves it to apply the workaround when
printing.

This does not fix the `{@link}` test added in the previous change as
the workaround looks for the first `.` to see if it is part of `e.g. `
but in this test it finds the first `.` in the fully qualified
`java.util.List`. That will be fixed in a follow-up change.

Bug: 447588621
Test: ./gradlew
Change-Id: Iadb6cb96ab29bbc4438d7d07fab36c625f6433a1