Make java summary truncation workaround inline tag aware
Previously, it would not ignore inline tags and would either miss
`e.g. ` that came after an inline tag with a `.` or it would apply the
workaround to an `e.g. ` in the inline tag. Neither of those was
correct as inline tags do not have any impact on the determination of
the summary line.

This change replaces the existing `String` based workaround with one
that uses a `JavadocContentRewriter` so can ignore inline tags.

This moves the application of the workaround from before it was fully
qualified until after. That was mainly because that is the point at
which the `DocComment` instance that was being printed was available
but it also matches the future behavior when relative references will
be fully qualified when parsing the Javadoc.

Bug: 447588621
Test: ./gradlew
Change-Id: Ibf25df10c96a1bae4536c753d59d178e1b86aa95