Avoid throwing a NPE when cannot find TypeBoundClass
Previously, `typeBoundClassForSymbol(ClassSymbol)` would throw a NPE if
it could not find a `TypeBoundClass` for `ClassSymbol`. That was based
off the assumption that if the latter would only be created if the
former exists. However, that does not appear to be strictly true.
`ClassSymbol`s can be created without a `TypeBoundClass`.

That causes problems downstream. It is unclear exactly why it happens
but `AnnoInfo.isContainerForRepeatableAnnotations()` attempts to get
a `TypeBoundClass` for an annotation `ClassSymbol` which cannot be
found and so `typeBoundClassForSymbol(...)` throws an exception.

While it cannot be reproduced it can be fixed. This change removes the
`null` check that was in `typeBoundClassForSymbol(...)` and leaves it
up the callers to decide what to do if it returns `null`. Some throw
an exception, others, like `isContainerForRepeatableAnnotations()` do
not.

Bug: 479907812
Test: ./gradlew
Change-Id: I2a0302a651babd199fd881e27da2b84b136598ff