Add record test with compact constructor
This is needed because a `Record` compact constructor has a unique
implicit parameter structure. A follow-up change will fix a problem
with Turbine's handling of implicit parameters and this will make sure
that it does not break.

It was necessary to set `API_LEVEL` to `34` as the `java.lang.Record`
super class of all `record` classes was added in that API version.

Bug: 458362596
Test: ./gradlew
Change-Id: Idcea52a6da0bc70d4bd30b5ac28e5201470ccb1f