Improve DefaultBaseModifierList.toString() method
Previously, it would just print the binary form of the internal `flags`
property but that made it difficult to understand. This change adds a
`ModifierKeyword` enum for all the modifier flags and uses that to
produce a list of keywords in `DefaultBaseModifierList.toString()`.

Preparation for future changes that add tests to improve the test
coverage of modifiers.

Bug: 458362596
Test: ./gradlew
Change-Id: If8cc56dd1ebd0f6d2a1427ced9dd3898d4bce199