Disambiguate overloads of `describe(...)` in Item.Companion
Previously, they were only differentiated by parameters and as they had
default parameters it was often hard to determine which version was
going to be used. This change renames them all based on their main
`Item` subclass to make it clearer which one is being used at each call
site.

Bug: 459795258
Test: ./gradlew
Change-Id: I46df99e39844a161a0aed0456f41d35d1bfdd7b9