Upgrade Turbine to 0.12.0
Turbine 0.12.0 changed the API Metalava relies on when it added
position information for Javadoc comments as well as javadoc comments
for `package` in `package-info.java` files so this change had to update
those parts or Metalava, allowing removal of some hacks.

Follow-up changes will take further advantage of some of the other
changes in Turbine 0.12.0.

Bug: 458362596
Test: ./gradlew
Change-Id: I8484d4e81b155cd2d0e9bd6c1e0de9001c241365