[Readability refactor] Move StubWriterConfig init logic to options.kt
Axioms/Assumptions:
* StubWriterConfig option is only used by createDocStubs()
* It only contains one attribute that will be true if docStubs is true
* It is possible to have the attribute = true if
  --exclude-documentation-from-stubs is not set on cmd

Because of the above, I opted to move the init logic for the config to
options so it is encapsulated in just one place. This makes it more
straightforward to see what the config's attributes depend on.

It is my hope that possibly the config's dependency on the exclude doc
stubs flag can be either simplified or removed

Bug: 445659663
Test: ./gradlew

Change-Id: I7eaac228db77e7cebdbbce7b989c301d76d8f68a