Report all syntax errors in Turbine
Previously, Turbine only reported the first syntax error it found. That
is because `Parser.parse(...)` throws a `TurbineError` on the first
syntax error it encounters in a file. This change catches the error and
adds it to a `TurbineLog`. While it will only report at most one error
per file, it can report errors from multiple files.

Bug: 458362596
Test: ./gradlew
Change-Id: Id8f53ed284e4b0f75b8a5909cd11883f3e302981