Simplify descriptionLineText rule
Previously, it was defined as one or more `descriptionLineNoSpaceNoAt`,
`SPACE` or `AT`. However, `descriptionLineNoSpaceNoAt` is only used
once so can be collapsed into `descriptionLineText`. Also,
`descriptionLineText` is a choice in `descriptionLineElement` which is
already allowed zero or more times so there is no need for
`descriptionLineText` to allow its content to be repeated.

So, this change collapses `descriptionLineNoSpaceNoAt` into
`descriptionLine` and removes `+` from `descriptionLine`.

Bug: 448905711
Test: ./gradlew
Change-Id: Idd2b3909f808372f0a458f2f3667beb8a3e34c2f