Remove AT lexer token
Initially, it was needed when the lexer was used when parsing the whole
javadoc and so needed to identify a block tag (started the line with
`AT`). However, now that `DocCommentParser` handles that and the lexer
is only used for parsing description blocks it no longer needs to treat
`@` any differently to plain text. So, this change removes it and
adds it to the `TEXT_CONTENT` token instead.

Bug: 448905711
Test: ./gradlew
Change-Id: Id7bd0597678eabd114bdb618500ca672141f311a