Dedup descriptionLineText and braceText
Prior to the preceding change the parser differentiated between the
states when braces had to be matched and when they did not. That
required two different rules for textual content, one when the braces
did not match (in which braces were treated as textual content) and one
where they did match (when braces were not treated as textual content).
However, the preceding change moved that into the lexer.

This change takes advantage of that to simplify the lexer rules and
dedup the two `...Text` parser rules.

Bug: 429965593
Test: ./gradlew
Change-Id: I292a7bcb5e5b059b3271c0e7e3c10ea45ea7834b