Move package consistency check back into ApiFile
This was previously moved from `ApiFile` to `PackageTracker` in change
https://r.android.com/3202058 because it did not provide all the
information that `ApiFile` needed to perform the check. However, it
made no sense for Psi and Turbine models. Since it was moved the code
has been substantially refactored and so it is now possible to
implement this in `ApiFile` again so this change moves it back.

Preparation for follow-up changes that refactor how package annotations
are managed.

Bug: 461491536
Test: ./gradlew
Change-Id: I47156d508f4cfe56dd6f2f09dba4479b79213521