Make Turbine handle packages annotation consistently
Previously, Turbine would extract package annotations correctly from
sources (i.e. `package-info.java`) but not from binary packages (i.e.
`package-info.class`). That was because for source packages it would
store the annotations in `PackageDoc` but that was not done for binary
packages.

Storing annotations for binary packages (which are created in
`createPackageFromUnderlyingModel(...)` in `PackageDoc.annotations`
would require duplicating functionality for getting the modifiers in
`createPackageFromUnderlyingModel(...)` that is already done in the
call to `gatherPackageJavadoc(...)`.

Instead of doing that this adds a new method
`createPackageAnnotations(String)` that takes a package name and
returns the annotations. That is used under the covers for both
`createPackageItem(...)` and `createPackageFromUnderlyingModel(...)`
avoiding duplication and ensuring consistent behavior.

The new mechanism is used by Turbine and it fixes the binary package
test that was previously failing.

Bug: 461491536
Test: ./gradlew
Change-Id: I72442ed2c3e09ed2726ada9550a756570207604e