Add PackageInfo and use in PackageTracker.createPackage(...)
Previously, `PackageTracker.createPackage(...)` took `PackageDocs`
and `annotations`. The `PackageDocs` was used to supply information for
the package and its containing packages and `annotations` contained the
annotations for the specific package. It would then get the`PackageDoc`
for the current package and pass that and the annotations down to
`CodebaseAssembler.createPackageItem(...)`. That was quite complex and
required `ApiFile` to construct `PackageDocs` and `PackageDoc` to
supply the file location for the package it was creating.

This change adds a `PackageInfo` data class which provides all the
information for the current package being created. That simplifies
`ApiFile`. `PackageTracker.findOrCreatePackage(...)` creates an
instance of that by combining information from the `PackageDoc` and
`annotations`.

This is preparation for follow-up changes that will use `PackageDoc`
solely for supplying information that was obtained from `package.html`
and `overview.html` files. Other information will be obtained from the
underlying model similarly to how annotations are currently processed.

Bug: 461491536
Test: ./gradlew
Change-Id: I47e62b3e05a553a44db2f23996187d1947c11bd3