Fix PsiFileLocation when used on a PsiFile
Previously, it would add a line number of `1` to the location, e.g.
`package.html:1`. That is misleading as it makes it seem as though the
problem is on that specific line. This change treats `PsiFile`
specially and avoids trying to find a line number, which is a waste of
time anyway.

Bug: 461491536
Test: ./gradlew
Change-Id: Ide51ad9ab30a5d438d023d47a66477b1d494d76e