Support package annotations in text resolve package tests
Previously, the `TextModelSuiteRunner` did not support annotations on
packages resolved by `ClassLoaderBasedCodebaseAssembler`. This change
adds some basic support for marker annotations using `URLClassLoader`
to access `RetentionPolicy.RUNTIME` annotations. It is not necessary
to add full support for annotations as that would only be adding test
code to test other test code.

This leverages the existing `URLClassLoader` used to resolve classes
to define a `j.l.Package` so its annotations can be accessed and then
they are used to create some marker annotations. The test was changed
from the default `RetentionPolicy.CLASS` to `RUNTIME` as otherwise
the annotations would not be accessible through `Package`. That does
not affect the test as the purpose was just to see that package
annotations were retrieved not check the behavior of the default
`RetentionPolicy`.

If better annotation support is needed then it will need to use some
`org.objectweb.asm` code to process the `package-info.class` files.

Bug: 462658629
Test: ./gradlew
Change-Id: Ib18d4a848d820928b82e324fc2083e0816ea4060