Replace comparison with REVERT_UNSTABLE_API with revertUnstableApi()
Previously, `getShowabilityForItem(...)` compared `itemShowability`
with `Showability.REVERT_UNSTABLE_API`. That would fail if either of
the `Showability` properties `recursive` or `forStubsOnly` were not
`ShowOrHide.REVERT_UNSTABLE_API`. At the moment that is not true but
it will change in the follow-up change.

This change switches it to call `Showability.revertUnstableApi()` which
is consistent with how other parts of the code checks to see whether an
item needs reverting. It also fixes some out of date documentation.

Bug: 463341334
Test: ./gradlew
Change-Id: I4bc402166278dd102c3a4922fe82d1ab69728009