Make SourceCodebaseAssembler return non-null PackageInfo.commentFactory
Previously, `SourceCodebaseAssembler`'s implementation of the
`getPackageInfoFromUnderlyingModel(...)` method just combined a
`PackageInfo` with an associated `PackageDoc`. It did not ensure that
the returned `PackageInfo.commentFactory` was not-`null`. It relied
on the subclasses to implement `emptyPackageDocumentationFactory()`.

This change removes those implementations and instead makes sure that
`PackageInfo.commentFactory` is not `null`, using the default
`ItemDocumentationFactory` where needed.

Bug: 461491536
Test: ./gradlew
Change-Id: I3c3aa9907ee265e19f564c6a5616bcc20bcc5036