Detect snapshot package error in getPackageInfoFromUnderlyingModel(...)
Previously, `CodebaseSnapshotTaker` would return `PackageInfo.EMPTY`
if it could not find a `PackageItem` in the original `Codebase`. It
relied on the `createPackageItem(...)` to detect that the
`commentFactory` was `null` and then call its
`emptyPackageDocumentationFactory()` implementation which then threw an
error. That was overly complex.

This change just throws the error directly in `CodebaseSnapshotTaker`'s
`getPackageInfoFromUnderlyingModel(...)` and removes its
`emptyPackageDocumentationFactory()` implementation as that is no
longer needed.

Bug: 461491536
Test: ./gradlew
Change-Id: I5a488ce8e76d89568fb692ed0023bc9cb05ac87e