Move extractAnnotations before migrating nulls
The external annotations file does not contain nullable annotations so
it is safe to write it before migrating nullable annotations.

This groups the stubs processing together and as the last thing that is
done which makes the code safer as stubs writing can modify the
`Codebase` so doing it last ensures that it does not affect other parts
outputs.

Bug: 464226866
Test: ./gradlew
Change-Id: I467772788f24b64c3a8daf9e08af1a3c9c03fae7