Remove duplicate check for public typedef class
A typedef class (e.g. annotated with `@IntDef`) should not be public.
It should either be marked `@hide` or not be `public` (or `protected`).
Previously, that was checked in two places, in `ApiLint` and in
`ExtractAnnotations`. This change removes the latter as `ApiLint` is
always run whereas `ExtractAnnotations` is only run when writing an
`annotations.zip` file. It also has no tests.

Bug: 399365344
Test: ./gradlew
Change-Id: Id2cd3ce7cc530d8c98cf7b062c578758b662a24e