Defer reporting missing previously released API
Previously, an error would be reported when `<api-flags>` were provided
in a config file without also providing a previously released API. That
was intended to ensure that disabling an API flag would work correctly.
However, it would fail, even if no `<api-flags>` were actually
disabled which is always the case when building the "everything" stubs
that are used to compile Android and when generating the "everything"
signature files which are checked in.

This replaces that early error checking with an error check that only
happens when an API change needs to be reverted but cannot because the
previously released API was not provided. That does mean that it will
not report problems if the flags are all enabled but that should not be
a problem. Existing sources that do have a previously released API are
not likely to have that removed as they will contain flags in a variety
of states, some disabled, some enabled. New sources that do not have a
previously released API will need one to be provided the first time a
flag is added as flags start in the disabled state.

Bug: 454856811
Test: ./gradlew
Change-Id: I5c51a1b21409f0415b5000553fa8a200e495e640