Stop pruning immutable/disabled flags from ApiFlags
Previously, `ApiFlags` would ignore any `ApiFlag` whose `ApiFlagAction`
would have been `REVERT`. That was an optimization to reduce the size
of the `ApiFlags` as the default behavior for an unknown flag was to
revert. However, that also lost the `isExported` setting which would
prevent the `UNEXPORTED_FLAGGED_API` check from working correctly.

This change removes `pruneDisabledFlags` option. A follow-up change
will provide a more configurable optimization.

Bug: 454856811
Test: ./gradlew
Change-Id: Idd72a2a9cfe0672a0b3b8593049ab3ea523140ee