Add ApiFlag.qualifiedName
Initially, `ApiFlag` was only one of three possible enums. Since then
the `isExported` has been added as well as a flag report command and
future changes will make more use of `ApiFlag`. This change adds
`qualifiedName` to `ApiFlag` which will make it easier to use.

This needed to make `ApiFlag` a `data class` as it is no longer
restricted to a fixed set of instances and needs to implement
`equals` and `hashCode()`.

Follow-up changes will simplify other code that uses `ApiFlag`.

Bug: 454856811
Test: ./gradlew
Change-Id: Id1f23bb4c7b607864492c8bc7bd3328cc08f8255