Remove TestFixture.reporter
Prior to change https://r.android.com/3838915 the `reporter` field was
provided to allow individual tests to track issues that were reported.
However, that change enforced issues were checked by every test so
having a customizable `reporter` field was unnecessary. This change
removes the field and replaces it with a `recordingReporter` field
instead that contains the `RecordingReporter` that is used by the test.

Bug: 454856811
Test: ./gradlew
Change-Id: I055b17480de3136ffe5d65199d07add0c4993602