Correct TurbineFileLocation for classes
Previously, it would report character position for class location which
differed from Psi. This change only reports character position for
ItemDocumentation.fileLocation` to match Psi.

Supporting `characterPosition` consistently across models would be
difficult as each model has its own idea as to the best location to
report for a specific item. So, this just keeps to lines except for
documentation location as Metalava ensures that is consistent across
models.

Bug: 454856811
Test: ./gradlew
Change-Id: I0a3b1b458aa419a7fad417aa1cc4994470e86f86