Ignore non-static const field values
Previously, `PsiFieldItem` would treat a Kotlin field as having a
constant value even if it was not `static` that produced an API
signature file that had errors as `ApiFile` will only expects values
on `static final` fields.

This change fixes that by making sure that Kotlin fields are only
treated as having a constant value if they are also `static`.

Bug: 467355045
Test: ./gradlew
Change-Id: I33cbcc7470eb65edc4e67b8a501bd143c42875cb