Ensure that the correct forms of tags are used
Previously, there was no checking to make sure that each tag used the
correct form for the type, i.e. inline or block. This change adds
checks to enforce that, reporting issues if there was a problem. Tag
types that have no specialized behavior in Metalava are assumed to be
usable in both forms.

Bug: 469362503
Test: ./gradlew
Change-Id: I72addc0769242a7812f84acd1a5f2f56c846ed4d